% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{absolute_error}
\alias{absolute_error}
\title{Absolute Error}
\usage{
absolute_error(
  true,
  pred,
  weights = NULL,
  rweights = NULL,
  multi = c("average", "sum", "raw")
)
}
\arguments{
\item{true}{a vector (or a matrix) of observed values. If a matrix is provided,
a multi-response is assumed}

\item{pred}{a vector (or a matrix) of predicted values}

\item{weights}{observation weights. This argument is for consistency, but not currently used.}

\item{rweights}{response weights}

\item{multi}{what to do when response has multiple output values
\describe{
\item{\code{average}}{errors of multiple outputs are averaged to get a single value for each observation}
\item{\code{sum}}{errors of multiple outputs are summed up to get a single value for each observation}
\item{\code{raw}}{returns a vector containing one error for each output}
}}
}
\value{
the absolute error, a vector of positive \code{double} values or a matrix,
having one column for each response, if response has multiple output values and \code{multi = "raw"}
}
\description{
This function computes the absolute error.
}
\details{
The absolute error (AE) is based on absolute values and it is defined as

\deqn{AE(y_{i},\hat{y}_{i}) = \lvert y_{i} - \hat{y}_{i} \rvert}

where \eqn{y_{i}} is the true value of the \eqn{i}-th sample and \eqn{\hat{y}_{i}}
is the predicted value.

If the observed data has multiple output values and weights are provided, then the error is weighted.

\deqn{wAE(w,y,\hat{y}) = \frac{1}{\sum_{j=1}^{n} w_{j}} w * \lvert y_{i} - \hat{y}_{i} \rvert}

where \eqn{w_{j}} is the weighting factor assigned to the \eqn{j}-th response.
The best possible score is zero.
}
\author{
Alessandro Barberis
}
