% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{absolute_percentage_error}
\alias{absolute_percentage_error}
\title{Absolute Percentage Error}
\usage{
absolute_percentage_error(
  true,
  pred,
  weights = NULL,
  rweights = NULL,
  multi = c("average", "sum", "raw"),
  eps = 1e-10
)
}
\arguments{
\item{true}{a vector (or a matrix) of observed values. If a matrix is provided,
a multi-response is assumed}

\item{pred}{a vector (or a matrix) of predicted values}

\item{rweights}{response weights}

\item{multi}{what to do when response has multiple output values
\describe{
\item{\code{average}}{errors of multiple outputs are averaged to get a single value for each observation}
\item{\code{sum}}{errors of multiple outputs are summed up to get a single value for each observation}
\item{\code{raw}}{returns a vector containing one error for each output}
}}
}
\value{
the absolute percentage error, a vector of positive \code{double} values or a matrix,
having one column for each response, if response has multiple output values and \code{multi = "raw"}
}
\description{
This function computes the absolute percentage error.
}
\details{
The absolute percentage error (APE) is based on absolute values.
The absolute percentage error is defined as

\deqn{APE(y,\hat{y}) = \frac{\lvert y_{i} - \hat{y}_{i} \rvert}{\max{\lvert y_{i} \rvert, \epsilon}}}

where \eqn{y_{i}} is the true value of the \eqn{i}-th sample, \eqn{\hat{y}_{i}}
is the predicted value, and \eqn{\epsilon} is an arbitrary positive small number used
to avoid a division by zero.

If response has multiple output values and weights are provided, then the error is weighted.

\deqn{wAPE(w,y,\hat{y}) = \frac{w * \lvert y_{i} - \hat{y}_{i} \rvert}{\max{\lvert y_{i} \rvert, \epsilon}}}

where \eqn{w} is a vector of \eqn{n} elements, \eqn{w_{j}} is the weighting factor assigned to the \eqn{j}-th response.
The best possible score is zero.
}
\author{
Alessandro Barberis
}
