% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Trained.R
\name{assess,Trained-method}
\alias{assess,Trained-method}
\title{Assess performance using test data}
\usage{
\S4method{assess}{Trained}(object, newx, newy, weights = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{renoir.trained.cv.glmnet} containing the fitted data}
}
\description{
Assess performance using test data
}
