% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_model_selection.R
\name{bm}
\alias{bm}
\title{Best model selection}
\usage{
bm(object, ...)
}
\arguments{
\item{object}{object to consider for extracting the best model}

\item{...}{further arguments}
}
\description{
Best model selection
}
