% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Sampler.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap}
\usage{
bootstrap(N, n, strata = NULL, balance = FALSE)
}
\arguments{
\item{N}{population size.}

\item{n}{sample size.}

\item{strata}{(optional) vector of stratification variables. If provided must be of length \code{N}}

\item{balance}{logical, whether to (try to) balance the strata frequency in the output sample.
Used only if \code{strata} is provided.}
}
\value{
A vector of length \code{n} containing the index of the computed random set of observations.
}
\description{
Takes a sample with replacement from the population.
}
\details{
This function is an interface to \code{\link{rsample}} with
\code{replace = TRUE}.
}
\seealso{
\code{\link{rsample}}
}
\author{
Alessandro Barberis
}
\keyword{internal}
