% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check.filename.and.generate}
\alias{check.filename.and.generate}
\title{Check filename and generate a new one}
\usage{
check.filename.and.generate(outdir, filename, ext)
}
\arguments{
\item{outdir}{path to the output directory}

\item{filename}{name of the file to check}

\item{ext}{file extension (e.g. 'txt')}
}
\description{
This function checks the existence of a file in the
specified directory, and in such case generates a new filename.
}
\keyword{internal}
