% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Logger.R
\name{check_file_path_and_open_con}
\alias{check_file_path_and_open_con}
\title{Check file path and open connection}
\usage{
check_file_path_and_open_con(file.path = NULL)
}
\description{
Check file path and open connection
}
\keyword{internal}
