% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_function_arg}
\alias{check_function_arg}
\title{Check if function parameter is valid}
\usage{
check_function_arg(arg = NULL, check.na = TRUE, check.null = TRUE)
}
\description{
Check if function parameter is valid
}
\keyword{internal}
