% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Forecaster.R
\name{check_predicted_by_glmnet}
\alias{check_predicted_by_glmnet}
\title{Check predict output}
\usage{
check_predicted_by_glmnet(object, type)
}
\value{
a standardised output prediction object
}
\description{
Function to check the output of predict function call and
standardise object. Expected objects: if output is multi-response,
a matrix nobs x nouts, if single response a vector of length nobs.
}
\keyword{internal}
