% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Tuner.R
\name{check_tuning_type_measure}
\alias{check_tuning_type_measure}
\title{Check tuning type measure}
\usage{
check_tuning_type_measure(type.measure = "mse", resp.type = "gaussian")
}
\arguments{
\item{type.measure}{the accuracy measure}

\item{resp.type}{the response type}
}
\description{
Check if the selected accuracy measure is supported for tuning.
}
\keyword{internal}
