% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{check_type_measure}
\alias{check_type_measure}
\title{Check the accuracy measure}
\usage{
check_type_measure(type.measure = "mse", resp.type = "gaussian")
}
\arguments{
\item{type.measure}{the accuracy measure}

\item{resp.type}{the response type}
}
\value{
Character string, the checked accuracy measure.
}
\description{
Check if the selected accuracy measure is supported for the considered
response type.
}
\author{
Alessandro Barberis
}
\keyword{internal}
