% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence.R
\name{ci}
\alias{ci}
\title{Confidence Interval}
\usage{
ci(estimate, se, confidence = 0.95, distribution = c("normal", "t"), n)
}
\arguments{
\item{estimate}{sample statistic from population having distribution \code{distribution}}

\item{se}{standard error (standard deviation of \code{estimate})}

\item{confidence}{the desired confidence level}

\item{distribution}{sampling distribution of the estimate.
Use \code{normal} if the population has unknown mean and known variance (or if \code{n} is large),
\code{t} if population has unknown mean and variance}

\item{n}{sample size, used to compute the degrees of freedom if \code{distribution = "t"}}
}
\value{
A named numeric vector with 2 elements, \code{low} and \code{up},
the lower and upper bounds of the confidence interval
for the point estimate.
}
\description{
This function computes the confidence interval of
a point estimate.
}
\details{
The confidence interval provides additional information about the
variability of a point estimate and it is generally defined as

\deqn{CI = estimate \pm \textit{margin of error} = estimate \pm \textit{critical value} \times \textit{standard error}}

where \eqn{\textit{estimate}} is the sample statistic estimating the population parameter of interest;
\eqn{\textit{critical value}} is a value based on the sampling distribution of the estimate
and the desired confidence level; \eqn{\textit{standard error}} is the standard deviation of the
point estimate.

Since estimates from different samples from the same population can be different, the estimate
can be seen as a random variable with its own distribution.
Given that we are not interested in the estimate for a specific sample but we rather
want to draw conclusions about the population, the confidence interval can be used to
obtain information about the population parameter.

Beware about the interpretation. The CI is not telling us if the population parameter is inside its range.
In fact, as generally we don't know the population parameter, we don't know if the interval
contains it. However, we know that a certain number of intervals (our confidence level, e.g. 95\\%)
formed in this way will contain the parameter.
}
\references{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC5723800/}
}
\author{
Alessandro Barberis
}
