% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence.R
\name{ciwm}
\alias{ciwm}
\title{Confidence Interval for the Weighted Mean}
\usage{
ciwm(x, weights, ..., confidence = 0.95, distribution = "normal", n)
}
\arguments{
\item{x}{vector of measurements}

\item{weights}{vector of weights}

\item{...}{further arguments to \link{sewm}}

\item{confidence}{the desired confidence level}

\item{distribution}{sampling distribution of the estimate.
Use \code{normal} if the population has unknown mean and known variance (or if \code{n} is large),
\code{t} if population has unknown mean and variance}

\item{n}{sample size, used to compute the degrees of freedom if \code{distribution = "t"}}
}
\value{
\code{list} containing three elements
\describe{
\item{m}{sample mean}
\item{sem}{standard error of the mean}
\item{ci}{confidence interval for the sample mean}
}
}
\description{
This function computes the confidence interval (CI) for the
weighted sample mean.
}
\details{
The confidence interval of the weighted sample mean provides additional information about the
variability of the population parameter estimate. It is defined as

\deqn{CI = mean \pm \textit{critical value} \times \textit{standard error of the mean (SEM)}}

where \eqn{mean = \frac{\sum_{i=1}^{n} w_{i} x_{i}}{\sum_{i=1}^{n} w_{i}}} is the weighted mean.
}
\references{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC5723800/}
}
\seealso{
\code{\link{ci()}} for more information on confidence intervals, \code{\link{sewm()}} for the standard error
of the weighted sample mean.
}
\author{
Alessandro Barberis
}
