% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{classification_metric}
\alias{classification_metric}
\title{Classification metric}
\usage{
classification_metric(true, pred, confusion = NULL, positive = 1, metric)
}
\arguments{
\item{true}{a vector (or a matrix) of observed values. If a matrix is provided,
a multi-response is assumed}

\item{pred}{a vector (or a matrix) of predicted values}

\item{confusion}{a confusion matrix as returned by \code{\link{confusion_matrix}}. If \code{confusion} is not provided,
a confusion matrix is internally computed by using \code{true} and \code{pred}}

\item{positive}{integer or character indicating the target class.
If \code{positive = 0} (default), global values are computed.}

\item{metric}{character string indicating the score to compute}
}
\value{
A classification metric
}
\description{
This function computes a classification metric.
}
\details{
This function has a common interface to compute a classification metric.
}
\author{
Alessandro Barberis
}
\keyword{internal}
