% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{classification_report}
\alias{classification_report}
\title{Classification report}
\usage{
classification_report(true, pred, confusion = NULL)
}
\arguments{
\item{true}{a vector (or a matrix) of observed values. If a matrix is provided,
a multi-response is assumed}

\item{pred}{a vector (or a matrix) of predicted values}

\item{confusion}{confusion matrix}
}
\value{
An object of class \code{ClassificationReport}.
}
\description{
Classification report
}
\examples{
\dontrun{
true = c(0, 1, 2, 2, 0)
pred = c(0, 0, 2, 1, 0)

classification_report(true = true, pred = pred)
x = confusion_matrix(true = true, pred = pred)
cr = classification_report(confusion = x)

true = c('dog', 'cat', 'dog', 'cat', 'dog', 'dog', 'cat', 'dog', 'cat', 'dog')
pred = c('dog', 'dog', 'dog', 'cat', 'dog', 'dog', 'cat', 'cat', 'cat', 'cat')
x = confusion_matrix(true = true, pred = pred)
print.ConfusionMatrix(x = x)
print.ConfusionMatrix(x = x, format = 'html')

true = sample(x = 1:4, size = 10, replace = T)
pred = sample(x = 1:4, size = 10, replace = T)
x = confusion_matrix(true = true, pred = pred)

print.ConfusionMatrix(x = x, format = 'html')
}
}
\keyword{internal}
