% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{classification_report_binomial}
\alias{classification_report_binomial}
\title{Classification report for binomial data}
\usage{
classification_report_binomial(true, pred, confusion = NULL, positive = 1)
}
\arguments{
\item{true}{a vector (or a matrix) of observed values. If a matrix is provided,
a multi-response is assumed}

\item{pred}{a vector (or a matrix) of predicted values}

\item{confusion}{confusion matrix}

\item{positive}{index of 'positive' class}
}
\value{
An object of class \code{BinaryClassificationReport}.
}
\description{
Classification report for binomial data
}
\keyword{internal}
