% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Trained.R
\name{clean,Trained-method}
\alias{clean,Trained-method}
\title{Clean Trained objects}
\usage{
\S4method{clean}{Trained}(object, rm.call = F, ...)
}
\arguments{
\item{object}{a \linkS4class{Trained} object}
}
\value{
the cleaned \linkS4class{Trained} object
}
\description{
This function tries to clean the trained object to avoid that the final
object's size raises too much when the initial set of features and observations is big.
}
\keyword{internal}
