% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean.cv.glmnet}
\alias{clean.cv.glmnet}
\title{Clean cv.glmnet objects}
\usage{
clean.cv.glmnet(object)
}
\arguments{
\item{object}{a \code{cv.glmnet} object}
}
\value{
the cleaned object
}
\description{
This function remove the \code{x} from the \code{call}
slots of the trained object. It is needed to avoid that the final
object's size raises too much when the initial set of features and observations is big.
}
