% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_EvaluatedList.R
\name{compute_importance}
\alias{compute_importance}
\title{Compute feature importance}
\usage{
compute_importance(
  object,
  marker,
  scorer,
  set,
  measure,
  logger = Logger(verbose = F),
  features,
  epsilon = 1,
  ...
)
}
\arguments{
\item{object}{an object of class \linkS4class{EvaluatedList}}

\item{marker}{an object of class \linkS4class{Marker}}

\item{scorer}{an object of class \linkS4class{Scorer}}

\item{set}{the data set to consider (\code{train}, \code{test} or \code{full})}

\item{measure}{the performance metric to consider}

\item{logger}{an object of class \linkS4class{Logger}}

\item{features}{(optional) vector containing the predictors to consider}

\item{epsilon}{small number to avoid division by zero}

\item{...}{not currently used}
}
\value{
The computed importance for each feature.
}
\description{
Compute feature importance
}
\keyword{internal}
