% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma.R
\name{compute_limma}
\alias{compute_limma}
\title{Linear model for series of arrays}
\usage{
compute_limma(
  x,
  y,
  weights = NULL,
  resp.type = c("gaussian", "mgaussian", "binomial", "multinomial", "poisson", "cox"),
  logged2 = FALSE,
  coef = NULL,
  ...
)
}
\arguments{
\item{x}{A matrix-like data object containing log-ratios or log-expression
values for a series of arrays, with rows corresponding to genes and columns to samples.}

\item{y}{the response variable}

\item{weights}{vector of observation weights}

\item{resp.type}{the response type}

\item{logged2}{logical, whether data was logged 2}

\item{coef}{column number or column name specifying which coefficient or contrast of the linear
model is of interest. For topTable, can also be a vector of column subscripts, in which
case the gene ranking is by F-statistic for that set of contrasts.
If coef has a single value, then the moderated t-statistics and p-values for that coefficient
or contrast are used.
If coef takes two or more values, the moderated F-statistics for that set of coefficients or
contrasts are used. If coef is left NULL, then all the coefficients or contrasts in the fitted
model are used, except that any coefficient named (Intercept) will be removed.}

\item{...}{further arguments to \code{\link[renoir]{limma}}}
}
\value{
a list of element as returned by \code{\link[renoir]{limma}}
}
\description{
Fit linear model for each feature given a series of observations.
This function is a wrapper for \code{\link[renoir]{limma}}
}
\seealso{
\code{\link{limma}}
}
\author{
Alessandro Barberis
}
\keyword{internal}
