% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruit.R
\name{compute_rank_product}
\alias{compute_rank_product}
\title{Compute Rank Product/Rank Sum}
\usage{
compute_rank_product(
  data,
  cl = rep(1, times = ncol(data)),
  logged = T,
  gene.names = rownames(data),
  ...
)
}
\arguments{
\item{data}{a matrix vars x obs}

\item{cl}{a vector containing the class labels of the samples.
In the two class unpaired case, the label of a sample
is either 0 (e.g., control group) or 1 (e.g., case group).
For one class data, the label for each sample should be 1}

\item{logged}{if "TRUE" data have been previously log transformed. Otherwise it 
should be set as "FALSE"}

\item{gene.names}{if "NULL", no gene name will be attached to the outputs,
otherwise it contains the vector of gene names}

\item{...}{further arguments to \code{\link[RankProd]{RankProducts}}}
}
\value{
a dataframe containing different elements
\describe{
  \item{code{pfp_}}{estimated percentage of false positive predictions for each feature
  positively and negatively correlated with the outcome}
  \item{code{p_val}}{estimated p-value for each feature
  positively and negatively correlated with the outcome}
  \item{code{Rank_Stat_}}{the rank-product/rank-sum statistics evaluated per each gene
  positively and negatively correlated with the outcome}

}
}
\description{
Compute Rank Product/Rank Sum
}
\keyword{internal}
