% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam.R
\name{compute_sam}
\alias{compute_sam}
\title{Significance analysis of microarrays}
\usage{
compute_sam(
  x,
  y,
  resp.type = c("gaussian", "mgaussian", "binomial", "multinomial", "poisson", "cox"),
  coef = NULL,
  logger,
  ...
)
}
\value{
The p-value associated with each feature
}
\description{
This function computes the significance analysis of microarrays.
This function is a wrapper for \code{\link[renoir]{sam}}
}
\keyword{internal}
