% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruit.R
\name{compute_significance_score}
\alias{compute_significance_score}
\title{Compute Significance Score}
\usage{
compute_significance_score(weights, recruitment)
}
\arguments{
\item{weights}{the weights for the computation}

\item{recruitment}{the recruitment terms}
}
\value{
the significance score for each feature
}
\description{
Compute Significance Score
}
\keyword{internal}
