% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruit.R
\name{compute_significance_score_recuitment_term}
\alias{compute_significance_score_recuitment_term}
\title{Compute Significance Score Term}
\usage{
compute_significance_score_recuitment_term(
  coefs,
  recruitment.term = c("coefficient", "presence")
)
}
\arguments{
\item{coefs}{list of coefficients from all the runs}

\item{recruitment.term}{whether to use a binary variable (\code{presence = 0} or \code{presence = 1})
multiplied by the sign of the coefficient, or to use directly the coefficient of a feature
in the computation of the significance score.}
}
\value{
a Matrix with the significance score terms for each feature,
for each model
}
\description{
This function computes the normalised coefficients or the
presence term for a single repsonse type (e.g. gaussian, binomial,
poisson, cox).
}
\keyword{internal}
