% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{compute_test_on_PC}
\alias{compute_test_on_PC}
\title{Compute significance on PC}
\usage{
compute_test_on_PC(
  PCA,
  pc = "PC1",
  strata = NULL,
  survival = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{PCA}{principal component analysis}

\item{pc}{which principal component to consider}

\item{...}{further arguments to compute_score}

\item{groups}{groups to consider while performing the test}
}
\value{
non adjusted p-value
}
\description{
Computes wilcoxon or Kruskal-Wallis test for the given groups
on the selected principal components
}
\keyword{internal}
