% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{confusion_matrix}
\alias{confusion_matrix}
\title{Confusion Matrix}
\source{
See \href{https://en.wikipedia.org/wiki/Confusion_matrix}{Wiki page} for further information.
}
\usage{
confusion_matrix(true, pred)
}
\arguments{
\item{true}{a vector (or a matrix) of observed values. If a matrix is provided,
a multi-response is assumed}

\item{pred}{a vector (or a matrix) of predicted values}
}
\value{
the confusion matrix, a \code{table} reporting the number of samples
for different groups. The matrix element \code{(i, j)} represents the number of samples
known to be in group \code{i} and predicted in group \code{j}.
}
\description{
This function produces a confusion matrix given a set of observed and
predicted values.
}
\details{
The confusion matrix is a specific table that reports the
classification accuracy of an algorithm. In this implementation,
a confusion matrix C has rows reporting instances of the actual classes,
while columns report instances of the predicted classes.
}
\author{
Alessandro Barberis
}
