% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{confusion_matrix_outcomes}
\alias{confusion_matrix_outcomes}
\title{Confusion Matrix outcomes}
\usage{
confusion_matrix_outcomes(true, pred, confusion = NULL, positive = 0)
}
\arguments{
\item{true}{a vector (or a matrix) of observed values. If a matrix is provided,
a multi-response is assumed}

\item{pred}{a vector (or a matrix) of predicted values}

\item{confusion}{a confusion matrix as returned by \code{\link{confusion_matrix}}. If \code{confusion} is not provided,
a confusion matrix is internally computed by using \code{true} and \code{pred}}

\item{positive}{integer or character indicating the target class.
If \code{positive = 0} (default), global values are computed.}
}
\value{
a list containing the number of false positives (\code{FP}), false negatives (\code{FN}),
true positives (\code{TP}), and true negatives (\code{TN}) of a confusion matrix.
}
\description{
This function returns the outcome of a confusion matrix in a list.
}
\details{
The outcomes of a \code{\link[confusion_matrix]{confusion matrix}} are returned as elements of a list.
The four possible outcomes are:
\describe{
  \item{TP}{true positives, when the actual presence of a condition is correctly predicted}
  \item{TN}{true negatives, when the actual absence of a condition is correctly predicted}
  \item{FP}{false positives, when the actual absence of a condition is wrongly predicted}
  \item{FN}{false negatives, when the actual presence of a condition is wrongly predicted}
}
}
\author{
Alessandro Barberis
}
