% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{create_dt}
\alias{create_dt}
\title{Create DataTable}
\usage{
create_dt(
  data,
  col.filter = TRUE,
  use.regex = TRUE,
  scrollX = FALSE,
  rownames = F
)
}
\arguments{
\item{data}{a data object (either a matrix or a data frame)}

\item{col.filter}{logical, whether to use column filters}

\item{use.regex}{logical, whether to use regular expressions
for searching}

\item{scrollX}{logical, whether to enable horizontal scrolling. \code{FALSE} by default}

\item{rownames}{logical, whether to show \code{data} rownames. \code{FALSE} by default}
}
\value{
an HTML widget to display rectangular data
}
\description{
This function sets up some arguments before
a call to \code{\link[DT]{datatable}}. It is used
in the generation of the interactive report.
}
\keyword{internal}
