% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_EvaluatedList.R
\name{create_EvaluatedList}
\alias{create_EvaluatedList}
\title{Create EvaluatedList Object}
\usage{
create_EvaluatedList(..., grouping = T)
}
\arguments{
\item{...}{list of \linkS4class{Evaluated} objects}

\item{grouping}{logical, whether to group \linkS4class{Evaluated} objects
by common id/config/response/sampling}
}
\value{
An object of class \linkS4class{EvaluatedList}
}
\description{
Create EvaluatedList Object
}
\keyword{internal}
