% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_filename}
\alias{create_filename}
\title{Create a filename}
\usage{
create_filename(filename, n)
}
\arguments{
\item{filename}{string, a file name}

\item{n}{sample size}
}
\value{
A string, the file name.
}
\description{
Create a filename
}
\keyword{internal}
