% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Sampler.R
\name{create_folds}
\alias{create_folds}
\title{Create k folds}
\usage{
create_folds(N, k, strata = NULL)
}
\arguments{
\item{N}{population size}

\item{k}{number of folds}

\item{strata}{(optional) vector of stratification variables. If provided must be of length \code{N}}
}
\value{
vector of length \code{N} containing the fold ids
}
\description{
Assigns the population to k folds. If \code{strata} is provided, a stratified approach
is adopted so that the percentage of each stratum in the population is preserved in each fold.
}
\author{
Alessandro Barberis
}
\keyword{internal}
