% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Renoir.R
\name{create_report,Renoir-method}
\alias{create_report,Renoir-method}
\title{Create an interactive report}
\usage{
\S4method{create_report}{Renoir}(
  object,
  output_format = "html_document",
  outdir = tempdir(),
  filename = NULL,
  report.type = c("full", "short"),
  annotation = NULL,
  feat.index = 0,
  ...
)
}
\arguments{
\item{object}{an object of class \linkS4class{Renoir}}

\item{outdir}{path to the output directory}

\item{filename}{(optional) name without extension for the output file}

\item{report.type}{the type of report to create. If \code{short}, the feature importance tables
are not created (this could solve the pandoc "out of memory" issue when the number of considered features is high)}

\item{annotation}{a data frame containing features annotation}

\item{feat.index}{numeric, the index of the column in \code{annotation} matching \code{x} feature names.
\code{0} means rownames}

\item{...}{further arguments to \code{\link[rmarkdown]{render}} \code{params} argument. Possible options are:
\describe{
\item{\code{yml_author}}{A character vector, name of the author(s)}
\item{\code{yml_affiliation}}{The author’s affiliation; must match length of name,
e.g. if name has length of two, affiliation must as well; use NA if you don’t want
to include an affiliation for a given author.Note that not all formats support the affiliation field.}
\item{\code{yml_email}}{The author email address. Note that not all formats support the email field.}
\item{\code{yml_title}}{A character vector, the title of the document}
\item{\code{yml_subtitle}}{A character vector, the subtitle of the document.
Not all R Markdown formats use subtitles, so it may depend on what you use in the output field.
It is available in \code{pdf_document}, \code{html_document}, and \code{word_document} by default}
\item{\code{yml_abstract}}{A character vector, the abstract. Long character vectors
are automatically wrapped using valid YAML syntax.
This field is not available in all output formats; it is available in \code{pdf_document} and \code{html_document} by default.}
\item{\code{yml_name}}{A character vector, the name of the document}
\item{\code{tabset}}{logical, whether to use tabbed sections (if any)}
}}
}
\value{
The compiled report is written into the output file.
}
\description{
Create an interactive report for the computed analysis.
}
\author{
Alessandro Barberis
}
