% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ScorerList.R
\name{create_ScorerList}
\alias{create_ScorerList}
\title{ScorerList Constructor}
\usage{
create_ScorerList(
  resp.type,
  measures,
  grouped = TRUE,
  parameters,
  logger = Logger()
)
}
\arguments{
\item{resp.type}{response type}
}
\value{
a \linkS4class{ScorerList}
}
\description{
Creates a list of supported scorers for the selected response type
}
