% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{create_text_for_tooltip}
\alias{create_text_for_tooltip}
\title{Create textual info for tooltip}
\usage{
create_text_for_tooltip(model, learner, show.config = FALSE, index = NULL)
}
\arguments{
\item{model}{a \linkS4class{Tuned} or a \linkS4class{Trained} object}

\item{learner}{a \linkS4class{Learner} object}

\item{show.config}{logical, whether to add information about the hyperparameters configuration}

\item{index}{(optional) model index}
}
\value{
A character string containing information to report in the plot tooltip.
}
\description{
Create textual info for tooltip
}
\keyword{internal}
