% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_varnames}
\alias{create_varnames}
\title{Create Feature Names}
\usage{
create_varnames(nm = NULL, n, prefix = "V")
}
\arguments{
\item{nm}{vector containing the feature names}

\item{n}{a length-one integer vector, number of features}

\item{prefix}{a length-one character vector, used in the
creation of the feature names}
}
\value{
A character vector containing the feature names.
}
\description{
This functions create feature names.
}
\details{
Function to standardise the creation of
feature names in renoir. If \code{nm} is provided
and is not \code{NULL}, the function returns \code{nm}.
If \code{nm = NULL}, a vector of length \code{n}
containing the feature names defined by the internal
default rule is returned.The rule creates the name so
that \code{prefix} is added to the number of the feature.
For example, creating names for \code{n = 2} features
with \code{prefix = "V"} will return \code{"V1", "V2"}.
}
\author{
Alessandro Barberis
}
\keyword{internal}
