% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Screener.R
\name{default_screener}
\alias{default_screener}
\title{Compute the significance of the filtering test}
\usage{
default_screener(
  x,
  y,
  weights = NULL,
  alternative = "two.sided",
  method = "pearson",
  conf.level = 0.95,
  resp.type = c("gaussian", "mgaussian", "binomial", "multinomial", "poisson", "cox"),
  observations = NULL,
  coef = NULL,
  adjust.method = "none",
  logger,
  multi = c("max", "raw", "average", "sum"),
  ...
)
}
\arguments{
\item{x}{the input matrix, where rows are observations and columns are variables.}

\item{y}{the response variable. Its number of rows must match the number of rows of \code{x}.}

\item{weights}{priors of the observations}

\item{alternative}{alternative hypotesis to use. Must be one of \code{"two.sided"} (default),
\code{"greater"} or \code{"less"}.}

\item{method}{which correlation method to use. Currently, only "pearson" is supported.}

\item{conf.level}{confidence levels used for the confidence intervals (where computed). A single number or a
numeric vector with value for each observation. All values must be in the range of \code{[0;1]}.}

\item{resp.type}{the response type}

\item{observations}{(optional) indices of observations to keep}

\item{coef}{(optional) an integer indicating the response variable to consider in multi-response data
when \code{multi = "raw"}}

\item{adjust.method}{method used to adjust the p-values for multiple testing. Options, in
increasing conservatism, include "none", "BH", "BY" and "holm"}

\item{logger}{a \linkS4class{Logger}}

\item{multi}{what to do when response has multiple output values
\describe{
\item{\code{max}}{the max value of scores across multiple outputs is selected to get a single value for each observation}
\item{\code{average}}{scores of multiple outputs are averaged to get a single value for each observation}
\item{\code{sum}}{scores of multiple outputs are summed up to get a single value for each observation}
\item{\code{raw}}{returns the scores for the multiple outputs}
}}

\item{...}{further arguments}
}
\value{
a \linkS4class{Screened} object
}
\description{
This function compute the significance of the screening test
}
\author{
Alessandro Barberis
}
