% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Screener.R
\name{ebayes_screener}
\alias{ebayes_screener}
\title{Compute the significance of the filtering test}
\usage{
ebayes_screener(
  x,
  y,
  weights = NULL,
  resp.type,
  observations = NULL,
  logged2 = TRUE,
  method = "ls",
  design = NULL,
  contrasts = NULL,
  assay.type = c("array", "seq"),
  proportion = 0.01,
  coef = NULL,
  number = Inf,
  adjust.method = "BH",
  p.value = 1,
  top.table.sort.by = "none",
  return.type = c("pvalue", "object"),
  multi = c("combine", "max", "average", "sum", "raw"),
  logger = Logger(verbose = F),
  ...
)
}
\arguments{
\item{x}{the input matrix, where rows are observations and columns are variables.}

\item{y}{the response variable. Its number of rows must match the number of rows of \code{x}.}

\item{weights}{priors of the observations}

\item{resp.type}{the response type}

\item{observations}{(optional) indices of observations to keep}

\item{logged2}{whether data was logged 2.}

\item{method}{fitting method; \code{"ls"} for least squares or \code{"robust"} for robust regression}

\item{design}{the design matrix of the microarray experiment, with rows corresponding to samples and columns to coefficients to be estimated.
  Defaults to \code{object$design} if that is non-NULL, otherwise to the unit vector meaning that all samples will be treated as replicates of a single treatment group.}

\item{contrasts}{numeric matrix with rows corresponding to coefficients in \code{fit} and columns containing contrasts. May be a vector if there is only one contrast. \code{NA}s are not allowed.}

\item{assay.type}{Assay type: "array" for microarray data, "seq" for sequencing data}

\item{proportion}{numeric value between 0 and 1, assumed proportion of genes which are differentially expressed}

\item{coef}{(optional) an integer indicating the response variable to consider in multi-response data
when \code{multi = "raw"}}

\item{number}{maximum number of genes to list}

\item{adjust.method}{method used to adjust the p-values for multiple testing. Options, in
increasing conservatism, include "none", "BH", "BY" and "holm"}

\item{p.value}{cutoff value for adjusted p-values. Only genes with lower p-values are listed.}

\item{top.table.sort.by}{character string specifying statistic to rank genes by.
Possible values for topTable are "logFC", "AveExpr", "t", "P", "p", "B"
or "none". (Permitted synonyms are "M" for "logFC", "A" or "Amean" for "AveExpr",
"T" for "t" and "p" for "P".)}

\item{multi}{what to do when response has multiple output values
\describe{
\item{\code{max}}{the max value of scores across multiple outputs is selected to get a single value for each observation}
\item{\code{average}}{scores of multiple outputs are averaged to get a single value for each observation}
\item{\code{sum}}{scores of multiple outputs are summed up to get a single value for each observation}
\item{\code{raw}}{returns the scores for the multiple outputs}
}}

\item{logger}{a \linkS4class{Logger}}

\item{...}{further arguments}
}
\value{
The p-value associated with each feature or NULL if the response type is not supported

a \linkS4class{Screened} object
}
\description{
This function compute the significance of the screening test by using an Empirical Bayes method.
}
\seealso{
\code{\link{permutation_screener}},
\code{\link{default_screener}}
}
\author{
Alessandro Barberis
}
