% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Evaluator.R
\name{evaluate,missing,ANY,Evaluator,numeric-method}
\alias{evaluate,missing,ANY,Evaluator,numeric-method}
\title{Evaluate Learning Method}
\usage{
\S4method{evaluate}{missing,ANY,Evaluator,numeric}(
  learner,
  evaluator,
  npoints = 3,
  ngrid,
  nmin = round(nrow(x)/2),
  x,
  looper = Looper(),
  logger,
  outdir = NULL,
  filename = "renoir",
  restore = T,
  rm.call = FALSE,
  rm.fit = FALSE,
  ...
)
}
\arguments{
\item{npoints}{number of sample sizes to consider (i.e. number of training set sizes in the grid)}

\item{ngrid}{(optional) grid of sample sizes to consider}

\item{nmin}{minimum sample size (i.e. number of samples to use for training)}

\item{rm.call}{logical, whether to remove the call from the models. Helpful if object size is expected to be big}

\item{rm.fit}{logical, whether to remove the model fits used for tuning the hyperparameters . Helpful if object size is expected to be big
object can be huge.}
}
\description{
Evaluate Learning Method
}
