% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Evaluator.R
\name{evaluate,missing,Learner,Evaluator,missing-method}
\alias{evaluate,missing,Learner,Evaluator,missing-method}
\title{Evaluate Learning Method}
\usage{
\S4method{evaluate}{missing,Learner,Evaluator,missing}(
  learner,
  evaluator,
  x,
  y,
  weights,
  offset,
  resp.type,
  observations,
  logger,
  outdir,
  filename,
  rm.call = T,
  rm.fit = F,
  ...
)
}
\arguments{
\item{observations}{(optional) index of samples to use for training.
If missing, observations will be sampled by using the \linkS4class{Sampler}
in \linkS4class{Evaluator}. This is an helper argument, and shouldn't be used
directly. Information of the sampling procedure in the output \linkS4class{Evaluated}
object are still retrieved from \linkS4class{Sampler} in \linkS4class{Evaluator} even if
\code{observation} is provided.}
}
\description{
Evaluate Learning Method
}
