% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Evaluator.R
\name{evaluate,missing,LearnerList,Evaluator,missing-method}
\alias{evaluate,missing,LearnerList,Evaluator,missing-method}
\title{Evaluate Learning Methods}
\usage{
\S4method{evaluate}{missing,LearnerList,Evaluator,missing}(learner, evaluator, x, looper = Looper(), logger, ...)
}
\arguments{
\item{looper}{a \linkS4class{Looper} to loop over the learning methods to evaluate}
}
\description{
Evaluate Learning Methods
}
