% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_generics.R, R/class_Evaluator.R
\name{evaluate}
\alias{evaluate}
\alias{evaluate,missing,Learner,Evaluator,missing-method}
\alias{evaluate,TrainedList,Learner,Evaluator,missing-method}
\alias{evaluate,TunedList,Learner,Evaluator,missing-method}
\alias{evaluate,missing,ANY,Evaluator,numeric-method}
\alias{evaluate,missing,LearnerList,Evaluator,missing-method}
\alias{evaluate,TrainedList,missing,Evaluator,missing-method}
\title{Evaluate Learning Method}
\usage{
evaluate(models, learner, evaluator, npoints, ...)

\S4method{evaluate}{missing,Learner,Evaluator,missing}(
  learner,
  evaluator,
  x,
  y,
  weights,
  offset,
  resp.type,
  observations,
  logger,
  outdir,
  filename,
  rm.call = T,
  rm.fit = F,
  ...
)

\S4method{evaluate}{TrainedList,Learner,Evaluator,missing}(
  models,
  learner,
  evaluator,
  x,
  y,
  weights,
  offset,
  resp.type,
  observations = NULL,
  logger,
  screened = NULL,
  config = character(),
  ...
)

\S4method{evaluate}{TunedList,Learner,Evaluator,missing}(models, learner, evaluator, npoints, ...)

\S4method{evaluate}{missing,ANY,Evaluator,numeric}(
  learner,
  evaluator,
  npoints = 3,
  ngrid,
  nmin = round(nrow(x)/2),
  x,
  looper = Looper(),
  logger,
  outdir = NULL,
  filename = "renoir",
  restore = T,
  rm.call = FALSE,
  rm.fit = FALSE,
  ...
)

\S4method{evaluate}{missing,LearnerList,Evaluator,missing}(learner, evaluator, x, looper = Looper(), logger, ...)

\S4method{evaluate}{TrainedList,missing,Evaluator,missing}(
  models,
  learner,
  evaluator,
  recorder,
  x,
  y,
  weights,
  offset,
  resp.type,
  samples = NULL,
  observations = NULL,
  logger,
  screened = NULL,
  config = character(),
  set = c("test", "train", "full"),
  min.obs = 0,
  auto.select = F,
  ...
)
}
\arguments{
\item{models}{an object of class \linkS4class{TrainedList}}

\item{learner}{an object of class \linkS4class{Learner}}

\item{evaluator}{an object of class \linkS4class{Evaluator}}

\item{npoints}{number of sample sizes to consider (i.e. number of training set sizes in the grid)}

\item{...}{further arguments to \code{learn} function}

\item{x}{the input matrix, where rows are observations and columns are variables.}

\item{y}{the response variable. Its number of rows must match the number of rows of \code{x}.}

\item{weights}{(optional) vector of observation weights}

\item{offset}{vector containing an offset, used for linear models. Default is \code{NULL}.}

\item{resp.type}{the response type}

\item{observations}{(optional) index of samples to use for training.
If missing, observations will be sampled by using the \linkS4class{Sampler}
in \linkS4class{Evaluator}. This is an helper argument, and shouldn't be used
directly. Information of the sampling procedure in the output \linkS4class{Evaluated}
object are still retrieved from \linkS4class{Sampler} in \linkS4class{Evaluator} even if
\code{observation} is provided.}

\item{logger}{an object of class \linkS4class{Logger}}

\item{outdir}{path to the output directory}

\item{filename}{(optional) name without extension for the output file}

\item{rm.call}{logical, whether to remove the call from the models. Helpful if object size is expected to be big}

\item{rm.fit}{logical, whether to remove the model fits used for tuning the hyperparameters . Helpful if object size is expected to be big
object can be huge.}

\item{ngrid}{(optional) grid of sample sizes to consider}

\item{nmin}{minimum sample size (i.e. number of samples to use for training)}

\item{looper}{a \linkS4class{Looper} to loop over the learning methods to evaluate}
}
\description{
Evaluate Learning Method

Evaluate Learning Method

Evaluate Learning Methods
}
