% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Evaluated.R
\docType{class}
\name{Evaluated-class}
\alias{Evaluated-class}
\title{Evaluator Class}
\description{
An S4 class representing an evaluation of a learning method.
}
\section{Slots}{

\describe{
\item{\code{id}}{evaluated learning method}

\item{\code{config}}{(optional) name of the configuration}

\item{\code{response}}{response type}

\item{\code{nout}}{a length-one integer vector}

\item{\code{sampling}}{sampling method used during evaluation}

\item{\code{N}}{population size}

\item{\code{n}}{sample size (i.e. training set size)}

\item{\code{k}}{the number of folds if \code{sampling = "cv"},
or the number of repeats if \code{sampling = "random"} and \code{sampling = "bootstrap"}}

\item{\code{observations}}{indices of sampled observations used for training}

\item{\code{models}}{list of Tuned or Trained elements}

\item{\code{stability}}{list containing the features stability for each response across the models}

\item{\code{performance}}{a \linkS4class{TestedList} object}
}}

\author{
Alessandro Barberis
}
