% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Evaluated.R
\name{Evaluated}
\alias{Evaluated}
\title{Constructor for the S4 Evaluated object.}
\usage{
Evaluated(
  id = character(),
  config = character(),
  response = character(),
  nout = integer(),
  sampling = character(),
  N = integer(),
  n = integer(),
  k = integer(),
  observations = list(),
  models = list(),
  stability = list(),
  performance = TestedList()
)
}
\arguments{
\item{id}{evaluated learning method}

\item{config}{(optional) name of the configuration}

\item{nout}{number of outcomes}

\item{sampling}{sampling method used during evaluation}

\item{N}{population size}

\item{n}{sample size (i.e. training set size)}

\item{k}{the number of folds if \code{sampling = "cv"},
or the number of repeats if \code{sampling = "random"} and \code{sampling = "bootstrap"}}

\item{observations}{indices of sampled observations used for training}

\item{models}{list of Tuned or Trained elements}

\item{stability}{list containing the features stability for each response across the models}

\item{performance}{a \linkS4class{TestedList} object}
}
\value{
An object of class \linkS4class{Evaluated}.
}
\description{
Constructor for the S4 \linkS4class{Evaluated} object.
}
\author{
Alessandro Barberis
}
\keyword{internal}
