% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_EvaluatedList.R
\name{EvaluatedList}
\alias{EvaluatedList}
\title{EvaluatedList Constructor}
\usage{
EvaluatedList(...)
}
\description{
Constructor for the S4 EvaluatedList object.
}
\details{
Constructor for the S4 \linkS4class{EvaluatedList} object.
}
