% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Evaluator.R
\docType{class}
\name{Evaluator-class}
\alias{Evaluator-class}
\title{Evaluator Class}
\description{
An S4 class providing the methods to perform an evaluation of a learning method.

The object consists of 3 slots
}
\section{Slots}{

\describe{
\item{\code{sampler}}{a \linkS4class{Sampler} object}

\item{\code{looper}}{a \linkS4class{Looper} object}

\item{\code{scorer}}{a \linkS4class{ScorerList} object}

\item{\code{logger}}{a \linkS4class{Logger} object}
}}

\author{
Alessandro Barberis
}
