% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Evaluator.R
\name{Evaluator}
\alias{Evaluator}
\title{Evaluator Class Constructor}
\usage{
Evaluator(
  sampler = Sampler(method = "random", k = 1L),
  looper = Looper(),
  scorer = ScorerList(Scorer(id = "mse")),
  logger = Logger()
)
}
\arguments{
\item{sampler}{a \linkS4class{Sampler}}

\item{looper}{a \linkS4class{Looper}}

\item{scorer}{a \linkS4class{ScorerList} or a \linkS4class{Scorer} object}

\item{logger}{a \linkS4class{Logger}}
}
\value{
An object of class \linkS4class{Evaluator}.
}
\description{
Constructor for the S4 \linkS4class{Evaluator} object.
}
\author{
Alessandro Barberis
}
