% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma.R
\name{extract_pval_limma}
\alias{extract_pval_limma}
\title{Extract adjusted p-value from limma result}
\usage{
extract_pval_limma(obj, coef = NULL, feat.order = NULL, pval = "adj.P.Val")
}
\arguments{
\item{obj}{object returned from \code{\link{compute_limma}}}

\item{coef}{column number or column name specifying which coefficient or contrast of the linear
model is of interest. For topTable, can also be a vector of column subscripts, in which
case the gene ranking is by F-statistic for that set of contrasts.
If coef has a single value, then the moderated t-statistics and p-values for that coefficient
or contrast are used.
If coef takes two or more values, the moderated F-statistics for that set of coefficients or
contrasts are used. If coef is left NULL, then all the coefficients or contrasts in the fitted
model are used, except that any coefficient named (Intercept) will be removed.}
}
\value{
the adjusted p-values of the test
}
\description{
Extract adjusted p-value from limma result
}
\keyword{internal}
