% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam.R
\name{extract_qval_sam_default}
\alias{extract_qval_sam_default}
\title{Extract q-value from SAM}
\usage{
extract_qval_sam_default(obj, feat.order = NULL)
}
\arguments{
\item{obj}{object as returned by \code{\link{sam}}}

\item{feat.order}{(optional) order of output features}
}
\value{
q-value
}
\description{
Extract q-value from SAM
}
\keyword{internal}
