% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{f1_score}
\alias{f1_score}
\title{F1 Score}
\usage{
f1_score(true, pred, weights = NULL, multi, ...)
}
\arguments{
\item{true}{a vector (or a matrix) of observed values. If a matrix is provided,
a multi-response is assumed}

\item{pred}{a vector (or a matrix) of predicted values}

\item{weights}{observation weights (not implemented yet)}

\item{multi}{what to do when response has multiple output values
\describe{
\item{\code{average}}{errors of multiple outputs are averaged to get a single value for each observation}
\item{\code{micro}}{the score is computed by using the global number of true positives, false negatives and false positives}
\item{\code{macro}}{scores of different classes are averaged by unweighted mean to get a single value. Class imbalance is not taken into account}
\item{\code{weighted}}{scores of different classes are averaged by weighted mean to get a single value. Weights are number of true instances per class. It takes into account label imbalance}
\item{\code{raw}}{returns a vector containing one score for each class}
\item{\code{binary}}{returns the score for the class specified by \code{positive}}
}}

\item{...}{further arguments to \code{\link{multiresponse_classification_metric}}}
}
\value{
A single score for the selected class if \code{multi = "binary"},
a vector containing one score for each class if \code{multi = "raw"},
a summary score computed by using the global metrics if \code{multi = "micro"} or
averaging the results from different classes if \code{multi = "macro"}. A summary score
produced by a weighted average of the results from different class is produced if
\code{multi = "weighted"}
}
\description{
This function computes the F1 score.
}
\details{
The F1 score is a measure of accuracy and is computed as
the harmonic mean of the precision and recall:

\deqn{F_{1} = \frac{2}{precision^{-1} + recall^{-1}} = 2\frac{precision \times recall}{precision + recall} = \frac{TP}{TP + \frac{1}{2}(FP + FN)}}

The highest possible value is 1.0, indicating perfect precision and recall. The lowest possible value is 0.

It is a specific case of the F-beta score, where \code{beta = 1}
}
\references{
\url{https://en.wikipedia.org/wiki/F-score}
}
\seealso{
\code{\link{fbeta_score}}
}
\author{
Alessandro Barberis
}
