% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{F1S}
\alias{F1S}
\title{F1 score}
\usage{
F1S(TP, TN, FP, FN)
}
\arguments{
\item{TP}{number of true positives}

\item{TN}{number of true negatives}

\item{FP}{number of false positives}

\item{FN}{number of false negatives}
}
\value{
A numeric value.
}
\description{
This function computes the F1 score.
}
\author{
Alessandro Barberis
}
