% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{factor.to.dummy.matrix}
\alias{factor.to.dummy.matrix}
\title{From factor to dummy matrix}
\usage{
factor.to.dummy.matrix(factor)
}
\arguments{
\item{factor}{a \code{factor}}
}
\value{
a dummy matrix
}
\description{
From factor to dummy matrix
}
\keyword{internal}
