% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{FBS}
\alias{FBS}
\title{Fbeta score}
\usage{
FBS(TP, TN, FP, FN, beta = 1)
}
\arguments{
\item{TP}{number of true positives}

\item{TN}{number of true negatives}

\item{FP}{number of false positives}

\item{FN}{number of false negatives}
}
\value{
A numeric value.
}
\description{
This function computes the F-beta score.
}
\author{
Alessandro Barberis
}
