% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_generics.R, R/class_Recorder.R
\name{features}
\alias{features}
\alias{features,Trained,Recorder-method}
\alias{features,Trained,missing-method}
\alias{features,Tuned,Recorder-method}
\alias{features,Tuned,missing-method}
\title{Get features in the model}
\usage{
features(object, recorder, ...)

\S4method{features}{Trained,Recorder}(object, recorder, type = c("all", "nonzero"), ...)

\S4method{features}{Trained,missing}(object, recorder, ...)

\S4method{features}{Tuned,Recorder}(object, recorder, ...)

\S4method{features}{Tuned,missing}(object, recorder, ...)
}
\description{
Get features in the model
}
