% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Filter.R
\docType{class}
\name{Filter-class}
\alias{Filter-class}
\title{Filter Class}
\description{
An S4 class representing a filter to pre-process the input data.

The object consists of 3 slots
}
\section{Slots}{

\describe{
\item{\code{id}}{a name to identify the filter}

\item{\code{filter}}{filter function}

\item{\code{parameters}}{list containing the parameters for the chosen filtering method}

\item{\code{logger}}{a \linkS4class{Logger}}
}}

\author{
Alessandro Barberis
}
