% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Filter.R
\name{filter,Filter-method}
\alias{filter,Filter-method}
\title{Unsupervised Screening}
\usage{
\S4method{filter}{Filter}(filter, x, logger, ...)
}
\arguments{
\item{filter}{an object of class \linkS4class{Filter}, providing the function
performing the unsupervised screening}

\item{x}{the input matrix, where rows are observations and columns are variables}

\item{logger}{an object of class \linkS4class{Logger}}

\item{...}{further arguments to the function performing the unsupervised screening}
}
\value{
An object of class \linkS4class{Filtered}.
}
\description{
This function filters the data in input by unsupervised screening.
}
\seealso{
\code{\link{filter_by_na}},
\code{\link{filter_by_intensity}},
\code{\link{filter_by_variability}}
}
\author{
Alessandro Barberis
}
