% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_generics.R, R/class_Filter.R,
%   R/class_FilterList.R
\name{filter}
\alias{filter}
\alias{filter,Filter-method}
\alias{filter,FilterList-method}
\title{Unsupervised Screening}
\usage{
filter(filter, ...)

\S4method{filter}{Filter}(filter, x, logger, ...)

\S4method{filter}{FilterList}(filter, x, na.omit = F, logger = Logger(), ...)
}
\arguments{
\item{filter}{an object of class \linkS4class{Filter}, providing the function
performing the unsupervised screening}

\item{...}{further arguments to the function performing the unsupervised screening}

\item{x}{the input matrix, where rows are observations and columns are variables}

\item{logger}{an object of class \linkS4class{Logger}}
}
\value{
An object of class \linkS4class{Filtered}.
}
\description{
This function filters the data in input by unsupervised screening.
}
\seealso{
\code{\link{filter_by_na}},
\code{\link{filter_by_intensity}},
\code{\link{filter_by_variability}}
}
\author{
Alessandro Barberis
}
