% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Filter.R
\name{filter_by_intensity}
\alias{filter_by_intensity}
\title{Filter by intensity}
\usage{
filter_by_intensity(x, threshold = 0.25, logger)
}
\arguments{
\item{x}{the input matrix, where rows are observations and columns are variables.}

\item{threshold}{the cutoff}

\item{logger}{a \linkS4class{Logger}}
}
\value{
A \linkS4class{Filtered} object
}
\description{
This function filters the input matrix \code{x} depending on the features intensity.
A variable is kept if value is greater than median across observations in a given percentage of observations defined by \code{threshold}.
Default is 0.25 (i.e. select variables with values > median in more than 25\\% of observations)
}
\author{
Alessandro Barberis
}
