% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Filter.R
\name{filter_by_na}
\alias{filter_by_na}
\title{Filter by missing values}
\usage{
filter_by_na(x, threshold = 0.5, logger)
}
\arguments{
\item{x}{the input matrix, where rows are observations and columns are variables.}

\item{threshold}{the cutoff}

\item{logger}{a \linkS4class{Logger}}
}
\value{
A \linkS4class{Filtered} object
}
\description{
This function filters the input matrix \code{x} depending on the presence of NAs.
A variable is removed if sum of NA elements is greater than a given percentage defined by \code{threshold}.
Default is 0.5 (i.e. select variables with less than 50\\% of NAs)
}
\author{
Alessandro Barberis
}
