% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Filter.R
\name{filter_by_variability}
\alias{filter_by_variability}
\title{Filter by variability}
\usage{
filter_by_variability(
  x,
  threshold = 0.25,
  logger,
  method = c("sd", "IQR", "mad")
)
}
\arguments{
\item{x}{the input matrix, where rows are observations and columns are variables.}

\item{threshold}{the cutoff}

\item{logger}{a \linkS4class{Logger}}

\item{method}{the measure of variability, i.e. the standard deviation (\code{sd}),
the interquartile range (\code{IQR}, the median absolute deviation (\code{mad}))}
}
\value{
A \linkS4class{Filtered} object
}
\description{
This function filters the input matrix \code{x} by features variability.
A feature is kept if part of the top variable features, where the cutoff is defined by \code{threshold}.
Default is 0.25 (i.e. select features in the top 25\% of variable features).
Two measures of variability are possible.
}
\author{
Alessandro Barberis
}
\keyword{internal}
