% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Filtered.R
\docType{class}
\name{Filtered-class}
\alias{Filtered-class}
\title{Filtered Class
An S4 class providing a container for the results of pre-processing.}
\description{
Filtered Class
An S4 class providing a container for the results of pre-processing.
}
\section{Slots}{

\describe{
\item{\code{method}}{a character string}

\item{\code{filtered}}{the filtered data}

\item{\code{index}}{vector containing the index of the retained features in the original matrix}

\item{\code{summary}}{a text summary of the pre-processing}
}}

\author{
Alessandro Barberis
}
