% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_FilterList.R
\docType{class}
\name{FilterList-class}
\alias{FilterList-class}
\alias{FilterList}
\title{FilterList Class}
\usage{
FilterList(...)
}
\arguments{
\item{...}{\linkS4class{Filter} objects}
}
\value{
a \linkS4class{FilterList} object
}
\description{
An S4 class to represent a list of filters to pre-process the data.

Constructor for the S4 FilterList object.
}
\details{
This subclass extends the S4  \linkS4class{List} virtual class by using
the  \linkS4class{SimpleList} implementation.

The object consists of 4 slots

Constructor for the S4 \linkS4class{FilterList} object.
}
\section{Slots}{

\describe{
\item{\code{listData}}{\code{list} storing the list elements}

\item{\code{elementType}}{the type of data represented in the sequence}

\item{\code{elementMetadata}}{\linkS4class{DataFrame} storing the element-wise metadata,
with a row for each element and a column for each metadata available. Default is \code{NULL}}

\item{\code{metadata}}{\code{list} storing the global metadata annotating the object as a whole}
}}

