% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_FilterList.R
\name{FilterList}
\alias{FilterList}
\title{FilterList Constructor}
\usage{
FilterList(...)
}
\arguments{
\item{...}{\linkS4class{Filter} objects}
}
\value{
a \linkS4class{FilterList} object
}
\description{
Constructor for the S4 FilterList object.
}
\details{
Constructor for the S4 \linkS4class{FilterList} object.
}
