% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{FNR}
\alias{FNR}
\title{False negative rate, miss rate, (1 - TPR)}
\usage{
FNR(TP, TN, FP, FN)
}
\arguments{
\item{TP}{number of true positives}

\item{TN}{number of true negatives}

\item{FP}{number of false positives}

\item{FN}{number of false negatives}
}
\value{
A numeric value.
}
\description{
This function computes the false negative rate.
}
\author{
Alessandro Barberis
}
