% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_generics.R, R/class_Forecaster.R,
%   R/class_ForecasterList.R
\name{forecast}
\alias{forecast}
\alias{forecast,Forecaster,Trained-method}
\alias{forecast,missing,Trained-method}
\alias{forecast,Forecaster,TrainedList-method}
\alias{forecast,Forecaster,Tuned-method}
\alias{forecast,missing,Tuned-method}
\alias{forecast,Forecaster,TunedList-method}
\alias{forecast,ForecasterList,Trained-method}
\alias{forecast,ForecasterList,TrainedList-method}
\title{Forecast method}
\usage{
forecast(forecaster, models, ...)

\S4method{forecast}{Forecaster,Trained}(forecaster, models, newx, type, newoffset, check.newx = F, ...)

\S4method{forecast}{missing,Trained}(forecaster, models, newx, type, newoffset, ...)

\S4method{forecast}{Forecaster,TrainedList}(forecaster, models, newx, type, newoffset, ...)

\S4method{forecast}{Forecaster,Tuned}(forecaster, models, newx, type, newoffset, ...)

\S4method{forecast}{missing,Tuned}(forecaster, models, newx, type, newoffset, ...)

\S4method{forecast}{Forecaster,TunedList}(forecaster, models, newx, type, newoffset, ...)

\S4method{forecast}{ForecasterList,Trained}(forecaster, models, ...)

\S4method{forecast}{ForecasterList,TrainedList}(forecaster, models, ...)
}
\arguments{
\item{forecaster}{an object of class \linkS4class{Forecaster}}

\item{models}{an object of class \code{\link{Trained}}}

\item{...}{further arguments to the forecaster}

\item{newx}{the new values at which prediction are to be made}

\item{type}{the type of prediction}

\item{newoffset}{the new offset, if provided it is added to the predicted value}

\item{check.newx}{logical, whether to check the \code{newx} to make sure
features are matching}
}
\value{
The prediction
}
\description{
Forecast method
}
