% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Forecaster.R
\name{forecast_by_gbm}
\alias{forecast_by_gbm}
\title{Forecaster for GBM Model Fits}
\usage{
forecast_by_gbm(
  object,
  newx,
  type = c("response", "class", "link"),
  newoffset,
  pthr = 0.5,
  ...
)
}
\arguments{
\item{object}{an object of class \code{\link[gbm]{gbm.object}}}

\item{newx}{the new values at which prediction are to be made}

\item{type}{the type of prediction. Three options are available
\describe{
\item{\code{"response"}}{predicted values are on the same scale as the outcome,
e.g. for Bernoulli the returned values are probabilities, for Poisson the expected counts}
\item{\code{"class"}}{the class corresponding to the max probability}
\item{\code{"link"}}{predictions are on the scale of f(x), e.g. for Bernoulli
the returned value is on the log odds scale}
}}

\item{newoffset}{the new offset, if provided it is added to the predicted value}

\item{pthr}{probability threshold, used when \code{type = "class"} and response
is binomial to assign label to observation}

\item{...}{further arguments to \code{\link[gbm]{predict.gbm}}}
}
\value{
A vector containing the predictions.
}
\description{
This function predicts values based on generalized
boosted model objects.
}
\author{
Alessandro Barberis
}
