% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Forecaster.R
\name{forecast_by_gknn}
\alias{forecast_by_gknn}
\title{Forecaster for Generalized kNN Model Fits}
\usage{
forecast_by_gknn(
  object,
  newx,
  type = c("response", "class", "prob", "votes"),
  newoffset,
  ...,
  target = 2L
)
}
\arguments{
\item{object}{an object of class \linkS4class{Trained}}

\item{newx}{the new values at which prediction are to be made}

\item{type}{character specifying the type of prediction. Three options are available
\describe{
\item{\code{"response"}}{predicted values are on the same scale as the outcome,
e.g. for binomial the returned values are classes, for gaussian the expected values}
\item{\code{"class"}}{the class labels corresponding to the max probability}
\item{\code{"prob"}}{the class distribution}
\item{\code{"votes"}}{the raw counts for each class}
}}

\item{newoffset}{currently not used}

\item{...}{further arguments to \code{\link[e1071]{predict.gknn}}}

\item{target}{(optional) integer, index of the target class. If provided, only the
prediction for the selected class are returned (default to second level
of factor response). Set to \code{NULL} to return all classes}
}
\value{
A vector containing the predictions.
}
\description{
This function predicts values based on Generalized k-Nearest
Neighbours model objects.
}
\seealso{
\code{\link[e1071]{predict.gknn}}
}
\author{
Alessandro Barberis
}
\keyword{internal}
