% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Forecaster.R
\name{forecast_by_glmnet}
\alias{forecast_by_glmnet}
\title{Forecaster for GLM Elastic-Net Fits}
\usage{
forecast_by_glmnet(
  object,
  newx,
  type = c("link", "response", "coefficients", "class", "nonzero"),
  newoffset,
  s,
  lambda,
  gamma,
  ...
)
}
\arguments{
\item{object}{an object of class \code{\link{Trained}}}

\item{newx}{Matrix of new values for \code{x} at which predictions are to be
made. Must be a matrix; can be sparse as in \code{Matrix} package. This
argument is not used for \code{type=c("coefficients","nonzero")}}

\item{type}{Type of prediction required. Type \code{"link"} gives the linear
predictors for \code{"binomial"}, \code{"multinomial"}, \code{"poisson"} or
\code{"cox"} models; for \code{"gaussian"} models it gives the fitted
values. Type \code{"response"} gives the fitted probabilities for
\code{"binomial"} or \code{"multinomial"}, fitted mean for \code{"poisson"}
and the fitted relative-risk for \code{"cox"}; for \code{"gaussian"} type
\code{"response"} is equivalent to type \code{"link"}. Type
\code{"coefficients"} computes the coefficients at the requested values for
\code{s}.  Note that for \code{"binomial"} models, results are returned only
for the class corresponding to the second level of the factor response.
Type \code{"class"} applies only to \code{"binomial"} or
\code{"multinomial"} models, and produces the class label corresponding to
the maximum probability. Type \code{"nonzero"} returns a list of the indices
of the nonzero coefficients for each value of \code{s}.}

\item{newoffset}{If an offset is used in the fit, then one must be supplied
for making predictions (except for \code{type="coefficients"} or
\code{type="nonzero"})}

\item{s}{Value(s) of the penalty parameter \code{lambda} at which
predictions are required. Default is the entire sequence used to create the
model.}

\item{lambda}{(optional) value of the penalty parameter. Used if \code{s} is
not provided}

\item{gamma}{Single value of \code{gamma} at which predictions are required,
for "relaxed" objects.}

\item{...}{further arguments to \code{\link[glmnet]{predict.glmnet}}}
}
\value{
A vector containing the predictions.
}
\description{
This function predicts values based on glmnet
model objects.
}
\seealso{
\code{\link[glmnet]{predict.glmnet}}
}
\author{
Alessandro Barberis
}
