% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renoir_nearest_shrunken_centroid.R
\name{forecast_by_nsc}
\alias{forecast_by_nsc}
\title{Forecaster for Nearest Shrunken Centroid Model Fits}
\usage{
forecast_by_nsc(
  object,
  newx,
  type = c("response", "class", "posterior", "centroid", "nonzero"),
  newoffset,
  ...,
  threshold,
  target = 2L
)
}
\arguments{
\item{object}{an object of class \linkS4class{Trained}}

\item{newx}{the new values at which prediction are to be made}

\item{type}{character specifying the type of prediction. Three options are available
\describe{
\item{\code{"response"}}{predicted values are on the same scale as the outcome,
e.g. for binomial the returned values are classes}
\item{\code{"class"}}{the class labels corresponding to the max probability}
\item{\code{"posterior"}}{posterior probabilities}
\item{\code{"centroid"}}{(unshrunken) class centroids}
\item{\code{"nonzero"}}{index of genes surviving the threshold}
}}

\item{newoffset}{currently not used}

\item{...}{further arguments to \code{\link[pamr]{pamr.predict}}}

\item{threshold}{The desired threshold value}

\item{target}{(optional) integer, index of the target class. If provided, only the
prediction for the selected class are returned (default to second level
of factor response). Set to \code{NULL} to return all classes. Used when
\code{type = "posterior"} or \code{type = "centroid"}}
}
\value{
A vector containing the predictions.
}
\description{
This function predicts values based on Nearest Shrunken
Centroid model objects.
}
\seealso{
\code{\link[pamr]{pamr.predict}}
}
\author{
Alessandro Barberis
}
\keyword{internal}
