% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Forecaster.R
\name{forecast_by_randomForest}
\alias{forecast_by_randomForest}
\title{Forecaster for Random Forest Model Fits}
\usage{
forecast_by_randomForest(
  object,
  newx,
  type = c("response", "prob", "votes"),
  newoffset = NULL,
  target = 2L,
  ...
)
}
\arguments{
\item{object}{an object of class \code{\link{Trained}}}

\item{newx}{the new values at which prediction are to be made}

\item{type}{the type of prediction. Three options are available
\describe{
\item{\code{"response"}}{predicted values are on the same scale as the outcome}
\item{\code{"prob"}}{matrix of class probabilities}
\item{\code{"votes"}}{matrix of vote counts}
}}

\item{newoffset}{the new offset, if provided it is added to the predicted value}

\item{target}{(optional) integer, index of the target class. If provided, only the
prediction for the selected class are returned (default to second level
of factor response). Set to \code{NULL} to return all classes}
}
\value{
A vector containing the predictions.
}
\description{
This function predicts values based on random forest
model objects.
}
\seealso{
\code{\link[randomForest]{predict.randomForest}}
}
\author{
Alessandro Barberis
}
\keyword{internal}
