% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Forecaster.R
\name{forecast_by_svm}
\alias{forecast_by_svm}
\title{Forecaster for SVM Model Fits}
\usage{
forecast_by_svm(
  object,
  newx,
  type = c("response", "class"),
  newoffset,
  pthr = 0.5,
  ...
)
}
\arguments{
\item{object}{an object of class \linkS4class{Trained}}

\item{newx}{the new values at which prediction are to be made}

\item{type}{character specifying the type of prediction. Three options are available
\describe{
\item{\code{"response"}}{predicted values are on the same scale as the outcome,
e.g. for binomial the returned values are classes, for gaussian the expected values}
\item{\code{"class"}}{the class corresponding to the max probability}
}}

\item{newoffset}{currently not used}

\item{pthr}{probability threshold, used when \code{type = "class"} and response
is binomial to assign label to observation}

\item{...}{further arguments to \code{\link[e1071]{predict.svm}}}
}
\value{
A vector containing the predictions.
}
\description{
This function predicts values based on Support Vector Machines model objects.
}
\details{
It seems there is an issue with the underlying function when asking for
probabilities
}
\seealso{
\code{\link[e1071]{predict.svm}}
}
\author{
Alessandro Barberis
}
