% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ForecasterList.R
\name{ForecasterList}
\alias{ForecasterList}
\title{Constructor for the S4 ForecasterList object.

Constructor for the S4 \linkS4class{ForecasterList} object.}
\usage{
ForecasterList(...)
}
\description{
ForecasterList Constructor
}
