% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Looper.R
\name{forloop}
\alias{forloop}
\title{Loop}
\usage{
forloop(n.iter, logger, cores = 1, .inorder = TRUE, fun, ...)
}
\arguments{
\item{n.iter}{number of iterations of the loop}

\item{logger}{a \linkS4class{Logger}}

\item{cores}{number of cores to use for parallel execution.}

\item{.inorder}{logical flag indicating whether the \code{.combine}
function requires the task results to be combined in the same order
that they were submitted.  If the order is not important, then it
setting \code{.inorder} to \code{FALSE} can give improved performance.
The default value is `TRUE.}

\item{fun}{A function to be executed inside the loop. It should contain an \code{iloop} parameter.}

\item{...}{further arguments to \code{fun}}
}
\description{
Common interface to run a for loop sequentially or in parallel.
The core of the loop is defined by the function \code{fun}.
}
\keyword{internal}
