% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_score}
\alias{format_score}
\title{Format score}
\usage{
format_score(
  x,
  scientific = TRUE,
  digits = 3,
  cutoff = 0.05,
  cutpoints = c(0, 1e-04, 0.001, 0.01, 0.05, 0.1, 1),
  symbols = c("****", "***", "**", "*", ".", "ns"),
  score = "p",
  name = NULL
)
}
\arguments{
\item{x}{vector or data.frame containing the data of interest (e.g. p-values).
If a data.frame, it must contain a \code{score} column.}

\item{scientific}{Either a logical specifying whether
    elements of a real or complex vector should be encoded in scientific
    format, or an integer penalty (see \code{\link[base]{options}("scipen")}).
    Missing values correspond to the current default penalty.}

\item{digits}{a positive integer indicating how many significant digits
    are to be used for
    numeric and complex \code{x}.  The default, \code{NULL}, uses
    \code{\link[base]{getOption}("digits")}.  This is a suggestion: enough decimal
    places will be used so that the smallest (in magnitude) number has
    this many significant digits, and also to satisfy \code{nsmall}.
    (For more, notably the interpretation for complex numbers see
    \code{\link[base]{signif}}.)}

\item{cutoff}{numerical cutoff for adding the symbol to the
\code{string} column}

\item{score}{score name to use in the \code{string} column}

\item{name}{(optional) name of the method used to obtain the score,
to use in \code{string}}
}
\value{
a \code{data.frame} with 4 columns
\describe{
\item{score}{contains the score as provided in input}
\item{score_f}{formatted scores}
\item{score_sym}{scores as symbols, as returned by \link{num_to_sym}}
\item{string}{scores as string, e.g. 'p = 0.005 **'}
}
}
\description{
Format score
}
\keyword{internal}
