% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{FPR}
\alias{FPR}
\title{False positive rate, probability of false alarm, fall-out, (1 - TNR)}
\usage{
FPR(TP, TN, FP, FN)
}
\arguments{
\item{TP}{number of true positives}

\item{TN}{number of true negatives}

\item{FP}{number of false positives}

\item{FN}{number of false negatives}
}
\value{
A numeric value.
}
\description{
This function computes the false positive rate.
}
\author{
Alessandro Barberis
}
