% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{get_available_measures}
\alias{get_available_measures}
\title{Available accuracy measures}
\usage{
get_available_measures(resp.type)
}
\arguments{
\item{resp.type}{the response type}
}
\value{
Character vector containing the available accuracy measures for the selected \code{resp.type}.
}
\description{
Get the available performance measures for the selected response type.
}
\author{
Alessandro Barberis
}
\keyword{internal}
