% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Sampler.R
\name{get_balanced_training_set_size_grid}
\alias{get_balanced_training_set_size_grid}
\title{Allowed Training set sizes}
\usage{
get_balanced_training_set_size_grid(strata, from, to)
}
\arguments{
\item{from}{Starting point of the sequence}

\item{to}{the maximum allowed size of a training set}

\item{y}{response variable}
}
\value{
a sequence of allowed sizes
}
\description{
The function computes a sequence of allowed training set size for
a logistic response variable based on the following rule:
given nc classes, training set (n) must include the same number of patients for each class.
}
\keyword{internal}
