% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruit.R
\name{get_features_importance}
\alias{get_features_importance}
\title{Features Importance}
\usage{
get_features_importance(
  object,
  recruitment.term = c("coefficient", "presence")
)
}
\arguments{
\item{object}{a list of \code{renoir.tat} objects}

\item{recruitment.term}{whether to use a binary variable (\code{presence = 0} or \code{presence = 1})
multiplied by the sign of the coefficient, or to use directly the coefficient of a feature
in the computation of the significance score.}
}
\value{
Features Importance
}
\description{
This function computes a significance score for each feature.
See \code{\link{features_importance}} for more details.
}
