% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_filename_results}
\alias{get_filename_results}
\title{Create filename for final output}
\usage{
get_filename_results(resampling.method, learning.method)
}
\value{
a generated filename for the final output
}
\description{
Create filename for final output
}
\keyword{internal}
