% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Logger.R
\name{get_log_levels}
\alias{get_log_levels}
\title{Get log levels}
\usage{
get_log_levels(level = c("OFF", "INFO", "DEBUG", "TRACE", "ALL"))
}
\arguments{
\item{level}{a log level}
}
\value{
A vector containing all the levels corresponding to \code{level}
}
\description{
Returns all the log levels included in the selected level
}
\author{
Alessandro Barberis
}
\keyword{internal}
