% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{get_measure_name}
\alias{get_measure_name}
\title{Accuracy measure name}
\usage{
get_measure_name(
  type.measure = c("mse", "mdev", "mclass", "auc", "mae", "C", "rmse", "mape", "r2",
    "msle", "squared_error", "deviance", "class", "absolute_error", "ape", "sle", "f1s",
    "acc", "precision", "sensitivity", "jaccard"),
  resp.type = NULL
)
}
\arguments{
\item{type.measure}{the accuracy measure}

\item{resp.type}{the response type}
}
\value{
Character string containing the accuracy measure name for the selected \code{resp.type}.
}
\description{
Get the accuracy measure name given the response type.
}
\author{
Alessandro Barberis
}
\keyword{internal}
