% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Evaluated.R
\name{get_model,Evaluated-method}
\alias{get_model,Evaluated-method}
\title{Get Trained/Tuned Model}
\usage{
\S4method{get_model}{Evaluated}(object, index)
}
\arguments{
\item{object}{\linkS4class{Evaluated} object}

\item{index}{index of model in the list}

\item{n}{sample size (i.e. training set size) to select}
}
\value{
a \linkS4class{Trained} or \linkS4class{Tuned} object
}
\description{
This function extracts a model from a
\linkS4class{Evaluated} object.
}
\author{
Alessandro Barberis
}
