% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_EvaluatedList.R
\name{get_model,EvaluatedList-method}
\alias{get_model,EvaluatedList-method}
\title{Get Trained/Tuned Model}
\usage{
\S4method{get_model}{EvaluatedList}(object, n, index)
}
\arguments{
\item{object}{\linkS4class{EvaluatedList} object}

\item{n}{sample size (i.e. training set size) to select}

\item{index}{index of model in the list}
}
\value{
a \linkS4class{Trained} object
}
\description{
This function extracts a model from a
\linkS4class{EvaluatedList} object.
}
\author{
Alessandro Barberis
}
