% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Renoir.R
\name{get_model,Renoir-method}
\alias{get_model,Renoir-method}
\title{Get Trained/Tuned Model}
\usage{
\S4method{get_model}{Renoir}(object, n, index)
}
\arguments{
\item{object}{an object of class \linkS4class{Renoir}}

\item{n}{sample size (i.e. training set size) to select}

\item{index}{index of model to select}
}
\value{
a \linkS4class{Trained} or \linkS4class{Tuned} object
}
\description{
This function extracts a model from a
\linkS4class{Renoir} object.
}
\author{
Alessandro Barberis
}
