% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{get_model_name}
\alias{get_model_name}
\title{Model name}
\usage{
get_model_name(resp.type)
}
\arguments{
\item{resp.type}{the response type}
}
\value{
Character string containing the model name.
}
\description{
Get the model name
}
\author{
Alessandro Barberis
}
\keyword{internal}
