% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_nobs}
\alias{get_nobs}
\title{Get the number of observations}
\usage{
get_nobs(object, na.rm = FALSE)
}
\arguments{
\item{na.rm}{logical, whether to remove NAs. If x is response variable
with multiple output values, an observation is removed if NA
in all responses.}

\item{x}{the predictors matrix or the response variable}
}
\value{
the number of observations
}
\description{
Get the number of observations
}
\keyword{internal}
