% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_num_of_classes}
\alias{get_num_of_classes}
\title{Get the number of classes}
\usage{
get_num_of_classes(y, resp.type)
}
\arguments{
\item{y}{response variable}

\item{resp.type}{response type}
}
\description{
Get the number of classes
}
\keyword{internal}
