% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_num_of_observations}
\alias{get_num_of_observations}
\title{Get the number of observations}
\usage{
get_num_of_observations(y, resp.type)
}
\arguments{
\item{y}{response variable}

\item{resp.type}{response type}
}
\description{
Get the number of observations
}
\keyword{internal}
