% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_num_of_responses}
\alias{get_num_of_responses}
\title{Get the number of responses}
\usage{
get_num_of_responses(y, resp.type)
}
\arguments{
\item{y}{response variable}

\item{resp.type}{response type}
}
\description{
Get the number of responses
}
\keyword{internal}
