% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Trained.R
\name{get_predict_param}
\alias{get_predict_param}
\title{Get parameters for predict method from Trained object}
\usage{
get_predict_param(object, ...)
}
\value{
a list of parameters to use for predict method. The list can be empty
if no parameter was found in input data
}
\description{
Get parameters for predict method from Trained object
}
\keyword{internal}
