% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Forecaster.R
\name{get_prediction_type,Forecaster-method}
\alias{get_prediction_type,Forecaster-method}
\title{Get prediction type}
\usage{
\S4method{get_prediction_type}{Forecaster}(object, type.measure)
}
\arguments{
\item{object}{a S4 \linkS4class{Forecaster} object}

\item{type.measure}{the accuracy measure}
}
\value{
the prediction type required to compute the prediction values expected
in input by a \linkS4class{Scorer} to calculate the score
}
\description{
Select the appropriate prediction type given the accuracy measure
}
