% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ForecasterList.R
\name{get_prediction_type,ForecasterList-method}
\alias{get_prediction_type,ForecasterList-method}
\title{Get prediction type}
\usage{
\S4method{get_prediction_type}{ForecasterList}(object, type.measure, learning.method, ...)
}
\arguments{
\item{object}{a S4 \linkS4class{ForecasterList} object}

\item{type.measure}{the accuracy measure}

\item{learning.method}{(optional) learning method(s) to consider. If provided,
the \code{learning.method} is matched against the learning methods in the \linkS4class{ForecasterList}
object.}
}
\value{
the prediction type required to compute the prediction values expected
in input by a \linkS4class{Scorer} to calculate the score.
If \code{learning.method} is missing or multiple values are provided then the
output is a vector containing the matching prediction types, one for each
learning method in \code{learning.method}.
}
\description{
Select the appropriate prediction type given the accuracy measure
}
