% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruit.R
\name{get_rank_product}
\alias{get_rank_product}
\title{Compute Rank Product/Rank Sum}
\usage{
get_rank_product(object, set = c("full.set", "test.set"), ...)
}
\arguments{
\item{object}{a \code{renoir} object}

\item{set}{the set of data used for the computation of the accuracy measures}

\item{...}{further arguments to \code{\link{compute_rank_product}}}
}
\value{
a data frame or a list of data frames containing the statistics and
associated estimated significance. See \code{\link{compute_rank_product}}
for more details.
}
\description{
Computes the rank product of the features importance to summarise
the results obtained by using from different configuration and accuracy measures.
Note that \code{AUC} measure is not considered.
}
\keyword{internal}
