% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_generics.R, R/class_Evaluated.R,
%   R/class_EvaluatedList.R
\name{get_sample}
\alias{get_sample}
\alias{get_sample,Evaluated,numeric-method}
\alias{get_sample,EvaluatedList,numeric-method}
\title{Get Sample}
\usage{
get_sample(object, index, ...)

\S4method{get_sample}{Evaluated,numeric}(object, index)

\S4method{get_sample}{EvaluatedList,numeric}(object, index, n)
}
\arguments{
\item{object}{\linkS4class{EvaluatedList} object}

\item{index}{index of sample in the \linkS4class{Evaluated} object}

\item{n}{integer, the sample size. Used to select the \linkS4class{Evaluated} object
from the \linkS4class{EvaluatedList}}
}
\value{
The selected sample.
}
\description{
Get Sample
}
