% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Marker.R
\name{get_significance_score}
\alias{get_significance_score}
\title{Compute Significance Score}
\usage{
get_significance_score(weights, recruitment)
}
\arguments{
\item{weights}{the weights for the computation}

\item{recruitment}{the recruitment terms}
}
\value{
the significance score for each feature
}
\description{
Compute Significance Score
}
\keyword{internal}
