% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Sampler.R
\name{get_size_grid}
\alias{get_size_grid}
\title{Get Sample Size Grid}
\usage{
get_size_grid(
  method = c("random", "bootstrap", "cv"),
  N,
  strata = NULL,
  balance = FALSE,
  nmin,
  npoints = 5L
)
}
\arguments{
\item{method}{The sampling strategy to use. The available options are:
\describe{
\item{\code{'random'}}{random sampling}
\item{\code{'cv'}}{k-fold cross-validation}
\item{\code{'bootstrap'}}{bootstrap (sampling with replacement)}
}}

\item{N}{the population size}

\item{strata}{(optional) vector of stratification variables. If provided must be of length \code{N}}

\item{balance}{logical, whether to (try to) balance the strata frequency in the output sample.
Used only if \code{strata} is provided when \code{method = "random"} or \code{method = "bootstrap"}}

\item{nmin}{integer, desired minimum number of observations (minimum sample size)}

\item{npoints}{integer, the desired number of sample sizes}
}
\value{
A vector containing the sample sizes.
}
\description{
Get Sample Size Grid
}
\keyword{internal}
