% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_EvaluatedList.R
\name{identify_primary_key_from_EvaluatedList}
\alias{identify_primary_key_from_EvaluatedList}
\title{Identify Primary Key}
\usage{
identify_primary_key_from_EvaluatedList(object)
}
\arguments{
\item{object}{object of class \linkS4class{EvaluatedList}}
}
\value{
A character vector.
}
\description{
This function identifies a primary key from a list
of \linkS4class{Evaluated} objects.
}
\details{
Currently, elements can be uniquely grouped by four identifiers:
1) the learning method, 2) the hyperparameter configuration,
3) the response type, 4) the sampling strategy.
This function extracts the related values from the \linkS4class{EvaluatedList}
object and check if multiple values are present in order to
identify a unique key.
}
\author{
Alessandro Barberis
}
