% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_generics.R, R/class_EvaluatedList.R
\name{importance}
\alias{importance}
\alias{importance,EvaluatedList-method}
\title{Compute feature importance}
\usage{
importance(object, ...)

\S4method{importance}{EvaluatedList}(object, marker, scorer, features, ...)
}
\arguments{
\item{object}{an object of class \linkS4class{EvaluatedList}}

\item{...}{further arguments to \code{compute_importance}}

\item{marker}{an object of class \linkS4class{Marker}}

\item{scorer}{an object of class \linkS4class{Scorer}}

\item{features}{(optional) vector containing the predictors to consider}
}
\description{
Compute feature importance
}
