% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_multi_response}
\alias{is_multi_response}
\title{Is the response type multi-variable?}
\usage{
is_multi_response(y, resp.type)
}
\arguments{
\item{y}{response variable}

\item{resp.type}{response type}
}
\value{
logical
}
\description{
Is the response type multi-variable?
}
\keyword{internal}
