% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_tests.R
\name{kw}
\alias{kw}
\title{Kruskal-Wallis Rank Sum Test}
\usage{
kw(x, g)
}
\arguments{
\item{x}{a numeric vector of data values, or a list of numeric data
    vectors.  Non-numeric elements of a list will be coerced, with a
    warning.}

\item{g}{a vector or factor object giving the group for the
    corresponding elements of \code{x}.  Ignored with a warning if
    \code{x} is a list.}
}
\description{
Kruskal-Wallis Rank Sum Test
}
\keyword{internal}
