% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Learner.R
\name{learn}
\alias{learn}
\title{Learn}
\usage{
\S4method{learn}{Learner}(learner, hyperparameters = NULL, ...)
}
\arguments{
\item{learner}{an object of class \linkS4class{Learner}}

\item{hyperparameters}{(optional) list containing the hyperparameters to tune}

\item{...}{additional arguments affecting the learning procedure}
}
\value{
The form of the returned object depends on the type
of learning. See the documentation of the particular method for
further details.
}
\description{
Function used to run a learning process.
The function invokes particular methods which depend on the
presence of the second argument. If \code{hyperparameters}
is provided then a tuning is performed, while a training
is run otherwise.
}
\seealso{
\code{\link{tune}},
\code{\link{train}}
}
\author{
Alessandro Barberis
}
