% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Learner.R
\name{Learner,character-method}
\alias{Learner,character-method}
\title{Learner Class Constructor}
\usage{
\S4method{Learner}{character}(
  id,
  tuning = "grid.search",
  screening,
  sampling = c("cv", "random", "bootstrap"),
  strata = NULL,
  balance = FALSE,
  k = 10L,
  score = "mse",
  resp.type,
  looper = Looper(),
  logger = Logger()
)
}
\arguments{
\item{id}{learning method}

\item{tuning}{the tuning method}

\item{screening}{the screening method}

\item{sampling}{the sampling strategy}

\item{strata}{(optional) vector of stratification variables. If provided must be of length \code{N}}

\item{balance}{logical, whether to (try to) balance the strata frequency in the output sample.
Used only if \code{strata} is provided when \code{method = "random"} or \code{method = "bootstrap"}}

\item{k}{the number of folds for \code{method = "cv"},
or the number of repeats for\code{method = "random"} and \code{method = "bootstrap"}}

\item{score}{the accuracy measure to use during tuning}

\item{resp.type}{the response type}

\item{looper}{a \linkS4class{Looper}}

\item{logger}{a \linkS4class{Logger}}
}
\value{
a \linkS4class{Learner} object
}
\description{
Constructor for the S4 \linkS4class{Learner} object.
}
\author{
Alessandro Barberis
}
