% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Learner.R
\docType{class}
\name{Learner-class}
\alias{Learner-class}
\title{Learner Class}
\description{
An S4 class representing a learner.

The object consists of 8 slots
}
\section{Slots}{

\describe{
\item{\code{tuner}}{a \linkS4class{Tuner} object}

\item{\code{trainer}}{a \linkS4class{Trainer} object}

\item{\code{forecaster}}{a \linkS4class{Forecaster} object}

\item{\code{scorer}}{a \linkS4class{ScorerList} object}

\item{\code{selector}}{a \linkS4class{Selector} object}

\item{\code{recorder}}{a \linkS4class{Recorder} object}

\item{\code{marker}}{a \linkS4class{Marker} object}

\item{\code{logger}}{a \linkS4class{Logger} object}
}}

\author{
Alessandro Barberis
}
