% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Learner.R
\name{Learner,missing-method}
\alias{Learner,missing-method}
\title{Learner Class Constructor}
\usage{
\S4method{Learner}{missing}(
  tuner = Tuner(),
  trainer,
  forecaster,
  scorer = Scorer(id = "mse"),
  selector,
  recorder,
  marker,
  logger = Logger()
)
}
\arguments{
\item{tuner}{a \linkS4class{Tuner} object}

\item{trainer}{a \linkS4class{Trainer} object}

\item{forecaster}{a \linkS4class{Forecaster} object}

\item{scorer}{a \linkS4class{ScorerList} or a \linkS4class{Scorer} object}

\item{selector}{a \linkS4class{Selector} object}

\item{recorder}{a \linkS4class{Recorder} object}

\item{marker}{a \linkS4class{Marker} object}

\item{logger}{a \linkS4class{Logger} object}
}
\value{
a \linkS4class{Learner} object
}
\description{
Constructor for the S4 \linkS4class{Learner} object.
}
\author{
Alessandro Barberis
}
