% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_LearnerList.R
\name{LearnerList}
\alias{LearnerList}
\title{Constructor for the S4 LearnerList object.

Constructor for the S4 \linkS4class{LearnerList} object.}
\usage{
LearnerList(...)
}
\description{
LearnerList Constructor
}
