% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Trainer.R
\name{list_supported_learning_methods}
\alias{list_supported_learning_methods}
\title{Supported Learning Methods}
\usage{
list_supported_learning_methods(x = c("all", "regression", "classification"))
}
\arguments{
\item{x}{character string indicating whether to return all learning methods,
the learning method for regression, or the learning methods for classification.}
}
\value{
A \code{data.frame} with 3 columns:
\describe{
\item{\code{id}}{contains the id used in renoir for the learning method}
\item{\code{method}}{contains the name of the learning method (e.g. 'random forest')}
\item{\code{default_hyperparameters}}{containg the name of the hyperparameters used by default}
}
}
\description{
This function returns a \code{data.frame} containing the currently
supported learning methods.
}
\author{
Alessandro Barberis
}
