% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{list_supported_performance_metrics}
\alias{list_supported_performance_metrics}
\title{Supported Performance Metrics}
\usage{
list_supported_performance_metrics(resp.type = NULL)
}
\arguments{
\item{resp.type}{(optional) the response type}
}
\value{
A \code{data.frame} with 3 columns:
\describe{
\item{\code{id}}{contains the id used in renoir for the performance metric (e.g. 'class')}
\item{\code{name}}{contains the name of the performance metric (e.g. 'Misclassification Error')}
\item{\code{problem}}{contains the type of problem for which is commonly used (e.g. 'classification')}
}
}
\description{
This function returns a \code{data.frame} containing the currently
supported performance metrics.
}
\details{
The currently implemented performance metrics are returned in a \code{data.frame}.
If \code{resp.type} is provided, only the metrics supported for the given response
are returned.
}
\seealso{
\code{\link{mean_absolute_error}},
\code{\link{mean_absolute_percentage_error}},
\code{\link{mean_squared_error}},
\code{\link{root_mean_squared_error}},
\code{\link{mean_squared_log_error}},
\code{\link{r2_score}},
\code{\link{classification_error_rate}},
\code{\link{accuracy_score}},
\code{\link{area_under_roc_curve}},
\code{\link{f1_score}},
\code{\link{fbeta_score}},
\code{\link{precision}},
\code{\link{sensitivity}},
\code{\link{jaccard_index}}
}
\author{
Alessandro Barberis
}
