% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Sampler.R
\name{list_supported_sampling_methods}
\alias{list_supported_sampling_methods}
\title{Supported Sampling Methods}
\usage{
list_supported_sampling_methods()
}
\value{
A \code{data.frame} with 3 columns:
\describe{
\item{\code{id}}{contains the id used in renoir for the sampling method (e.g. 'cv')}
\item{\code{name}}{contains the name of the sampling method (e.g. 'cross-validation')}
\item{\code{supported}}{contains the supported strategy}
}
}
\description{
This function returns a \code{data.frame} containing the currently
supported sampling methods.
}
\seealso{
\code{\link{rsample}},
\code{\link{bootstrap}},
\code{\link{create_folds}}
}
\author{
Alessandro Barberis
}
