% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Filter.R
\name{list_supported_unsupervised_screening_methods}
\alias{list_supported_unsupervised_screening_methods}
\title{Supported Unsupervised Screening}
\usage{
list_supported_unsupervised_screening_methods()
}
\value{
A \code{data.frame} with 2 columns:
\describe{
\item{\code{id}}{contains the id used in renoir for the unsupervised screening method (e.g. 'na')}
\item{\code{description}}{contains a short description of the method (e.g. 'filter by NA presence')}
}
}
\description{
This function returns a \code{data.frame} containing the currently
supported unsupervised screening methods.
}
\details{
The currently implemented unsupervised screening methods are returned in a \code{data.frame}.
}
\seealso{
\code{\link{filter_by_na}},
\code{\link{filter_by_intensity}},
\code{\link{filter_by_variability}}
}
\author{
Alessandro Barberis
}
