% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Logger.R
\name{log_all}
\alias{log_all}
\title{Log Method}
\usage{
\S4method{log_all}{Logger}(object, message, sep, add.level = FALSE, add.time = FALSE)
}
\arguments{
\item{object}{an object of class \linkS4class{Logger}}

\item{message}{a character string, the message to print}

\item{sep}{a character vector of strings to append after \code{message}}

\item{add.level}{logical, whether to add the log level to \code{message}}

\item{add.time}{logical, whether to add the time to \code{message}}
}
\value{
None (invisible \code{NULL}).
}
\description{
This function prints the log information of level \code{ALL}.
}
\details{
A log information of level \code{ALL} is printed if the provided logger
has a >= log level.
}
\seealso{
\code{\link{log_trace()}},
\code{\link{log_debug()}},
\code{\link{log_info()}}
}
\author{
Alessandro Barberis
}
