% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Logger.R
\name{log_default}
\alias{log_default}
\title{Log Method}
\usage{
log_default(
  object,
  log.level,
  message,
  sep,
  add.level = FALSE,
  add.time = FALSE
)
}
\arguments{
\item{object}{an object of class \linkS4class{Logger}}

\item{log.level}{a character string, the log level of the information}

\item{message}{a character string, the message to print}

\item{sep}{a character vector of strings to append after \code{message}}

\item{add.level}{logical, whether to add the log level to \code{message}}

\item{add.time}{logical, whether to add the time to \code{message}}
}
\value{
None (invisible \code{NULL}).
}
\description{
Generic function to print the log information.
}
\seealso{
\code{\link{log_all}},
\code{\link{log_trace}},
\code{\link{log_debug}},
\code{\link{log_info}}
}
\author{
Alessandro Barberis
}
\keyword{internal}
