% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Looper.R
\name{loop,Looper-method}
\alias{loop,Looper-method}
\title{Loop}
\usage{
\S4method{loop}{Looper}(looper, n.iter, .inorder = TRUE, fun, ...)
}
\arguments{
\item{looper}{a \linkS4class{Looper}}

\item{n.iter}{number of iterations of the loop}

\item{.inorder}{logical flag indicating whether the \code{.combine}
function requires the task results to be combined in the same order
that they were submitted.  If the order is not important, then it
setting \code{.inorder} to \code{FALSE} can give improved performance.
The default value is `TRUE.}

\item{fun}{A function to be executed inside the loop. It should contain an \code{iloop} parameter.}

\item{...}{further arguments to \code{fun}}
}
\description{
Common interface to run a for loop sequentially or in parallel.
The core of the loop is defined by the function \code{fun}.
}
