% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Looper.R
\docType{class}
\name{Looper-class}
\alias{Looper-class}
\alias{Looper}
\title{Looper Class
An S4 class providing methods for loop constructs.}
\usage{
Looper(logger = Logger(), cores = 1L)
}
\description{
The object consists of 2 slots
}
\section{Slots}{

\describe{
\item{\code{logger}}{a \linkS4class{Logger}}

\item{\code{cores}}{the number of cores to use. If \code{cores > 1} a parallel environment is setup.}
}}

\author{
Alessandro Barberis
}
