% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_generics.R, R/class_Marker.R
\name{mark}
\alias{mark}
\alias{mark,Trained-method}
\alias{mark,Tuned-method}
\alias{mark,list-method}
\title{Mark the features of the trained model}
\usage{
mark(object, ...)

\S4method{mark}{Trained}(object, marker, features, weights, ...)

\S4method{mark}{Tuned}(object, marker, weights, ...)

\S4method{mark}{list}(object, marker, ...)
}
\arguments{
\item{object}{an object of class \linkS4class{Trained}}

\item{...}{further arguments to \code{marker}}

\item{marker}{an object of class \linkS4class{Marker}}

\item{features}{(optional) vector containing all the predictors. If provided,
the results from the marker function are checked and the missing features are added
with NA mark}

\item{weights}{(optional) feature weights to be used in the computation}
}
\value{
A list with one element per response, containing the marks for the features
}
\description{
Mark the features of the trained model
}
\author{
Alessandro Barberis
}
