% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Marker.R
\name{mark_gbm}
\alias{mark_gbm}
\title{Features Marker}
\usage{
mark_gbm(
  object,
  marking.system = c("presence", "influence", "normalized_influence")
)
}
\arguments{
\item{marking.system}{a length-one character vector, indicating a strategy
for marking the features. Available options are \code{"presence"} (default),
\code{"influence"}, and \code{"normalized_influence"}}

\item{obejct}{an object of class \linkS4class{Trained}}
}
\value{
A list with one element per response, containing the marks for the features
}
\description{
This function marks the features of an object of class
\code{\link[gbm]{gbm.object}}.
}
\details{
The features in the \linkS4class{Trained} object in input
are marked, i.e. a numerical value is assigned to each feature
based on the choice of \code{marking.system}.

If \code{marking.system = "presence"}, a binary value of 1/0 is assigned
to the features by their presence in the model:
\code{1} is given to features that were used in the forest,
\code{0} otherwise.

If \code{marking.system = "influence"}, the relative influence of each
variable in reducing the loss function in the gbm object is used.

If \code{marking.system = "normalized_influence"}, the relative influence
normalized to sum to 100 is used.
}
\seealso{
\code{\link[gbm]{relative.influence}},
\code{\link[gbm]{permutation.test.gbm}}
}
\author{
Alessandro Barberis
}
