% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Marker.R
\name{mark_svm}
\alias{mark_svm}
\title{Features Marker}
\usage{
mark_svm(object, marking.system = c("presence", "weights"))
}
\arguments{
\item{marking.system}{a length-one character vector, indicating a strategy
for marking the features.}

\item{obejct}{an object of class \linkS4class{Trained}}
}
\value{
A list with one element per response, containing the marks for the features
}
\description{
This function marks the features of an object of class
\code{\link[e1071]{svm}}.
}
\details{
The features in the \linkS4class{Trained} object in input
are marked, i.e. a numerical value is assigned to each feature
based on the choice of \code{marking.system}.

If \code{marking.system = "presence"}, a binary value of 1/0 is assigned
to the features by their presence in the model:
\code{1} is given to features that were used in the forest,
\code{0} otherwise.

If \code{marking.system = "weights"}, the feature weights are used
as marks.
}
\author{
Alessandro Barberis
}
