% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Marked.R
\docType{class}
\name{Marked-class}
\alias{Marked-class}
\alias{Marked}
\title{Marked
An S4 class to represent a marked model}
\usage{
Marked(mark = data.frame(), measure = character(), set = character())
}
\description{
The Marked object consists of 3 slots

Constructor for the S4 \linkS4class{Marked} object.
}
\section{Slots}{

\describe{
\item{\code{mark}}{A \code{data.frame} containing the feature grades computed on the data}

\item{\code{measure}}{the considered accuracy measure}

\item{\code{set}}{the data set considered}
}}

