% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Marked.R
\name{Marked}
\alias{Marked}
\title{Constructor for the S4 Marked object.}
\usage{
Marked(mark = data.frame(), measure = character(), set = character())
}
\description{
Constructor for the S4 \linkS4class{Marked} object.
}
