% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MarkedList.R
\docType{class}
\name{MarkedList-class}
\alias{MarkedList-class}
\alias{MarkedList}
\title{MarkedList Class
An S4 class to represent a renoir trained and tested models list}
\usage{
MarkedList(...)
}
\description{
This subclass extends the S4  \linkS4class{List} virtual class by using
the  \linkS4class{SimpleList} implementation.

Constructor for the S4 \linkS4class{LearnedList} object.
}
\details{
The object consists of 4 slots
}
\section{Slots}{

\describe{
\item{\code{listData}}{\code{list} storing the list elements}

\item{\code{elementType}}{the type of data represented in the sequence}

\item{\code{elementMetadata}}{\linkS4class{DataFrame} storing the element-wise metadata,
with a row for each element and a column for each metadata available. Default is \code{NULL}}

\item{\code{metadata}}{\code{list} storing the global metadata annotating the object as a whole}
}}

