% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MarkedList.R
\name{MarkedList}
\alias{MarkedList}
\title{LearnedList Constructor
Constructor for the S4 LearnedList object.}
\usage{
MarkedList(...)
}
\description{
Constructor for the S4 \linkS4class{LearnedList} object.
}
