% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Marker.R
\docType{class}
\name{Marker-class}
\alias{Marker-class}
\title{Marker Class
An S4 class to represent a features marker.}
\description{
The object consists of 3 slots
}
\section{Slots}{

\describe{
\item{\code{id}}{a length-one character vector}

\item{\code{marker}}{a function able to assign a numerical value to the features in a trained model}

\item{\code{parameters}}{a list}

\item{\code{logger}}{a \linkS4class{Logger} object}
}}

\author{
Alessandro Barberis
}
