% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Marker.R
\name{Marker}
\alias{Marker}
\title{Marker Class Constructor}
\usage{
Marker(id, marker, parameters, logger = Logger())
}
\arguments{
\item{id}{the learning method}

\item{marker}{(optional) function able to assign a numerical value to
the features in a trained model. Used if \code{id} is not one of the supported by renoir.
If provided, it must return a list with the number of elements equivalent to the number
of responses in the outcome variable.}

\item{parameters}{list containing the parameters to fix for the marking}

\item{logger}{a \linkS4class{Logger}}
}
\value{
An object of class \linkS4class{Marker}.
}
\description{
Constructor for the S4 \linkS4class{Marker} object.
}
\author{
Alessandro Barberis
}
