% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{MCC}
\alias{MCC}
\title{Matthews correlation coefficient}
\usage{
MCC(TP, TN, FP, FN)
}
\arguments{
\item{TP}{number of true positives}

\item{TN}{number of true negatives}

\item{FP}{number of false positives}

\item{FN}{number of false negatives}
}
\value{
A numeric value.
}
\description{
This function computes the Matthews correlation coefficient.
}
\author{
Alessandro Barberis
}
