% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{mean_absolute_error}
\alias{mean_absolute_error}
\title{Mean Absolute Error}
\usage{
mean_absolute_error(
  true,
  pred,
  weights = NULL,
  multi = c("average", "sum", "raw"),
  rweights = NULL
)
}
\arguments{
\item{true}{a vector (or a matrix) of observed values. If a matrix is provided,
a multi-response is assumed}

\item{pred}{a vector (or a matrix) of predicted values}

\item{weights}{observation weights}

\item{multi}{what to do when response has multiple output values
\describe{
\item{\code{average}}{errors of multiple outputs are averaged to get a single value for each observation}
\item{\code{sum}}{errors of multiple outputs are summed up to get a single value for each observation}
\item{\code{raw}}{returns a vector containing one error for each output}
}}

\item{rweights}{response weights}
}
\value{
the mean absolute error, a positive \code{double} or a vector of positive \code{double} values,
one for each response, if response has multiple output values and \code{multi = "raw"}
}
\description{
This function computes the mean absolute error.
}
\details{
The mean absolute error (MAE) is a measure of errors based on absolute values.
It treats small and large errors equally so it is less sensitive to outliers compared
to other measures (e.g. MSE).
The MAE estimated over n observations is defined as

\deqn{MAE(y,\hat{y}) = \frac{1}{n}\sum_{i=1}^{n} \lvert y_{i} - \hat{y}_{i} \rvert}

where \eqn{y_{i}} is the true value of the \eqn{i}-th sample and \eqn{\hat{y}_{i}}
is the predicted value.
If observation weights are provided, then the weighted mean absolute error is computed as

\deqn{wMAE(w,y,\hat{y}) = \frac{1}{\sum_{i=1}^{n} w_{i}}\sum_{i=1}^{n} w_{i} * \lvert y_{i} - \hat{y}_{i} \rvert}

where \eqn{w_{i}} is the weighting factor assigned to the \eqn{i}-th observation.
The best possible score is zero.
}
\author{
Alessandro Barberis
}
