% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{mean_absolute_percentage_error}
\alias{mean_absolute_percentage_error}
\title{Mean Absolute Percentage Error}
\usage{
mean_absolute_percentage_error(
  true,
  pred,
  weights = NULL,
  multi = c("average", "sum", "raw"),
  eps = 1e-10,
  rweights = NULL
)
}
\arguments{
\item{true}{a vector (or a matrix) of observed values. If a matrix is provided,
a multi-response is assumed}

\item{pred}{a vector (or a matrix) of predicted values}

\item{weights}{observation weights}

\item{multi}{what to do when response has multiple output values
\describe{
\item{\code{average}}{errors of multiple outputs are averaged to get a single value for each observation}
\item{\code{sum}}{errors of multiple outputs are summed up to get a single value for each observation}
\item{\code{raw}}{returns a vector containing one error for each output}
}}

\item{eps}{an arbitrary positive small number, used to correct the MAPE formula and avoid the division by zero}

\item{rweights}{response weights}
}
\value{
the mean absolute percentage error, a positive \code{double} or a vector of positive \code{double} values,
one for each response, if response has multiple output values and \code{multi = "raw"}
}
\description{
This function computes the mean absolute percentage error.
}
\details{
The mean absolute percentage error (MAPE) is a measure of errors based on relative absolute values.
The MAPE estimated over \emph{n} observations is defined as

\deqn{MAPE(y,\hat{y}) = \frac{1}{n}\sum_{i=1}^{n} \frac{\lvert y_{i} - \hat{y}_{i} \rvert}{\max{\lvert y_{i} \rvert, \epsilon}}}

where \eqn{y_{i}} is the true value of the \eqn{i}-th sample, \eqn{\hat{y}_{i}}
is the predicted value, and \eqn{\epsilon} is an arbitrary positive small number used
to avoid a division by zero.

If observation weights are provided, then the weighted mean absolute percentage error is computed as

\deqn{wMAPE(w,y,\hat{y}) = \frac{1}{\sum_{i=1}^{n} w_{i}}\sum_{i=1}^{n} w_{i} * \frac{\lvert y_{i} - \hat{y}_{i} \rvert}{\max{\lvert y_{i} \rvert, \epsilon}}}

where \eqn{w_{i}} is the weighting factor assigned to the \eqn{i}-th observation.
The best possible score is zero.
}
\author{
Alessandro Barberis
}
