% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{multiresponse_classification_metric}
\alias{multiresponse_classification_metric}
\title{Multi-response Classification Metric}
\usage{
multiresponse_classification_metric(
  true,
  pred,
  multi = c("weighted", "average", "micro", "macro", "raw", "binary"),
  metric,
  positive = 1,
  weights,
  confusion = NULL,
  ...
)
}
\arguments{
\item{true}{a vector (or a matrix) of observed values. If a matrix is provided,
a multi-response is assumed}

\item{pred}{a vector (or a matrix) of predicted values}

\item{multi}{what to do when response has multiple output values
\describe{
\item{\code{average}}{errors of multiple outputs are averaged to get a single value for each observation}
\item{\code{micro}}{the score is computed by using the global number of true positives, false negatives and false positives}
\item{\code{macro}}{scores of different classes are averaged by unweighted mean to get a single value. Class imbalance is not taken into account}
\item{\code{weighted}}{scores of different classes are averaged by weighted mean to get a single value. Weights are number of true instances per class. It takes into account label imbalance}
\item{\code{raw}}{returns a vector containing one score for each class}
\item{\code{binary}}{returns the score for the class specified by \code{positive}}
}}

\item{metric}{character string indicating the score to compute}

\item{positive}{integer or character indicating the target class.
If \code{positive = 0} (default), global values are computed.}

\item{weights}{observation weights (not implemented yet)}

\item{confusion}{a confusion matrix as returned by \code{\link{confusion_matrix}}. If \code{confusion} is not provided,
a confusion matrix is internally computed by using \code{true} and \code{pred}}
}
\value{
A single score for the selected class if \code{multi = "binary"},
a vector containing one score for each class if \code{multi = "raw"},
a summary score computed by using the global metrics if \code{multi = "micro"} or
averaging the results from different classes if \code{multi = "macro"}. A summary score
produced by a weighted average of the results from different class is produced if
\code{multi = "weighted"}
}
\description{
This function computes a multi-response classification metric.
}
\details{
Common interface for computing a multi-response classification metric.
}
\author{
Alessandro Barberis
}
\keyword{internal}
