% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{multiresponse_error}
\alias{multiresponse_error}
\title{Multi-response error}
\usage{
multiresponse_error(error, weights = NULL, multi = c("average", "sum", "raw"))
}
\arguments{
\item{error}{vector or matrix containing accuracy measures}

\item{weights}{response weights}

\item{multi}{what to do when response has multiple output values
\describe{
  \item{\code{average}}{errors of multiple outputs are averaged to get a single value for each observation}
  \item{\code{sum}}{errors of multiple outputs are summed up to get a single value for each observation}
  \item{\code{raw}}{returns a vector containing one error for each output}
}}
}
\value{
the error after applied the selected strategy
}
\description{
If error is from multi-response data, this function allows to adopt different strategies.
}
\author{
Alessandro Barberis
}
