% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Screener.R
\name{multiresponse_screener}
\alias{multiresponse_screener}
\title{Multi-response screen}
\usage{
multiresponse_screener(
  screened,
  multi = c("max", "average", "sum", "raw"),
  coef = NULL
)
}
\arguments{
\item{screened}{vector or matrix containing the significance for the screened features}

\item{multi}{what to do when response has multiple output values
\describe{
  \item{\code{max}}{for each feature, the maximal significance value across responses is kept}
  \item{\code{average}}{significances of multiple outputs are averaged to get a single value for each feature}
  \item{\code{sum}}{significances of multiple outputs are summed up to get a single value for each feature}
  \item{\code{raw}}{for each feature, returns one score for each output}
}}

\item{coef}{(optional) integer value indicating the response to keep}
}
\value{
The significance of the screening test for each feature after the
application of the selected strategy.
}
\description{
If screened features are from multi-response data,
this function allows to adopt different strategies.
}
\author{
Alessandro Barberis
}
\keyword{internal}
