% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_generics.R, R/class_Recorder.R
\name{nfeatures}
\alias{nfeatures}
\alias{nfeatures,Trained,Recorder-method}
\alias{nfeatures,Tuned,Recorder-method}
\title{Get number of features in the model}
\usage{
nfeatures(object, recorder, ...)

\S4method{nfeatures}{Trained,Recorder}(object, recorder, ...)

\S4method{nfeatures}{Tuned,Recorder}(object, recorder, ...)
}
\description{
Get number of features in the model
}
