% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{NLR}
\alias{NLR}
\title{Negative likelihood ratio (LR-)}
\usage{
NLR(TP, TN, FP, FN)
}
\arguments{
\item{TP}{number of true positives}

\item{TN}{number of true negatives}

\item{FP}{number of false positives}

\item{FN}{number of false negatives}
}
\value{
A numeric value.
}
\description{
This function computes the negative likelihood ratio.
}
\author{
Alessandro Barberis
}
