% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{NPV}
\alias{NPV}
\title{Negative predictive value (1 - FOR)}
\usage{
NPV(TP, TN, FP, FN)
}
\arguments{
\item{TP}{number of true positives}

\item{TN}{number of true negatives}

\item{FP}{number of false positives}

\item{FN}{number of false negatives}
}
\value{
A numeric value.
}
\description{
This function computes the negative predictive value.
}
\author{
Alessandro Barberis
}
