% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{num_to_sym}
\alias{num_to_sym}
\title{Number to symbols}
\usage{
num_to_sym(
  x,
  cutpoints = c(0, 1e-04, 0.001, 0.01, 0.05, 0.1, 1),
  symbols = c("****", "***", "**", "*", ".", "ns")
)
}
\arguments{
\item{x}{number to convert}

\item{cutpoints}{cutpoints linked to \code{symbols}}

\item{symbols}{character strings representing the numer in \code{x} falling
in the range defined by the corresponding \code{cutpoints}}
}
\value{
A symbolic representation of the number in input.
}
\description{
Number to symbols
}
\keyword{internal}
