% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Screener.R
\name{permutation_screener}
\alias{permutation_screener}
\title{Compute the significance of the filtering test}
\usage{
permutation_screener(
  x,
  y,
  weights = NULL,
  resp.type = c("gaussian", "mgaussian", "binomial", "multinomial", "poisson", "cox"),
  observations = NULL,
  coef = NULL,
  sam.resp.type = c("Quantitative", "Two class unpaired", "Survival", "Multiclass",
    "One class", "Two class paired", "Two class unpaired timecourse",
    "One class timecourse", "Two class paired timecourse", "Pattern discovery"),
  geneid = NULL,
  genenames = NULL,
  assay.type = c("array", "seq"),
  s0 = 1e-04,
  s0.perc = NULL,
  nperms = 100,
  center.arrays = FALSE,
  testStatistic = c("standard", "wilcoxon"),
  time.summary.type = c("slope", "signed.area"),
  regression.method = c("standard", "ranks"),
  return.x = FALSE,
  knn.neighbors = 10,
  random.seed = NULL,
  logged2 = TRUE,
  fdr.output = 1,
  eigengene.number = 1,
  nresamp = 20,
  all.genes = TRUE,
  return.type = c("pvalue", "object"),
  logger = Logger(verbose = F),
  multi = c("max", "average", "sum", "raw")
)
}
\arguments{
\item{x}{the input matrix, where rows are observations and columns are variables.}

\item{y}{the response variable. Its number of rows must match the number of rows of \code{x}.}

\item{weights}{priors of the observations}

\item{resp.type}{the response type}

\item{observations}{(optional) indices of observations to keep}

\item{coef}{(optional) an integer indicating the response variable to consider in multi-response data
when \code{multi = "raw"}}

\item{sam.resp.type}{Problem type:\cr
\describe{
\item{\code{"Quantitative"}}{for a continuous parameter (Available for both array and sequencing data)}
\item{\code{"Two class unpaired"}}{for both array and sequencing data}
\item{\code{"Survival"}}{for censored survival outcome (for both array and sequencing data)}
\item{\code{"Multiclass"}}{more than 2 groups (for both array and sequencing data)}
\item{\code{"One class"}}{for a single group (only for array data)}
\item{\code{"Two class paired"}}{for two classes with paired observations (for both array and sequencing data)}
\item{\code{"Two class unpaired timecourse"}}{only for array data}
\item{\code{"One class time course"}}{only for array data}
\item{\code{"Two class.paired timecourse"}}{only for array data}
\item{\code{"Pattern discovery"}}{only for array data}
}}

\item{geneid}{Optional character vector of geneids for output.}

\item{genenames}{Optional character vector of genenames for output.}

\item{s0}{Exchangeability factor  for denominator of test statistic; Default
is automatic choice. Only used for array data.}

\item{s0.perc}{Percentile of standard  deviation values to use for s0; default is
automatic choice; -1 means s0=0 (different from s0.perc=0, meaning
s0=zeroeth percentile of  standard  deviation values= min of sd values.
Only used for array data.}

\item{nperms}{Number of permutations used to estimate false discovery rates}

\item{center.arrays}{Should the data for each sample (array) be median centered
at the outset? Default =FALSE. Only used for array data.}

\item{testStatistic}{Test statistic to use in two class unpaired case.Either
"standard" (t-statistic) or ,"wilcoxon" (Two-sample wilcoxon or Mann-Whitney
test). Only used for array data.}

\item{time.summary.type}{Summary measure for each time course: "slope", or "signed.area"). Only used for array data.}

\item{regression.method}{Regression method for quantitative case:
"standard", (linear least squares) or "ranks" (linear least squares 
on ranked data). Only used for array data.}

\item{return.x}{Should the matrix of feature values be returned?
Only useful for time course data, where x contains summaries of the features
over time. Otherwise x is the same as the input data data\\$x}

\item{knn.neighbors}{Number of nearest neighbors to use for imputation
of missing features values. Only used for array data.}

\item{random.seed}{Optional initial seed for random number generator (integer)}

\item{logged2}{Has the data been transformed by log (base 2)? This information is used only 
for computing fold changes }

\item{fdr.output}{(Approximate) False Discovery Rate cutoff for output in significant genes table}

\item{eigengene.number}{Eigengene to be used  (just for resp.type="Pattern discovery")}

\item{nresamp}{Number of resamples used to construct test statistic. Default 20.}

\item{return.type}{debug argument, if \code{object} the output from samr is reported}

\item{logger}{a \linkS4class{Logger}}

\item{multi}{what to do when response has multiple output values
\describe{
\item{\code{max}}{the max value of scores across multiple outputs is selected to get a single value for each observation}
\item{\code{average}}{scores of multiple outputs are averaged to get a single value for each observation}
\item{\code{sum}}{scores of multiple outputs are summed up to get a single value for each observation}
\item{\code{raw}}{returns the scores for the multiple outputs}
}}

\item{...}{further arguments}
}
\value{
a \linkS4class{Screened} object
}
\description{
This function compute the significance of the screening test by using
a permutation approach (significance analysis of microarray). The score stored
in the reported \linkS4class{Screened} object is the computed q-value.
}
\author{
Alessandro Barberis
}
