% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_PCA}
\alias{plot_PCA}
\title{Plot PCA}
\usage{
plot_PCA(
  PCA,
  groups,
  title = "",
  grid.arrange.ncol = 1,
  palette = NULL,
  palette.type = "qual",
  order = NULL,
  width = NULL,
  height = NULL,
  ...
)
}
\arguments{
\item{PCA}{object as returned by \code{\link[stats]{prcomp}}}

\item{groups}{data frame containing the groups (each column is a different group)}

\item{title}{plot title}

\item{grid.arrange.ncol}{number of columns to arrange plots in final figure}

\item{palette}{(optional) when a single value,
if a string, it is considered as a name of palette;
if a number, it is the index into the lists of palettes of appropriate type.
If \code{palette} is a single (unnamed) value, will be used for all the \code{groups}.
If is a named vector, each element matching a column in \code{groups} will be used.
The parameter is passed to \code{\link[ggplot2]{scale_colour_brewer}}}

\item{palette.type}{palette type. One of seq (sequential), div (diverging) or qual (qualitative)}

\item{order}{the order to force before arranging the list of plots in the grid.
If order is set to \code{order = 'auto'} and \code{grid.arrange.ncol=2}, the resulting grid
has PC components from same group across the columns,
while the same PC components for different groups are reported across the rows}

\item{width, height}{size of each plot in inches}

\item{...}{further arguments to \code{\link[gridExtra]{arrangeGrob}}}
}
\value{
a \code{ggplot2} object
}
\description{
Plot PCA
}
\author{
Alessandro Barberis
}
\keyword{internal}
