% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_single_evaluation}
\alias{plot_single_evaluation}
\title{Plot the Evaluation of a Learning Method}
\usage{
plot_single_evaluation(
  data,
  thr = NULL,
  colour = "darksalmon",
  name.ms = "mean score",
  add.uncertainty = T,
  add.boxplot = T,
  colour.box = "darkgreen",
  add.scores = T,
  colour.point = "darkgreen",
  add.best = T,
  colour.best = "red",
  shape.best = 22,
  size.best = 3,
  scale.x = F,
  title = "Evaluation",
  subtitle = ggplot2::waiver(),
  caption = ggplot2::waiver(),
  xlab = "Training-set size",
  ylab = "Performance",
  ...
)
}
\arguments{
\item{data}{\code{data.frame} containing the data to plot. The function expects
specific columns:
\describe{
\item{\code{training_set_size}}{contains the considered training-set sizes}
\item{\code{score}}{contains the performance metric for each model}
\item{\code{mean_score}}{contains the mean performance metric for the specific training-set size}
\item{\code{lower_ci}}{contains the lower bound of the confidence interval for the mean score}
\item{\code{upper_ci}}{contains the upper bound of the confidence interval for the mean score}
\item{\code{best_resample}}{contains the index of the automatically selected optimal training-set size}
\item{\code{best_model}}{contains the index of the best model for the optimal training-set size}
\item{\code{name}}{contains a grouping key, e.g. the learning method}
}}

\item{thr}{numerical value, if provided it is used to draw an horizontal line}

\item{colour}{character string, containing the colour for the performance estimate}

\item{add.uncertainty}{logical, whether to include the quantified uncertainty of the
performance estimate in the plot}

\item{add.boxplot}{logical, whether to include a boxplot in the figure}

\item{colour.box}{character string, the colour of the boxplot}

\item{add.scores}{logical, whether to add the performance metric of
individual models as points in the plot}

\item{colour.point}{character string, the colour of the points}

\item{add.best}{logical, whether to add a point indicating the performance
of what is reported as best model in \code{data}}

\item{colour.best}{character string, the colour of the best model}

\item{shape.best}{integer, \code{shape} aesthetic passed to \code{\link[ggplot2]{geom_point}}}

\item{size.best}{integer, \code{size} aesthetic passed to \code{\link[ggplot2]{geom_point}}}

\item{scale.x}{logical, whether to force the scaling of the x-axis}

\item{title}{character string, the title of the plot}

\item{subtitle}{character string, the subtitle of the plot}

\item{caption}{character string, the caption of the plot}

\item{xlab, ylab}{character string, axes labels}

\item{...}{further arguments to \code{\link[ggplot2]{ggplot}}}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
This function creates a plot representing
the evaluation of a learning method across different
training-set sizes.
}
\details{
A plot showing the mean performance and the related
95\\% confidence interval of a learning method
across different training-set sizes is produced.
Individual scores and summary metrics in the form of boxplots
can be also added (default) via the \code{add.scores} and
\code{add.boxplot} arguments, respectively.
}
\seealso{
\code{\link{plot_multi_evaluation}}
}
\author{
Alessandro Barberis
}
\keyword{internal}
