% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{plotly.density}
\alias{plotly.density}
\title{Create a plotly density plot}
\usage{
plotly.density(
  data,
  groups,
  bw = "nrd0",
  fill = "tozeroy",
  yaxis = "Density",
  xaxis = "",
  title.plot = "Density curve",
  na.rm = TRUE
)
}
\arguments{
\item{data}{the data from which the estimate is to be computed. For the default method a numeric vector: long vectors are not supported.}

\item{groups}{a list of matrices, each representing a group}

\item{bw}{the smoothing bandwidth as in \code{\link[stats]{density}}}

\item{yaxis}{y axis' title}

\item{xaxis}{x axis' title}

\item{title.plot}{layout's title}
}
\value{
a plotly object
}
\description{
Create a plotly density plot
}
\keyword{internal}
