% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{plotly_evaluation}
\alias{plotly_evaluation}
\title{Plot the Evaluation of a Learning Method}
\usage{
plotly_evaluation(
  data,
  tooltip = c("imodel", "screened", "nfeatures", "configuration"),
  ...
)
}
\arguments{
\item{data}{\code{data.frame} containing the summary of an object of class
\linkS4class{Renoir} as returned by \code{\link{summary_table()}}.
This function expects specific columns:
\describe{
  \item{\code{training_set_size}}{contains the considered training-set sizes}
  \item{\code{score}}{contains the performance metric for each model}
  \item{\code{mean_score}}{contains the mean performance metric for the specific training-set size}
  \item{\code{lower_ci}}{contains the lower bound of the confidence interval for the mean score}
  \item{\code{upper_ci}}{contains the upper bound of the confidence interval for the mean score}
  \item{\code{best_resample}}{contains the index of the automatically selected optimal training-set size}
  \item{\code{best_model}}{contains the index of the best model for the optimal training-set size}
  \item{\code{name}}{contains a grouping key, e.g. the learning method}
}
The \code{name} column is used to identify the
number of considered evaluations}

\item{tooltip}{a character vector, the subset of columns to select for creating a text for tooltip.
It is used unless a \code{tooltip} column is found in \code{data}}

\item{...}{further arguments to \code{\link{plotly_single_evaluation}}
or \code{\link{plotly_multi_evaluation}}}
}
\value{
An object of class \code{plotly}
}
\description{
This function creates an interactive plot representing
the evaluation of a learning method across different
training-set sizes.
}
\details{
An interactive plot showing the mean performance and the related
95\% confidence interval of a learning method
across different training-set sizes is produced.
The evaluated element is identified by the \code{name}
column in the \code{data}. If a unique key is found then
\code{\link{plotly_single_evaluation}} is dispatched, while if
multiple keys are found then the dispatched method is
\code{\link{plotly_multi_evaluation}}. In the latter case,
multiple evaluations are reported in the same plot.
}
\seealso{
\code{\link{plotly_single_evaluation}},
\code{\link{plotly_multi_evaluation}}
}
\author{
Alessandro Barberis
}
