% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{plotly_multi_evaluation}
\alias{plotly_multi_evaluation}
\title{Plot Multiple Evaluations}
\usage{
plotly_multi_evaluation(
  data,
  thr = NULL,
  add.uncertainty = T,
  alpha.ci = NULL,
  add.boxplot = T,
  add.scores = T,
  add.best = T,
  shape.best = 0,
  size.best = 10,
  title = "Evaluation",
  xlab = "Training-set size",
  ylab = "Performance",
  ...
)
}
\arguments{
\item{data}{\code{data.frame} containing the data to plot. The function expects
specific columns:
\describe{
\item{\code{training_set_size}}{contains the considered training-set sizes}
\item{\code{score}}{contains the performance metric for each model}
\item{\code{mean_score}}{contains the mean performance metric for the specific training-set size}
\item{\code{lower_ci}}{contains the lower bound of the confidence interval for the mean score}
\item{\code{upper_ci}}{contains the upper bound of the confidence interval for the mean score}
\item{\code{best_resample}}{contains the index of the automatically selected optimal training-set size}
\item{\code{best_model}}{contains the index of the best model for the optimal training-set size}
\item{\code{name}}{contains a grouping key, e.g. the learning method}
\item{\code{tooltip}}{contains the hover text}
}}

\item{thr}{numerical value, if provided it is used to draw an horizontal line}

\item{add.uncertainty}{logical, whether to include the quantified uncertainty of the
performance estimate in the plot}

\item{alpha.ci}{alpha value for \code{color.ci}}

\item{add.boxplot}{logical, whether to include a boxplot in the figure}

\item{add.scores}{logical, whether to add the performance metric of
individual models as points in the plot}

\item{add.best}{logical, whether to add a point indicating the performance
of what is reported as best model in \code{data}}

\item{shape.best}{integer, \code{shape} aesthetic passed to \code{\link[plotly]{add_trace}}}

\item{size.best}{integer, \code{size} aesthetic passed to \code{\link[plotly]{add_trace}}}

\item{title}{character string, the title of the plot}

\item{xlab, ylab}{character string, axes labels}

\item{...}{further arguments to \code{\link[plotly]{plot_ly}}}
}
\value{
A \code{plotly} object.
}
\description{
This function creates an interactive plot representing
multiple evaluations of a learning method across different
training-set sizes.
}
\details{
An interactive plot showing the mean performance and the related
95\\% confidence interval of learning methods
across different training-set sizes is produced.
Individual scores and summary metrics in the form of boxplots
can be also added (default) via the \code{add.scores} and
\code{add.boxplot} arguments, respectively.
}
\seealso{
\code{\link{plotly_single_evaluation}}
}
\author{
Alessandro Barberis
}
\keyword{internal}
