% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{plotly_single_evaluation}
\alias{plotly_single_evaluation}
\title{Plot the Evaluation of a Learning Method}
\usage{
plotly_single_evaluation(
  data,
  thr = NULL,
  colour = "darksalmon",
  add.uncertainty = T,
  color.ci = I("gray95"),
  alpha.ci = NULL,
  add.boxplot = T,
  colour.box = "rgba(136, 194, 136, 1)",
  colour.box.line = "rgba(44, 160, 44, 1)",
  add.scores = T,
  colour.point = "rgba(44, 160, 44, 1)",
  add.best = T,
  colour.best = "rgba(255, 182, 193, .9)",
  colour.best.line = "rgba(152, 0, 0, .8)",
  size.best = 10,
  shape.best = 0,
  title = "Evaluation",
  xlab = "Training-set size",
  ylab = "Performance",
  ...
)
}
\arguments{
\item{data}{\code{data.frame} containing the data to plot. The function expects
specific columns:
\describe{
  \item{\code{training_set_size}}{contains the considered training-set sizes}
  \item{\code{score}}{contains the performance metric for each model}
  \item{\code{mean_score}}{contains the mean performance metric for the specific training-set size}
  \item{\code{lower_ci}}{contains the lower bound of the confidence interval for the mean score}
  \item{\code{upper_ci}}{contains the upper bound of the confidence interval for the mean score}
  \item{\code{best_resample}}{contains the index of the automatically selected optimal training-set size}
  \item{\code{best_model}}{contains the index of the best model for the optimal training-set size}
  \item{\code{name}}{contains a grouping key, e.g. the learning method}
  \item{\code{tooltip}}{contains the hover text}
}}

\item{thr}{numerical value, if provided it is used to draw an horizontal line}

\item{colour}{character string, containing the colour for the performance estimate}

\item{add.uncertainty}{logical, whether to include the quantified uncertainty of the
performance estimate in the plot}

\item{color.ci}{character string, containing the colour for the confidence interval}

\item{alpha.ci}{alpha value for \code{color.ci}}

\item{add.boxplot}{logical, whether to include a boxplot in the figure}

\item{colour.box}{character string, the colour of the boxplot}

\item{colour.box.line}{character string, the colour of the lines in the boxplot}

\item{add.scores}{logical, whether to add the performance metric of
individual models as points in the plot}

\item{colour.point}{character string, the colour of the points}

\item{add.best}{logical, whether to add a point indicating the performance
of what is reported as best model in \code{data}}

\item{colour.best}{character string, the colour of the best model}

\item{colour.best.line}{character string, the colour of the line of the best model}

\item{size.best}{integer, \code{size} aesthetic passed to \code{\link[plotly]{add_trace}}}

\item{shape.best}{integer, \code{shape} aesthetic passed to \code{\link[plotly]{add_trace}}}

\item{title}{character string, the title of the plot}

\item{xlab, ylab}{character string, axes labels}

\item{...}{further arguments to \code{\link[plotly]{plot_ly}}}
}
\value{
A \code{plotly} object.
}
\description{
This function creates an interactive plot representing
the evaluation of a learning method across different
training-set sizes.
}
\details{
An interactive plot showing the mean performance and the related
95\% confidence interval of a learning method
across different training-set sizes is produced.
Individual scores and summary metrics in the form of boxplots
can be also added (default) via the \code{add.scores} and
\code{add.boxplot} arguments, respectively.
}
\seealso{
\code{\link{plotly_multi_evaluation}}
}
\author{
Alessandro Barberis
}
