% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.R
\name{predict.relaxed}
\alias{predict.relaxed}
\title{Predict}
\usage{
\method{predict}{relaxed}(
  object,
  newx,
  s = NULL,
  gamma = 1,
  type = c("link", "response", "coefficients", "nonzero", "class"),
  exact = FALSE,
  newoffset,
  ...
)
}
\description{
Corrects a bug in blend.relaxed
}
