% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ConfusionMatrix}
\alias{print.ConfusionMatrix}
\title{Create a confusion table in html}
\source{
https://haozhu233.github.io/kableExtra/awesome_table_in_html.html
}
\usage{
\method{print}{ConfusionMatrix}(
  x,
  digits = max(3, getOption("digits") - 3),
  format = c("text", "html"),
  html.pkg = c("DT", "kable"),
  taglist = T,
  ...
)
}
\arguments{
\item{x}{a confusion matrix as returned by \code{\link{confusion_matrix}}}

\item{digits}{Maximum number of digits for numeric columns, passed to
\code{round()}. This can also be a vector of length \code{ncol(x)}, to set
the number of digits for individual columns.}

\item{format}{A character string. Possible values are \code{latex},
\code{html}, \code{pipe} (Pandoc's pipe tables), \code{simple} (Pandoc's
simple tables), and \code{rst}. The value of this argument will be
automatically determined if the function is called within a \pkg{knitr}
document. The \code{format} value can also be set in the global option
\code{knitr.table.format}. If \code{format} is a function, it must return a
character string.}

\item{...}{Other arguments (see Examples and References).}
}
\value{
A character vector of the table source code.
}
\description{
Create a confusion table in html
}
\examples{
\dontrun{
true = c(0, 1, 2, 2, 0)
pred = c(0, 0, 2, 1, 0)
x = confusion_matrix(true = true, pred = pred)
print.ConfusionMatrix(x = x)
print.ConfusionMatrix(x = x, format = 'html')

true = c('dog', 'cat', 'dog', 'cat', 'dog', 'dog', 'cat', 'dog', 'cat', 'dog')
pred = c('dog', 'dog', 'dog', 'cat', 'dog', 'dog', 'cat', 'cat', 'cat', 'cat')
x = confusion_matrix(true = true, pred = pred)
print.ConfusionMatrix(x = x)
print.ConfusionMatrix(x = x, format = 'html')

true = sample(x = 1:4, size = 10, replace = T)
pred = sample(x = 1:4, size = 10, replace = T)
x = confusion_matrix(true = true, pred = pred)

print.ConfusionMatrix(x = x, format = 'html')
}
}
\keyword{internal}
