% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Logger.R
\name{print.log.line}
\alias{print.log.line}
\title{Print log line}
\usage{
\method{print}{log.line}(
  line = c("long.line.1", "long.line.2", "long.line.3", "short.line.1", "short.line.2",
    "short.line.3"),
  verbose,
  con,
  sep
)
}
\description{
Print log line
}
\keyword{internal}
