% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_confusion_table}
\alias{print_confusion_table}
\title{Create a confusion table in html}
\usage{
print_confusion_table(
  x,
  digits = max(3, getOption("digits") - 3),
  format = "html",
  ...
)
}
\arguments{
\item{x}{a confusion matrix as returned by \code{\link[glmnet]{confusion.glmnet}}}

\item{digits}{Maximum number of digits for numeric columns, passed to
\code{round()}. This can also be a vector of length \code{ncol(x)}, to set
the number of digits for individual columns.}

\item{format}{A character string. Possible values are \code{latex},
\code{html}, \code{pipe} (Pandoc's pipe tables), \code{simple} (Pandoc's
simple tables), and \code{rst}. The value of this argument will be
automatically determined if the function is called within a \pkg{knitr}
document. The \code{format} value can also be set in the global option
\code{knitr.table.format}. If \code{format} is a function, it must return a
character string.}

\item{...}{Other arguments (see Examples and References).}
}
\value{
A character vector of the table source code.
}
\description{
Create a confusion table in html
}
\keyword{internal}
