% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Quantifier.R
\docType{class}
\name{Quantifier-class}
\alias{Quantifier-class}
\title{Quantifier Class
An S4 class providing the methods to quantify uncertainty.}
\description{
The object consists of 6 slots
}
\section{Slots}{

\describe{
\item{\code{id}}{a character string}

\item{\code{quantifier}}{a function}

\item{\code{parameters}}{a list}

\item{\code{logger}}{a \linkS4class{Logger}}
}}

\author{
Alessandro Barberis
}
