% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Quantifier.R
\name{Quantifier}
\alias{Quantifier}
\title{Constructor for the S4 Quantifier object.}
\usage{
Quantifier(id = "ciwm", quantifier, parameters, logger = Logger())
}
\arguments{
\item{id}{the quantifier method name associated to this \linkS4class{Quantifier}.
If the method is one of the supported by renoir, the constructor will
automatically select a \code{quantifier}. See \code{supported_uncertainty_quantifiers()}
for the supported methods.}

\item{quantifier}{(optional) function to quantify uncertainty.
Must be provided if \code{id} is not one of the supported by renoir.}

\item{parameters}{list containing the parameters to fix for the chosen quantifier}

\item{logger}{a \linkS4class{Logger}}
}
\description{
Constructor for the S4 \linkS4class{Quantifier} object.
}
\author{
Alessandro Barberis
}
