% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{r2_score}
\alias{r2_score}
\title{R squared}
\source{
See \href{https://en.wikipedia.org/wiki/Coefficient_of_determination}{Wiki page} for further information.
}
\usage{
r2_score(true, pred, weights = NULL, multi = c("average", "sum", "raw"))
}
\arguments{
\item{true}{a vector (or a matrix) of observed values. If a matrix is provided,
a multi-response is assumed}

\item{pred}{a vector (or a matrix) of predicted values}

\item{weights}{observation weights}

\item{multi}{what to do when response has multiple output values
\describe{
  \item{\code{average}}{errors of multiple outputs are averaged to get a single value for each observation}
  \item{\code{sum}}{errors of multiple outputs are summed up to get a single value for each observation}
  \item{\code{raw}}{returns a vector containing one error for each output}
}}
}
\value{
the R squared score, a \code{double} or a vector of \code{double} values,
one for each response, if response has multiple output values and \code{multi = "raw"}
}
\description{
This function computes the coefficient of determination (R squared).
}
\details{
The coefficient of determination (R squared) is a measure of the proportion
of variance in the dependent variable that is explained by the independent variables.
The R2 estimated over n observations is defined as

\deqn{R^{2}(y,\hat{y}) = 1 - \frac{\sum_{i=1}^{n} (y_{i} - \hat{y}_{i})^{2}}{\sum_{i=1}^{n} (y_{i} - \bar{y})^{2}}}

where \eqn{\bar{y}} is the mean of the observed data, \eqn{y_{i}} is the true value of
the \eqn{i}-th sample, and \eqn{\hat{y}_{i}} is the predicted value.

If observation weights are provided, then the weighted R squared is computed as

\deqn{wR^{2}(w,y,\hat{y}) = 1 - \frac{\sum_{i=1}^{n} w_{i}*(y_{i} - \hat{y}_{i})^{2}}{\sum_{i=1}^{n} w_{i}*(y_{i} - \bar{y})^{2}}}

where \eqn{w_{i}} is the weighting factor assigned to the \eqn{i}-th observation.

The best possible score is one.
}
\author{
Alessandro Barberis
}
