% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_generics.R, R/class_Recorder.R
\name{record}
\alias{record}
\alias{record,Trained,Recorder-method}
\alias{record,Trained,missing-method}
\alias{record,TrainedList,Recorder-method}
\alias{record,Tuned,Recorder-method}
\title{Features Recorder}
\usage{
record(object, recorder, ...)

\S4method{record}{Trained,Recorder}(object, recorder, features, ...)

\S4method{record}{Trained,missing}(object, recorder, features, ...)

\S4method{record}{TrainedList,Recorder}(object, recorder, ...)

\S4method{record}{Tuned,Recorder}(object, recorder, ...)
}
\arguments{
\item{object}{a \linkS4class{Trained} object}

\item{recorder}{a \linkS4class{Recorder} object}

\item{features}{(optional) vector containing all the predictor names. If provided,
the results from the recorder function are checked and the missing features are added
with NA record}
}
\description{
Features Recorder
}
