% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Recorder.R
\name{record_gknn}
\alias{record_gknn}
\title{Features Recorder}
\usage{
record_gknn(object)
}
\arguments{
\item{obejct}{an object of class \code{gknn}}
}
\value{
A list with one element per response, each element being
an integer vector of length equivalent to the number of features.
}
\description{
This function records the presence of the features
in an object of class \code{\link[e1071]{gknn}}.
}
\details{
The presence of the features in the \code{gknn}
object in input is recorded and returned in a named vector
per output response.

If a feature is present in the model then the feature is returned with a \code{1}.
If a feature is missing in the model then the feature is returned with a \code{0}.
}
\author{
Alessandro Barberis
}
