% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Recorder.R
\name{record_randomForest}
\alias{record_randomForest}
\title{Features Recorder}
\usage{
record_randomForest(object)
}
\arguments{
\item{obejct}{an object of class \linkS4class{randomForest}}
}
\value{
A list with one element per response, each element being
an integer vector of length equivalent to the number of features.
}
\description{
This function records the presence of the features
in an object of class \code{\link{randomForest}}.
}
\details{
The presence of the features in the \linkS4class{randomForest}
object in input is recorded and returned in a named vector
per output response.

If a feature is present in the model (i.e. the feature is used in
the random forest) then the feature is returned with a \code{1}.
If a feature is missing in the model (i.e. the feature is not used in
the random forest) then the feature is returned with a \code{0}.
}
\author{
Alessandro Barberis
}
