% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Recorder.R
\docType{class}
\name{Recorder-class}
\alias{Recorder-class}
\title{Recorder Class
An S4 class to represent a recorder.}
\description{
The object consists of 3 slots
}
\section{Slots}{

\describe{
\item{\code{id}}{a length-one character vector}

\item{\code{recorder}}{a function able to record the presence of a feature in a trained model}

\item{\code{logger}}{a \linkS4class{Logger} object}
}}

\author{
Alessandro Barberis
}
