% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Recorder.R
\name{Recorder}
\alias{Recorder}
\title{Recorder Class Constructor}
\usage{
Recorder(id, recorder, logger = Logger())
}
\arguments{
\item{id}{the learning method name associated to this \linkS4class{Recorder}.
If learning method is one of the supported by renoir, the constructor will
automatically select a \code{recorder}. See \code{supported_learning_methods()}
for the supported methods.}

\item{recorder}{(optional) function to record if a predictor is present in the trained model.
Used if \code{id} is not one of the supported by renoir.
If \code{recorder} is provided it must conform to the renoir common interface,
and must have the following formal arguments:
\describe{
   \item{object}{a \linkS4class{Trained} object}
   \item{...}{additional arguments}
}}

\item{logger}{a \linkS4class{Logger}}
}
\value{
An object of class \linkS4class{Recorder}.
}
\description{
Constructor for the S4 \linkS4class{Recorder} object.
}
\author{
Alessandro Barberis
}
