% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Renoir.R
\docType{class}
\name{Renoir-class}
\alias{Renoir-class}
\alias{Renoir}
\title{Renoir Class}
\usage{
Renoir(
  id = character(),
  config = character(),
  response = character(),
  nout = integer(),
  grid = integer(),
  k = integer(),
  sampling = character(),
  learning = character(),
  screening = character(),
  scoring = character(),
  filter = Filtered(),
  evaluation = EvaluatedList(),
  stability = list(),
  marks = MarkedList(),
  nbest = data.frame(),
  call = list()
)
}
\arguments{
\item{id}{evaluated learning method}

\item{config}{name of the configuration}

\item{response}{response type}

\item{nout}{the number of responses}

\item{grid}{integer vector, the considered training set sizes}

\item{k}{integer, the number of sampling repeats}

\item{sampling}{string, sampling method used for evaluation}

\item{learning}{string, the learning procedure (i.e. training or tuning)}

\item{screening}{string, the feature screening strategy adopted}

\item{scoring}{string, the accuracy measure(s) used during the evaluation}

\item{filter}{a \linkS4class{Filtered} object containing the summary of the pre-processing}

\item{evaluation}{a \linkS4class{EvaluatedList} object containing the evaluation of the
learning method}

\item{stability}{list, containing the features stability for each response,
i.e. the frequency of features recruitment per response across all the computed models}

\item{marks}{a \linkS4class{MarkedList} object containing the features importance}

\item{nbest}{data.frame, it contains the indices of the automatically selected best training set size for
different settings (e.g. assessment on training, testing, full set of data)}

\item{call}{list, containing the arguments to the "renoir" function call that creates the \linkS4class{Renoir} object}
}
\value{
a \linkS4class{Renoir} object
}
\description{
An S4 class representing the evaluation of a learning method by repeated sampling.

Constructor for the S4 \linkS4class{Renoir} object.
}
\section{Slots}{

\describe{
\item{\code{id}}{a character string}

\item{\code{config}}{a character string}

\item{\code{response}}{a character string}

\item{\code{nout}}{a one-length integer vector}

\item{\code{grid}}{a integer vector}

\item{\code{k}}{a one-length integer vector}

\item{\code{sampling}}{a character string}

\item{\code{learning}}{a character string}

\item{\code{screening}}{a character string}

\item{\code{scoring}}{a character string}

\item{\code{filter}}{a \linkS4class{Filtered} object}

\item{\code{evaluation}}{a \linkS4class{EvaluatedList} object}

\item{\code{stability}}{a list object}

\item{\code{marks}}{a \linkS4class{MarkedList} object}

\item{\code{nbest}}{a data.frame object}

\item{\code{call}}{list}
}}

\seealso{
\code{\link{Renoir}}
}
\author{
Alessandro Barberis
}
