% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Renoir.R
\name{renoir}
\alias{renoir}
\title{Evaluation of a learning method}
\usage{
renoir(
  filter,
  npoints = 3,
  ngrid,
  nmin = round(nrow(x)/2),
  looper = Looper(),
  filename = "renoir",
  outdir = NULL,
  restore = TRUE,
  learner,
  evaluator,
  logger,
  hyperparameters,
  x,
  y,
  weights = NULL,
  offset = NULL,
  resp.type,
  rm.call = FALSE,
  rm.fit = FALSE,
  grouping = TRUE,
  ...
)
}
\arguments{
\item{filter}{a \linkS4class{Filter} or \linkS4class{FilterList} object}

\item{npoints}{number of sample sizes to consider (i.e. number of training set sizes in the grid)}

\item{ngrid}{(optional) grid of sample sizes to consider}

\item{nmin}{minimum sample size (i.e. number of samples to use for training)}

\item{looper}{a \linkS4class{Looper} object}

\item{filename}{(optional) name without extension for the output file}

\item{outdir}{path to the output directory}

\item{restore}{logical, whether to try to restore a previously ran analysis
if it stopped before the end (\code{outdir} should have been set in the previous analysis)}

\item{learner}{a \linkS4class{Learner} object}

\item{evaluator}{a \linkS4class{Evaluator} object}

\item{logger}{a \linkS4class{Logger} object}

\item{hyperparameters}{named list containing the hyperparameters of the learning method to tune}

\item{x}{the input matrix, where rows are observations and columns are variables.}

\item{y}{the response variable. Its number of rows must match the number of rows of \code{x}.}

\item{weights}{(optional) vector of observation weights}

\item{offset}{vector containing an offset, used for linear models. Default is \code{NULL}.}

\item{resp.type}{the response type}

\item{rm.call}{logical, whether to remove the call from the models. Helpful if object size is expected to be big}

\item{rm.fit}{logical, whether to remove the model fits used for tuning the hyperparameters . Helpful if object size is expected to be big
object can be huge.}

\item{grouping}{logical, whether to group results by learning method, configuration, response type, sampling method}

\item{...}{further arguments to the evaluate function}
}
\value{
The output depends on the adopted learning strategy.
If the tuning was required then the output is a list containing
2 \linkS4class{Renoir} objects, one per assessed configuration.
If tuning was not required, the output is a \linkS4class{Renoir}
object.
}
\description{
This function runs an evaluation of a learning method
applied over a predictors matrix \code{x} and a response variable
\code{y}.
See the \strong{Details} section below for further information.
}
\details{
This function uses the provided \linkS4class{Learner} to
train a model.

The \linkS4class{Evaluator} in input contains
}
\author{
Alessandro Barberis
}
