% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Trainer.R
\name{renoir_default_arg_glmnet}
\alias{renoir_default_arg_glmnet}
\title{Get default arguments for glmnet call}
\usage{
renoir_default_arg_glmnet(name)
}
\value{
a list with default argument for \code{\link[glmnet]{glmnet}} call
}
\description{
Get default arguments for glmnet call
}
\keyword{internal}
