% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Trainer.R
\name{renoir_gbm}
\alias{renoir_gbm}
\title{Generalized Boosted Regression Modeling (GBM)}
\usage{
renoir_gbm(
  x,
  y = NULL,
  weights = NULL,
  offset = NULL,
  resp.type,
  observations = NULL,
  features = NULL,
  clean = FALSE,
  keep.call = TRUE,
  ...,
  eta,
  ntree,
  keep.data = FALSE,
  verbose = F
)
}
\arguments{
\item{x}{the input matrix, where rows are observations and columns are variables}

\item{y}{the response variable. Its number of rows must match the number of rows of \code{x}}

\item{weights}{priors of the observations}

\item{offset}{used only for GLM methods, it is an a priori known component to be included in the linear predictor during fitting}

\item{resp.type}{the response type}

\item{observations}{indices of observations to keep}

\item{features}{indices of predictors to keep}

\item{...}{further arguments to \code{\link[gbm]{gbm.fit}}}

\item{eta}{The shrinkage parameter applied to each tree in the expansion.
Also known as the learning rate or step-size reduction; 0.001 to 0.1 usually work,
but a smaller learning rate typically requires more trees. Default is 0.1.}

\item{ntree}{the total number of trees to fit. This is equivalent to the
number of iterations and the number of basis functions in the additive expansion.}

\item{keep.data}{logical, whether or not to keep the data and an index of the data stored with the object.}

\item{verbose}{Logical indicating whether or not to print out progress and 
performance indicators (\code{TRUE}). If this option is left unspecified for 
\code{gbm.more}, then it uses \code{verbose} from \code{object}. Default is
\code{FALSE}.}

\item{distribution}{either a character string specifying the name of the distribution to use or
a list with a component name specifying the distribution and any additional parameters needed.
If not specified, it is set by \code{resp.type}}
}
\value{
An object of class \linkS4class{Trained}
}
\description{
Generalized Boosted Regression Modeling (GBM)
}
\author{
Alessandro Barberis
}
