% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Trainer.R
\name{renoir_glmnet}
\alias{renoir_glmnet}
\title{GLM with elastic-net}
\usage{
renoir_glmnet(
  x,
  y,
  weights = NULL,
  offset = NULL,
  clean,
  keep.call,
  alpha = 1,
  relax = F,
  gamma = 1,
  resp.type,
  observations = NULL,
  features = NULL,
  learning.method,
  ...
)
}
\arguments{
\item{x}{the input matrix, where rows are observations and columns are variables}

\item{y}{the response variable. Its number of rows must match the number of rows of \code{x}}

\item{weights}{priors of the observations}

\item{offset}{used only for GLM methods, it is an a priori known component to be included in the linear predictor during fitting}

\item{alpha}{The elasticnet mixing parameter, with \eqn{0\le\alpha\le 1}.
The penalty is defined as
\deqn{(1-\alpha)/2||\beta||_2^2+\alpha||\beta||_1.} \code{alpha=1} is the
lasso penalty, and \code{alpha=0} the ridge penalty.}

\item{relax}{If \code{TRUE} then for each \emph{active set} in the path of
solutions, the model is refit without any regularization. See \code{details}
for more information. This argument is new, and users may experience convergence issues
with small datasets, especially with non-gaussian families. Limiting the
value of 'maxp' can alleviate these issues in some cases.}

\item{gamma}{dummy variable, not used in training but set in config}

\item{resp.type}{the response type}

\item{observations}{indices of observations to keep}

\item{features}{indices of predictors to keep}

\item{...}{further arguments to \code{\link[glmnet]{glmnet}}}
}
\value{
An object of class \linkS4class{Trained}
}
\description{
GLM with elastic-net
}
\author{
Alessandro Barberis
}
