% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renoir_nearest_shrunken_centroid.R
\name{renoir_nearest_shrunken_centroid}
\alias{renoir_nearest_shrunken_centroid}
\title{Nearest Shrunken Centroid}
\usage{
renoir_nearest_shrunken_centroid(
  x,
  y,
  weights = NULL,
  offset = NULL,
  clean,
  keep.call,
  resp.type = c("binomial", "multinomial"),
  observations = NULL,
  features = NULL,
  learning.method,
  ...,
  threshold
)
}
\arguments{
\item{...}{further arguments to \code{\link[pamr]{pamr.train}}}

\item{threshold}{A vector of threshold values for the centroid
    shrinkage.Default is a set of 30 values chosen by the software}
}
\description{
This function trains a nearest shrunken centroid classifier.
}
\details{
Renoir interface to \code{\link[pamr]{pamr.train}}.
}
\seealso{
\code{\link[pamr]{pamr.train}}
}
\author{
Alessandro Barberis
}
