% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Trainer.R
\name{renoir_random_forest}
\alias{renoir_random_forest}
\title{Random Forest}
\usage{
renoir_random_forest(
  x,
  y = NULL,
  weights = NULL,
  offset = NULL,
  resp.type,
  observations = NULL,
  features = NULL,
  ntree = 500,
  clean = FALSE,
  keep.call = TRUE,
  ...
)
}
\arguments{
\item{x}{the input matrix, where rows are observations and columns are variables}

\item{y}{the response variable. Its number of rows must match the number of rows of \code{x}}

\item{weights}{priors of the observations}

\item{offset}{used only for GLM methods, it is an a priori known component to be included in the linear predictor during fitting}

\item{resp.type}{the response type}

\item{observations}{indices of observations to keep}

\item{features}{indices of predictors to keep}

\item{ntree}{Number of trees to grow.  This should not be set to too
    small a number, to ensure that every input row gets predicted at
    least a few times. }

\item{...}{further arguments to \code{\link[randomForest]{randomForest}}}
}
\value{
An object of class \linkS4class{Trained}
}
\description{
Random Forest
}
\author{
Alessandro Barberis
}
