% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Trainer.R
\name{renoir_split_glmnet}
\alias{renoir_split_glmnet}
\title{Split glmnet object}
\usage{
renoir_split_glmnet(object)
}
\arguments{
\item{object}{the output of a \code{\link[glmnet]{glmnet}} function call}
}
\value{
a list of glmnet objects
}
\description{
Split the output object of a \code{\link[glmnet]{glmnet}} function call to have
one configuration (i.e. one lambda value) per object. It updates the main
slots of the object (\code{a0, beta, dfmat, df, dev.ratio, lambda, dim}) accordingly.
\code{npasses} is not updated.
}
\keyword{internal}
