% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renoir_nearest_shrunken_centroid.R
\name{renoir_split_nsc}
\alias{renoir_split_nsc}
\title{Split Trained Object}
\usage{
renoir_split_nsc(object)
}
\arguments{
\item{object}{an object of class \code{pamtrained}}
}
\value{
An object of class \code{RenoirSplitList}, which is a list
containing the objects resulting from the split.
}
\description{
This function splits a trained object containing multiple
configurations into several objects, one for each configuration.
}
\details{
The following elements stored in the object and linked to
the hyperparameter \code{threshold} are split:
yhat, prob, threshold, nonzero, errors
}
\author{
Alessandro Barberis
}
\keyword{internal}
