% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Trainer.R
\name{renoir_subset_glmnet}
\alias{renoir_subset_glmnet}
\title{Subset glmnet object}
\usage{
renoir_subset_glmnet(object, i)
}
\arguments{
\item{object}{a glmnet fit}

\item{i}{lambda}
}
\value{
Modified object.
}
\description{
Subset glmnet object
}
\keyword{internal}
