% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Trainer.R
\name{renoir_svm}
\alias{renoir_svm}
\title{Support Vector Machine}
\usage{
renoir_svm(
  x,
  y = NULL,
  weights = NULL,
  offset = NULL,
  resp.type,
  observations = NULL,
  features = NULL,
  clean = FALSE,
  keep.call = TRUE,
  ...,
  use.nusvm = FALSE,
  kernel = "radial",
  degree = 3,
  gamma = NULL,
  coef0 = 0,
  cost = 1,
  nu = 0.5,
  epsilon = 0.1,
  fitted = FALSE,
  verbose = F
)
}
\arguments{
\item{x}{the input matrix, where rows are observations and columns are variables}

\item{y}{the response variable. Its number of rows must match the number of rows of \code{x}}

\item{weights}{priors of the observations}

\item{offset}{used only for GLM methods, it is an a priori known component to be included in the linear predictor during fitting}

\item{resp.type}{the response type}

\item{observations}{indices of observations to keep}

\item{features}{indices of predictors to keep}

\item{...}{further arguments to \code{\link[e1071]{svm}}}

\item{kernel}{the kernel used in training and predicting. You
    might consider changing some of the following parameters, depending
    on the kernel type.\cr
    \describe{
      \item{linear:}{\eqn{u'v}{u'*v}}
      \item{polynomial:}{\eqn{(\gamma u'v + coef0)^{degree}}{(gamma*u'*v + coef0)^degree}}
      \item{radial basis:}{\eqn{e^(-\gamma |u-v|^2)}{exp(-gamma*|u-v|^2)}}
      \item{sigmoid:}{\eqn{tanh(\gamma u'v + coef0)}{tanh(gamma*u'*v + coef0)}}
      }
    }

\item{degree}{parameter needed for kernel of type \code{polynomial} (default: 3)}

\item{gamma}{parameter needed for all kernels except \code{linear}
    (default: 1/(data dimension))}

\item{coef0}{parameter needed for kernels of type \code{polynomial}
    and \code{sigmoid} (default: 0)}

\item{cost}{cost of costraint violation (must be >0)}

\item{nu}{re-parametrization of \code{cost} which controls the
number of support vectors and the margin errors.
It must be in range (0, 1]. Note that for classification nu must be
nu * length(y)/2 <= min(table(y))}

\item{epsilon}{epsilon in the insensitive-loss function (default: 0.1)}

\item{fitted}{logical indicating whether the fitted values should be computed
    and included in the model or not (default: \code{TRUE})}
}
\value{
An object of class \linkS4class{Trained}
}
\description{
Support Vector Machine
}
\author{
Alessandro Barberis
}
