% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Sampler.R
\name{resample,Sampler-method}
\alias{resample,Sampler-method}
\title{Resample}
\usage{
\S4method{resample}{Sampler}(object)
}
\value{
A \code{list} containing the samples
}
\description{
Takes a sample with or without replacement from the population.
}
