% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_generics.R, R/class_Resampler.R
\name{resample}
\alias{resample}
\alias{resample,Resampler-method}
\title{Resample}
\usage{
resample(object, ...)

\S4method{resample}{Resampler}(object, .inorder = TRUE, fun, ...)
}
\arguments{
\item{...}{further arguments to \code{fun}}

\item{.inorder}{logical flag indicating whether the \code{.combine}
function requires the task results to be combined in the same order
that they were submitted.  If the order is not important, then it
setting \code{.inorder} to \code{FALSE} can give improved performance.
The default value is `TRUE.}

\item{fun}{A function to be executed on re-sampled data. It should contain an \code{indices} parameter}

\item{do.parallel}{logical, whether to try to parallelise the train-and-test procedure}

\item{cores}{number of cores to use for parallel execution}
}
\value{
A \code{list} containing the output of function \code{fun}
}
\description{
Takes repeated samples with or without replacement from the population and feed them to function \code{fun}
}
\author{
Alessandro Barberis
}
