% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Resampler.R
\docType{class}
\name{Resampler-class}
\alias{Resampler-class}
\title{Resampler Class
An S4 class providing the re-sampling methods.}
\description{
The object consists of 2 slots
}
\section{Slots}{

\describe{
\item{\code{sampler}}{a \linkS4class{Sampler}}

\item{\code{logger}}{a \linkS4class{Logger}}

\item{\code{looper}}{a \linkS4class{Looper}}
}}

\author{
Alessandro Barberis
}
