% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reshape_recruitment_matrix}
\alias{reshape_recruitment_matrix}
\title{Reshape precruit}
\usage{
reshape_recruitment_matrix(x, annotation, feat.index = 0, check.names = TRUE)
}
\arguments{
\item{x}{data frame containing the probability of recruitment for each feature of the model}

\item{annotation}{a data frame with rownames (or one column) matching \code{x} rownames.}

\item{feat.index}{numeric, the index of the column in \code{annotation} matching \code{x} feature names.
\code{0} means rownames.}

\item{check.names}{logical.  If \code{TRUE} then the names of the
    variables in the data frame are checked to ensure that they are
    syntactically valid variable names and are not duplicated.
    If necessary they are adjusted (by \code{\link[base]{make.names}})
    so that they are.}
}
\value{
a data frame with at least 2 columns (features and probability of recruitment),
plus eventual annotation.
}
\description{
Reshape recruitment matrix to improve printing.
}
\keyword{internal}
