% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{root_mean_squared_error}
\alias{root_mean_squared_error}
\title{Root-Mean-Square Error}
\usage{
root_mean_squared_error(true, pred, weights, multi, ...)
}
\arguments{
\item{...}{arguments to \link{mean_squared_error}}
}
\value{
the root-mean-square error, a positive \code{double} or a vector of positive \code{double} values,
one for each response, if response has multiple output values and \code{multi = "raw"}
}
\description{
This function computes the root-mean-square error.
}
\details{
The root-mean-square error (RMSE) is a measure of errors based on squared losses.
It is defined as the squared root of the \link{mean_squared_error}:

\deqn{RMSE(y,\hat{y}) = \sqrt{MSE(y,\hat{y})} = \sqrt{\frac{1}{n}\sum_{i=1}^{n} (y_{i} - \hat{y}_{i})^{2}}}

If observation weights are provided, then the weighted root mean squared error is computed as

\deqn{wRMSE(w,y,\hat{y}) = \sqrt{wMSE(w,y,\hat{y})} = \sqrt{\frac{1}{\sum_{i=1}^{n} w_{i}}\sum_{i=1}^{n} w_{i} * (y_{i} - \hat{y}_{i})^{2}}}

This function is a shortcut for \code{mean_squared_error(root = TRUE, ...)}.
}
\author{
Alessandro Barberis
}
