% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam.R
\name{SAM.mod}
\alias{SAM.mod}
\title{Significance analysis of microarrays - simple user interface}
\usage{
SAM.mod(
  x,
  y = NULL,
  censoring.status = NULL,
  resp.type = c("Quantitative", "Two class unpaired", "Survival", "Multiclass",
    "One class", "Two class paired", "Two class unpaired timecourse",
    "One class timecourse", "Two class paired timecourse", "Pattern discovery"),
  geneid = NULL,
  genenames = NULL,
  s0 = NULL,
  s0.perc = NULL,
  nperms = 100,
  center.arrays = FALSE,
  testStatistic = c("standard", "wilcoxon"),
  time.summary.type = c("slope", "signed.area"),
  regression.method = c("standard", "ranks"),
  return.x = TRUE,
  knn.neighbors = 10,
  random.seed = NULL,
  logged2 = FALSE,
  fdr.output = 0.2,
  eigengene.number = 1,
  all.genes = FALSE
)
}
\arguments{
\item{x}{Feature matrix: p (number of features) by n (number of samples),
one observation per column (missing values allowed)}

\item{y}{n-vector of outcome measurements}

\item{censoring.status}{n-vector of censoring censoring.status (1= died or event occurred, 0=survived, or event was censored), needed for a censored survival outcome}

\item{resp.type}{Problem type:
"Quantitative" for a continuous parameter;
"Two class unpaired"; 
"Survival" for censored survival outcome; 
"Multiclass": more than 2 groups;  
"One class" for a single group; 
"Two class paired" for two classes with paired observations;
"Two class unpaired timecourse", 
"One class time course",
"Two class.paired timecourse",
"Pattern discovery"}

\item{geneid}{Optional character vector of geneids for output.}

\item{genenames}{Optional character vector of genenames for output.}

\item{s0}{Exchangeability factor  for denominator of test statistic; Default
is automatic choice. Only used for array data.}

\item{s0.perc}{Percentile of standard  deviation values to use for s0; default is
automatic choice; -1 means s0=0 (different from s0.perc=0, meaning
s0=zeroeth percentile of  standard  deviation values= min of sd values.
Only used for array data.}

\item{nperms}{Number of permutations used to estimate false discovery rates}

\item{center.arrays}{Should the data for each sample (array) be median centered
at the outset? Default =FALSE. Only used for array data.}

\item{testStatistic}{Test statistic to use in two class unpaired case.Either
"standard" (t-statistic) or ,"wilcoxon" (Two-sample wilcoxon or Mann-Whitney
test). Only used for array data.}

\item{time.summary.type}{Summary measure for each time course: "slope", or "signed.area"). Only used for array data.}

\item{regression.method}{Regression method for quantitative case:
"standard", (linear least squares) or "ranks" (linear least squares 
on ranked data). Only used for array data.}

\item{return.x}{Should the matrix of feature values be returned?
Only useful for time course data, where x contains summaries of the features
over time. Otherwise x is the same as the input data data\\$x}

\item{knn.neighbors}{Number of nearest neighbors to use for imputation
of missing features values. Only used for array data.}

\item{random.seed}{Optional initial seed for random number generator (integer)}

\item{logged2}{Has the data been transformed by log (base 2)? This information is used only 
for computing fold changes }

\item{fdr.output}{(Approximate) False Discovery Rate cutoff for output in significant genes table}

\item{eigengene.number}{Eigengene to be used  (just for resp.type="Pattern discovery")}

\item{all.genes}{Should all genes be listed? Default FALSE}
}
\description{
Modified from \code{\link[samr]{SAM}} to solve a bug.
Correlates a large number of features (eg genes) with an outcome variable,
such as a group indicator, quantitative variable or survival time.
This is a simple user interface for the samr function applied to array data.
For sequencing data applications, see the function SAMseq
}
\keyword{internal}
