% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam.R
\name{sam}
\alias{sam}
\title{Significance analysis of microarrays (wrapper function to \code{\link[samr]{SAM}})}
\usage{
sam(
  x,
  y = NULL,
  censoring.status = NULL,
  resp.type = c("Quantitative", "Two class unpaired", "Survival", "Multiclass",
    "One class", "Two class paired", "Two class unpaired timecourse",
    "One class timecourse", "Two class paired timecourse", "Pattern discovery"),
  geneid = NULL,
  genenames = NULL,
  assay.type = c("array", "seq"),
  s0 = 1e-04,
  s0.perc = NULL,
  nperms = 100,
  center.arrays = FALSE,
  testStatistic = c("standard", "wilcoxon"),
  time.summary.type = c("slope", "signed.area"),
  regression.method = c("standard", "ranks"),
  return.x = FALSE,
  knn.neighbors = 10,
  random.seed = NULL,
  logged2 = FALSE,
  fdr.output = 0.2,
  eigengene.number = 1,
  nresamp = 20,
  plotSam = FALSE,
  verbose = FALSE,
  outdir = NULL,
  logger = Logger(),
  all.genes = FALSE
)
}
\arguments{
\item{x}{Feature matrix: p (number of features) by n (number of samples), one observation per column (missing values allowed)}

\item{y}{n-vector of outcome measurements}

\item{censoring.status}{n-vector of censoring censoring.status
(1=died or event occurred, 0=survived, or event was censored), needed for a censored survival outcome}

\item{resp.type}{Problem type:\cr
\describe{
\item{\code{"Quantitative"}}{for a continuous parameter (Available for both array and sequencing data)}
\item{\code{"Two class unpaired"}}{for both array and sequencing data}
\item{\code{"Survival"}}{for censored survival outcome (for both array and sequencing data)}
\item{\code{"Multiclass"}}{more than 2 groups (for both array and sequencing data)}
\item{\code{"One class"}}{for a single group (only for array data)}
\item{\code{"Two class paired"}}{for two classes with paired observations (for both array and sequencing data)}
\item{\code{"Two class unpaired timecourse"}}{only for array data}
\item{\code{"One class time course"}}{only for array data}
\item{\code{"Two class.paired timecourse"}}{only for array data}
\item{\code{"Pattern discovery"}}{only for array data}
}}

\item{geneid}{Optional character vector of geneids for output.}

\item{genenames}{Optional character vector of genenames for output.}

\item{assay.type}{Assay type: "array" for microarray data, "seq" for counts from sequencing}

\item{s0}{Exchangeability factor for denominator of test statistic; Default in original function is a automatic choice.
Only used for array data. Our default is a small number since sometimes the automatic choice could get weird results.}

\item{s0.perc}{Percentile of standard deviation values to use for s0;
default is automatic choice;
-1 means s0=0 (different from s0.perc=0, meaning s0=zeroeth percentile of standard deviation values=
min of sd values. Only used for array data.}

\item{nperms}{Number of permutations used to estimate false discovery rates}

\item{center.arrays}{Should the data for each sample (array) be median centered at the outset? Default =FALSE. Only used for array data.}

\item{testStatistic}{Test statistic to use in two class unpaired case.
Either "standard" (t-statistic) or ,"wilcoxon" (Two-sample wilcoxon or Mann-Whitney test).
Only used for array data.}

\item{time.summary.type}{Summary measure for each time course: "slope", or "signed.area"). Only used for array data.}

\item{return.x}{= Should the matrix of feature values be returned? Only useful for time course data, where x contains summaries of the features over time. Otherwise x is the same as the input data data\$x}

\item{knn.neighbors}{Number of nearest neighbors to use for imputation of missing features values. Only used for array data.}

\item{random.seed}{Optional initial seed for random number generator (integer)}

\item{logged2}{Has the data been transformed by log (base 2)? This information is used only for computing fold changes}

\item{fdr.output}{(Approximate) False Discovery Rate cutoff for output in significant genes table}

\item{eigengene.number}{Eigengene to be used (just for resp.type="Pattern discovery")}

\item{plotSam}{whether to make Q-Q plot for SAM analysis, default TRUE.}

\item{outdir}{output directory}

\item{logger}{a \linkS4class{Logger}}

\item{all.genes}{Should all genes be listed? Default FALSE}

\item{nresapm}{For assay.type="seq", number of resamples used to construct test statistic. Default 20. Only used for sequencing data.
Default is equal to nresamp and it must be at most nresamp. Only used for sequencing data}
}
\value{
A list with components (the same as returned by \code{\link[samr]{SAM}})
}
\description{
Correlates a large number of features (eg genes) with an outcome variable,
such as a group indicator, quantitative variable or survival time by using \code{\link[samr]{SAM}}
function.
}
\keyword{internal}
