% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam.R
\name{samr.compute.siggenes.table.mod}
\alias{samr.compute.siggenes.table.mod}
\title{Compute significant genes table (with one bug corrected)}
\usage{
samr.compute.siggenes.table.mod(
  samr.obj,
  del,
  data,
  delta.table,
  min.foldchange = 0,
  all.genes = FALSE,
  compute.localfdr = FALSE
)
}
\arguments{
\item{samr.obj}{ Object returned from call to samr}

\item{del}{Value of delta to define cutoff rule}

\item{data}{Data object, same as that used in call to samr}

\item{delta.table}{ Object returned from call to
  samr.compute.delta.table}

\item{min.foldchange}{The minimum fold change desired; should be >1;
default is zero, meaning no fold change criterion is applied}

\item{all.genes}{Should all genes be listed? Default FALSE}

\item{compute.localfdr}{Should the local fdrs be computed (this can take some time)? Default FALSE}
}
\description{
Computes significant genes table, starting with samr object "samr.obj" and delta.table "delta.table"
Modified from \code{\link[samr]{samr.compute.siggenes.table}} to correct a bug
}
\keyword{internal}
