% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Scorer.R
\name{score,Scorer,ANY,ANY-method}
\alias{score,Scorer,ANY,ANY-method}
\title{Score}
\usage{
\S4method{score}{Scorer,ANY,ANY}(scorer, true, pred, weights = NULL, ...)
}
\arguments{
\item{scorer}{a \linkS4class{Scorer} object}

\item{true}{a vector (or a matrix) of observed values. If a matrix is provided,
a multi-response is assumed}

\item{pred}{a vector (or a matrix) of predicted values}

\item{weights}{observation weights}

\item{...}{further arguments to scorer function}
}
\value{
the computed accuracy measure
}
\description{
Score
}
