% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ScorerList.R
\name{score,ScorerList,ANY,ANY-method}
\alias{score,ScorerList,ANY,ANY-method}
\title{Score}
\usage{
\S4method{score}{ScorerList,ANY,ANY}(
  scorer,
  true,
  pred,
  weights = NULL,
  multi = c("average", "sum", "raw"),
  logger,
  ...
)
}
\arguments{
\item{scorer}{a \linkS4class{ScorerList} object}

\item{true}{a vector (or a matrix) of observed values. If a matrix is provided,
a multi-response is assumed}

\item{pred}{a vector (or a matrix) of predicted values}

\item{weights}{observation weights}

\item{multi}{what to do when response has multiple output values
\describe{
  \item{\code{average}}{errors of multiple outputs are averaged to get a single value for each observation}
  \item{\code{sum}}{errors of multiple outputs are summed up to get a single value for each observation}
  \item{\code{raw}}{returns a vector containing one error for each output}
}}

\item{logger}{a \linkS4class{Logger} object}

\item{...}{further arguments to scorer function}
}
\value{
the computed accuracy measure for each \linkS4class{Scorer} in the \linkS4class{ScorerList}
}
\description{
Score
}
