% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_generics.R, R/class_Scorer.R,
%   R/class_ScorerList.R
\name{score}
\alias{score}
\alias{score,Scorer,ANY,ANY-method}
\alias{score,Scorer,list,list-method}
\alias{score,ScorerList,ANY,ANY-method}
\alias{score,ScorerList,list,list-method}
\title{Score}
\usage{
score(scorer, true, pred, weights, ...)

\S4method{score}{Scorer,ANY,ANY}(scorer, true, pred, weights = NULL, ...)

\S4method{score}{Scorer,list,list}(
  scorer,
  true,
  pred,
  weights = NULL,
  multi = c("average", "sum"),
  grouped = TRUE,
  min.obs = 3,
  logger,
  ...
)

\S4method{score}{ScorerList,ANY,ANY}(
  scorer,
  true,
  pred,
  weights = NULL,
  multi = c("average", "sum", "raw"),
  logger,
  ...
)

\S4method{score}{ScorerList,list,list}(
  scorer,
  true,
  pred,
  weights = NULL,
  multi = c("average", "sum"),
  grouped = TRUE,
  logger,
  ...
)
}
\arguments{
\item{scorer}{a \linkS4class{ScorerList} object}

\item{true}{a list of vectors (or matrices) of observed values. If list elements are matrices,
a multi-response is assumed}

\item{pred}{a list of vectors (or matrices) of predicted values}

\item{weights}{a list of vectors of observation weights}

\item{...}{further arguments to scorer function}

\item{multi}{what to do when response has multiple output values
\describe{
\item{\code{average}}{errors of multiple outputs are averaged to get a single value for each observation}
\item{\code{sum}}{errors of multiple outputs are summed up to get a single value for each observation}
}}

\item{grouped}{logical, whether to compute separate statistics when lists of values are provided}

\item{min.obs}{integer, minimum number of observations per list. If the number of
observations per list is less than \code{min.obs}, an attempt to compute errors at the
observation level (for the unlisted observations)
and then summarise is made (equivalent to \code{grouped = FALSE}). It is working only for certain
accuracy measures. Default is \code{min.obs = 3}. Set \code{min.obs = 0} to suppress.}

\item{logger}{a \linkS4class{Logger} object}
}
\value{
the computed accuracy measure

the computed accuracy measure

the computed accuracy measure for each \linkS4class{Scorer} in the \linkS4class{ScorerList}

the computed accuracy measure for each \linkS4class{Scorer} in the \linkS4class{ScorerList}
}
\description{
Computes scores for different sets of data

Computes scores for different sets of data
}
