% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Scorer.R
\docType{class}
\name{Scorer-class}
\alias{Scorer-class}
\alias{Scorer}
\title{Scorer Class}
\usage{
Scorer(
  id,
  scorer,
  parameters,
  resp.type,
  grouped = TRUE,
  selector,
  optimum,
  quantifier = Quantifier(),
  logger = Logger()
)
}
\arguments{
\item{id}{the scoring method name associated to this \linkS4class{Scorer}.
If scoring method is one of the accuracy measures supported by renoir, the constructor will
automatically select a \code{scorer} and a \code{selector}. See \code{supported_scoring_methods()}
for the supported methods.}

\item{scorer}{(optional) function to evaluate a model performance.
Must be provided if \code{id} is not one of the accuracy measures supported by renoir.
If \code{scorer} is provided it must conform to the renoir common interface,
and must have the following formal arguments:
\describe{
\item{true}{a vector (or a matrix) of observed values. If a matrix is provided, a multi-response is assumed}
\item{pred}{a vector (or a matrix) of predicted values}
\item{weights}{priors of the observations}
\item{multi}{what to do when response has multiple output values}
\item{...}{additional arguments can be added}
}}

\item{parameters}{list containing the parameters to fix for the chosen scoring method}

\item{resp.type}{(optional) response type. If provided, the selected \code{id} is checked
to see if available for the considered response type.}

\item{grouped}{logical, whether to compute separate statistics when lists of values are provided}

\item{selector}{(optional) function to select optimal accuracy measure from multiple values
Must be provided if \code{id} is not one of the accuracy measures supported by renoir.
If \code{selector} is provided it must conform to the renoir common interface,
and must have the following formal arguments:
\describe{
\item{measures}{a vector of accuracy measures returned by this \linkS4class{Scorer}}
}}

\item{optimum}{the optimisation problem to solve for selecting the optimal accuracy measure
from multiple values}

\item{quantifier}{a \linkS4class{Quantifier}}

\item{logger}{a \linkS4class{Logger}}
}
\description{
An S4 class providing the methods to compute an accuracy measure.

The object consists of 6 slots

Constructor for the S4 \linkS4class{Scorer} object.
}
\section{Slots}{

\describe{
\item{\code{id}}{the accuracy measure}

\item{\code{scorer}}{function to be used for scoring}

\item{\code{parameters}}{list containing the parameters to fix for the scoring function}

\item{\code{grouped}}{logical, whether to compute separate statistics when lists of values are provided}

\item{\code{selector}}{function to select optimal measure given a vector of values}

\item{\code{optimum}}{the optimisation problem to solve for selecting the optimal accuracy measure.
Two possible values are \code{min} and \code{max}}

\item{\code{quantifier}}{a \linkS4class{Quantifier}}

\item{\code{logger}}{a \linkS4class{Logger}}
}}

\author{
Alessandro Barberis
}
