% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ScorerList.R
\docType{class}
\name{ScorerList-class}
\alias{ScorerList-class}
\alias{ScorerList}
\alias{create_ScorerList}
\title{ScorerList Class}
\usage{
ScorerList(...)

create_ScorerList(
  resp.type,
  measures,
  grouped = TRUE,
  parameters,
  logger = Logger()
)
}
\arguments{
\item{...}{\linkS4class{Scorer} objects}

\item{resp.type}{response type}
}
\value{
a \linkS4class{ScorerList} object

a \linkS4class{ScorerList}
}
\description{
An S4 class to represent a scorer list.

Constructor for the S4 ScorerList object.

Creates a list of supported scorers for the selected response type
}
\details{
This subclass extends the S4  \linkS4class{List} virtual class by using
the  \linkS4class{SimpleList} implementation.

The object consists of 4 slots

Constructor for the S4 \linkS4class{ScorerList} object.
}
\section{Slots}{

\describe{
\item{\code{listData}}{\code{list} storing the list elements}

\item{\code{elementType}}{the type of data represented in the sequence}

\item{\code{elementMetadata}}{\linkS4class{DataFrame} storing the element-wise metadata,
with a row for each element and a column for each metadata available. Default is \code{NULL}}

\item{\code{metadata}}{\code{list} storing the global metadata annotating the object as a whole}
}}

