% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ScorerList.R
\name{ScorerList}
\alias{ScorerList}
\title{ScorerList Constructor}
\usage{
ScorerList(...)
}
\arguments{
\item{...}{\linkS4class{Scorer} objects}
}
\value{
a \linkS4class{ScorerList} object
}
\description{
Constructor for the S4 ScorerList object.
}
\details{
Constructor for the S4 \linkS4class{ScorerList} object.
}
