% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_generics.R, R/class_Screener.R,
%   R/class_ScreenerList.R
\name{screen}
\alias{screen}
\alias{screen,Screener-method}
\alias{screen,ScreenerList-method}
\title{Features screening}
\usage{
screen(screener, ...)

\S4method{screen}{Screener}(
  screener,
  x,
  y,
  weights = NULL,
  resp.type,
  observations = NULL,
  multi = c("max", "average", "sum"),
  order = c("increasing", "decreasing", "none"),
  cutoff = NULL,
  maxvars = NULL,
  logger,
  ...
)

\S4method{screen}{ScreenerList}(
  screener,
  x,
  y,
  weights = NULL,
  resp.type,
  observations = NULL,
  combine = c("max", "average", "sum"),
  order = c("increasing", "decreasing"),
  cutoff = NULL,
  maxvars = NULL,
  logger = Logger(verbose = F),
  ...
)
}
\arguments{
\item{screener}{a \linkS4class{ScreenerList} object}

\item{...}{further arguments to \code{screener} function}

\item{x}{the input matrix, where rows are observations and columns are variables.}

\item{y}{the response variable. Its number of rows must match the number of rows of \code{x}.}

\item{weights}{priors of the observations}

\item{resp.type}{the response type}

\item{observations}{(optional) indices of observations to keep}

\item{multi}{what to do when response has multiple output values
\describe{
\item{\code{max}}{the max value of scores across multiple outputs is selected to get a single value for each observation}
\item{\code{average}}{scores of multiple outputs are averaged to get a single value for each observation}
\item{\code{sum}}{scores of multiple outputs are summed up to get a single value for each observation}
}}

\item{order}{whether to order the results by the score values
\describe{
\item{\code{increasing}}{results are sorted in increasing order}
\item{\code{decreasing}}{results are sorted in decreasing order}
\item{\code{none}}{results are not sorted}
}}

\item{cutoff}{(optional) numeric, the value to use as a threshold for filtering not significant variables}

\item{maxvars}{integer, the maximum number of variables to keep. If the filtering is returning an higher number of variables,
only the most significant \code{maxvars} variables are kept.}

\item{logger}{a \linkS4class{Logger}}

\item{combine}{how to combine the results from multiple screeners
\describe{
\item{\code{max}}{the max value of scores across multiple \linkS4class{Screened} objects is selected to get a single value for each observation}
\item{\code{average}}{scores of multiple \linkS4class{Screened} objects are averaged to get a single value for each observation}
\item{\code{sum}}{scores of multiple \linkS4class{Screened} objects are summed up to get a single value for each observation}
}}

\item{coef}{(optional) an integer indicating the response variable to consider in multi-response data
when \code{multi = "raw"}}
}
\value{
a \linkS4class{Screened} object

a \linkS4class{Screened} object
}
\description{
This function performs a feature screening, and return a \linkS4class{Screened} object

This function performs a feature screening by using the provided screeners. It combines
the results from the different screeners and return a \linkS4class{Screened} object
}
\author{
Alessandro Barberis
}
