% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ScreenerList.R
\name{ScreenerList}
\alias{ScreenerList}
\title{ScreenerList Constructor}
\usage{
ScreenerList(...)
}
\arguments{
\item{...}{\linkS4class{Screener} objects to be stored in a \linkS4class{ScreenerList}}
}
\value{
a \linkS4class{ScreenerList} object
}
\description{
Constructor for the S4 ScreenerList object.
}
\details{
Constructor for the S4 \linkS4class{ScreenerList} object.
}
\author{
Alessandro Barberis
}
