% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Screener.R
\name{screening_by_default}
\alias{screening_by_default}
\title{Compute the significance of the filtering test}
\usage{
screening_by_default(
  x,
  y,
  weights = NULL,
  alternative = "two.sided",
  method = "pearson",
  conf.level = 0.95,
  resp.type = c("gaussian", "mgaussian", "binomial", "multinomial", "poisson", "cox"),
  observations = NULL,
  coef = NULL,
  logger,
  multi = c("raw", "max", "average", "sum"),
  ...
)
}
\arguments{
\item{x}{numeric matrix}

\item{y}{response type (must be a factor for binomial and multinomial)}

\item{alternative}{alternative hypotesis to use. Must be one of \code{"two.sided"} (default),
\code{"greater"} or \code{"less"}.}

\item{method}{which correlation method to use. Currently, only "pearson" is supported.}

\item{conf.level}{confidence levels used for the confidence intervals (where computed). A single number or a
numeric vector with value for each observation. All values must be in the range of [0;1].}

\item{...}{further arguments}
}
\value{
the p-value of the selected test
}
\description{
This function compute the significance of the screening test
}
\keyword{internal}
