% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Screener.R
\name{screening_by_ebayes}
\alias{screening_by_ebayes}
\title{Compute the significance of the filtering test}
\usage{
screening_by_ebayes(
  x,
  y,
  weights = NULL,
  resp.type = c("gaussian", "mgaussian", "binomial", "multinomial", "poisson", "cox"),
  observations = NULL,
  logged2 = FALSE,
  coef = NULL,
  design = NULL,
  return.type = c("pvalue", "object"),
  logger = Logger(verbose = F),
  multi = c("raw", "max", "average", "sum"),
  ...
)
}
\arguments{
\item{x}{A matrix-like data object containing log-ratios or log-expression
values for a series of arrays, with rows corresponding to genes and columns to samples.}

\item{y}{the response variable}

\item{weights}{vector of observation weights}

\item{resp.type}{the response type}

\item{logged2}{whether data was logged 2}

\item{coef}{column number or column name specifying which coefficient or contrast of the linear
model is of interest. For topTable, can also be a vector of column subscripts, in which
case the gene ranking is by F-statistic for that set of contrasts.
If coef has a single value, then the moderated t-statistics and p-values for that coefficient
or contrast are used.
If coef takes two or more values, the moderated F-statistics for that set of coefficients or
contrasts are used. If coef is left NULL, then all the coefficients or contrasts in the fitted
model are used, except that any coefficient named (Intercept) will be removed.}

\item{design}{the design matrix of the microarray experiment, with rows corresponding to arrays
and columns to coefficients to be estimated. Defaults to the unit vector meaning
that the arrays are treated as replicates.}

\item{logger}{a \linkS4class{Logger}}

\item{multi}{what to do when response has multiple output values
\describe{
  \item{\code{max}}{for each feature, the maximal significance value across responses is kept}
  \item{\code{average}}{significances of multiple outputs are averaged to get a single value for each feature}
  \item{\code{sum}}{significances of multiple outputs are summed up to get a single value for each feature}
  \item{\code{raw}}{for each feature, returns one score for each output}
}}

\item{...}{further arguments to \code{\link[renoir]{limma}}}
}
\value{
The p-value associated with each feature or NULL if the response type is not supported
}
\description{
This function compute the significance of the screening test by using
\code{\link[renoir]{compute_limma}}
}
\keyword{internal}
