% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Screener.R
\name{screening_default_binomial}
\alias{screening_default_binomial}
\title{Compute the significance of the filtering test for the binomial family}
\usage{
screening_default_binomial(x, y, conf.level = 0.95, ...)
}
\description{
Compute the significance of the filtering test for the binomial family
}
\keyword{internal}
