% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Screener.R
\name{screening_default_cox}
\alias{screening_default_cox}
\title{Compute the significance of the filtering test for the cox family}
\usage{
screening_default_cox(x, y, weights = NULL, ...)
}
\arguments{
\item{...}{further arguments to \code{\link[survival]{coxph}}}
}
\description{
Compute the significance of the filtering test for the cox family
}
\keyword{internal}
