% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Screener.R
\name{screening_default_gaussian}
\alias{screening_default_gaussian}
\title{Screener Class
An S4 class providing the methods to perform a features screening.}
\usage{
screening_default_gaussian(
  x,
  y,
  alternative = "two.sided",
  method = "pearson",
  conf.level = 0.95
)
}
\arguments{
\item{x}{the input matrix, where rows are observations and columns are variables.}

\item{observations}{(optional) indices of observations to keep}
}
\value{
the index of features to keep.
}
\description{
This function performs a feature screening, and return the index of the retained elements
in order of significance.
}
\details{
The object consists of 10 slots
}
\section{Slots}{

\describe{
\item{\code{method}}{the type of screening to apply
\describe{
  \item{\code{default}}{if default is selected, the adopted test is different depending on the response type:
       \describe{
           \item{\code{gaussian}}{test is correlation}
           \item{\code{mgaussian}}{test is correlation (for each gene, one correlation is computed with each response variable)}
           \item{\code{binomial}}{test is t-test or Wilcox test, depending on the number of observations}
           \item{\code{multinomial}}{test is anova or Kruskall-Wallis test, depending on the number of observations}
           \item{\code{poisson}}{test is correlation}
           \item{\code{cox}}{test is a Cox proportional hazards regression model}
       }
  }
  \item{\code{ebayes}}{}
  \item{\code{permutation}}{}
}}

\item{\code{resp.type}}{the response type}

\item{\code{x}}{the input matrix, where rows are observations and columns are variables.}

\item{\code{y}}{the response variable. Its number of rows must match the number of rows of \code{x}.}

\item{\code{cutoff}}{numeric, the p-value to use as a threshold for filtering not significant variables.}

\item{\code{coef}}{(optional) an integer indicating the response variable to consider in multi-response data}

\item{\code{maxvars}}{integer, the maximum number of variables to keep. If the filtering is returning an higher number of variables,
only the most significant \code{maxvars} variables are kept.}

\item{\code{default}}{list, arguments to the function performing a default screening}

\item{\code{ebayes}}{list, arguments to the function performing a screening based on empirical Bayes methods}

\item{\code{permutation}}{list, arguments to the function performing a screening based on permutation}

\item{\code{logger}}{a \linkS4class{Logger}}
}}

\author{
Alessandro Barberis
Features screening

Alessandro Barberis
Compute the significance of the filtering test for the gaussian family
}
\keyword{internal}
