% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Screener.R
\name{screening_default_mgaussian}
\alias{screening_default_mgaussian}
\title{Compute the significance of the filtering test for the multivariate gaussian family}
\usage{
screening_default_mgaussian(
  x,
  y,
  alternative = "two.sided",
  method = "pearson",
  conf.level = 0.95
)
}
\description{
Compute the significance of the filtering test for the multivariate gaussian family
}
\keyword{internal}
