% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Screener.R
\name{screening_default_multinomial}
\alias{screening_default_multinomial}
\title{Compute the significance of the filtering test for the multinomial family}
\usage{
screening_default_multinomial(x, y, ...)
}
\description{
Compute the significance of the filtering test for the multinomial family
}
\keyword{internal}
