% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Selector.R
\name{select,Selector,TrainedList,numeric,numeric,Scorer-method}
\alias{select,Selector,TrainedList,numeric,numeric,Scorer-method}
\title{Select a model}
\usage{
\S4method{select}{Selector,TrainedList,numeric,numeric,Scorer}(selector, models, merr, sderr, scorer, ...)
}
\arguments{
\item{selector}{a \linkS4class{Selector} object}

\item{models}{a \linkS4class{TrainedList} object containing trained models.
Each element corresponds to an element in \code{merr} and \code{sderr}.}

\item{merr}{vector of error estimates}

\item{sderr}{vector of standard errors of the estimates}

\item{scorer}{the \linkS4class{Scorer} object used to generate \code{merr} and \code{sderr}}
}
\value{
list of 2 elements, containing the index of the model having optimal accuracy
and the index of the model having accuracy within 1 standard error from the optimal accuracy
}
\description{
Selects a model from a list
}
