% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Selector.R
\name{select_model}
\alias{select_model}
\title{Model Selector}
\usage{
select_model(
  merr,
  sderr,
  type.measure,
  trainedlist,
  rule = c("opt", "1se"),
  FUN
)
}
\arguments{
\item{merr}{vector of error estimates}

\item{sderr}{vector of standard errors}

\item{type.measure}{accuracy measure used in the computation of \code{merr} and \code{sderr}}

\item{trainedlist}{list of trained models. Each element corresponds to an element in \code{merr} and \code{sderr},
and it is a list of trained models with the same configuration over different sampled data}
}
\description{
Model Selector
}
