% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ForecasterList.R
\name{setAs,Forecaster-method}
\alias{setAs,Forecaster-method}
\title{Coerce to ForecasterList}
\arguments{
\item{from}{a \linkS4class{Forecaster}}
}
\value{
a \linkS4class{ForecasterList}
}
\description{
Function to coerce a \linkS4class{Forecaster} to a \linkS4class{ForecasterList}
}
\keyword{internal}
