% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence.R
\name{sewm}
\alias{sewm}
\title{Standard Error of Weighted Mean}
\usage{
sewm(x, weights, var.est = "default", ...)
}
\arguments{
\item{x}{vector of measurements}

\item{weights}{vector of weights}

\item{var.est}{a character string, the estimator for the variance of the weighted mean}

\item{...}{further arguments to \code{var.est}}
}
\value{
A length-one numeric vector.
}
\description{
Computes the Standard Error of the weighted sample mean.
}
\details{
The standard error of the sample mean is its standard deviation, i.e.
the square root of the variance of the estimate

\deqn{SEM = \hat{\sigma}_{W}}

Be aware that while stratification during sampling is not affecting the
estimate of the mean, it may affect its standard error.

Different options for the calculation of the variance are here provided.
}
\seealso{
\code{\link{uevarwm}} for more information on the default estimate
of the variance.
}
\author{
Alessandro Barberis
}
