% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_generics.R, R/class_Renoir.R
\name{signature}
\alias{signature}
\alias{signature,Renoir,numeric,numeric,missing,missing,missing-method}
\alias{signature,Renoir,missing,missing,numeric,character,character-method}
\alias{signature,Renoir,missing,missing,missing,character,character-method}
\alias{signature,Renoir,missing,missing,numeric,missing,missing-method}
\alias{signature,Renoir,missing,missing,missing,missing,missing-method}
\alias{signature,Renoir,numeric,numeric,numeric,missing,missing-method}
\title{Extract a signature}
\usage{
signature(object, index, n, cutoff, measure, set, ...)

\S4method{signature}{Renoir,numeric,numeric,missing,missing,missing}(object, index, n, cutoff, measure, set, grouped = F)

\S4method{signature}{Renoir,missing,missing,numeric,character,character}(
  object,
  index,
  n,
  cutoff,
  measure,
  set = c("train", "test", "full"),
  ...
)

\S4method{signature}{Renoir,missing,missing,missing,character,character}(
  object,
  index,
  n,
  cutoff,
  measure,
  set = c("train", "test", "full"),
  ...
)

\S4method{signature}{Renoir,missing,missing,numeric,missing,missing}(
  object,
  index,
  n,
  cutoff,
  measure,
  set = c("train", "test", "full"),
  ...
)

\S4method{signature}{Renoir,missing,missing,missing,missing,missing}(object, index, n, cutoff, measure, set = c("train", "test", "full"))

\S4method{signature}{Renoir,numeric,numeric,numeric,missing,missing}(object, index, n, cutoff, grouped = F)
}
\arguments{
\item{grouped}{logical, whether to merge results from a multi-response
to obtain a single set of features}
}
\description{
Features with non-zero coefficients (where applicable) from
selected model are considered as signature
}
