% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_generics.R, R/class_EvaluatedList.R
\name{stability}
\alias{stability}
\alias{stability,EvaluatedList-method}
\title{Stability}
\usage{
stability(object, ...)

\S4method{stability}{EvaluatedList}(object)
}
\arguments{
\item{object}{an object of class \linkS4class{EvaluatedList}}
}
\description{
Stability
}
