% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stts}
\alias{stts}
\title{Summary Table to String}
\usage{
stts(object, sep.el = ": ", collapse.el = ", ", sep.col = "<br>")
}
\arguments{
\item{object}{a \code{data.frame}}

\item{sep.el}{a character string to separate the terms in a column from the column name}

\item{collapse.el}{a character string to separate the terms in a column}

\item{sep.col}{a character string to separate the resulting strings}
}
\value{
A character string
}
\description{
Converts a summary table into a
character string.
}
\details{
This function create a string containing the elements
of a \code{data.frame} so that each column is reported as
"colname: element1<sep.el>element2<sep.el>...". The different strings are then pasted together,
separated by the \code{sep.col} parameter.
}
