% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ForecasterList.R
\name{subset_forecaster,ForecasterList,character-method}
\alias{subset_forecaster,ForecasterList,character-method}
\title{Subset forecaster}
\usage{
\S4method{subset_forecaster}{ForecasterList,character}(object, by)
}
\arguments{
\item{object}{a S4 \linkS4class{ForecasterList} object}

\item{by}{the learning method to match. If missing, the \code{object} is returned
as it is. In case of no match \code{NULL} is returned.}
}
\value{
a \linkS4class{ForecasterList} object
}
\description{
Select the forecasters from the list matching the learning method in \code{by}
}
\keyword{internal}
