% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{summarise_features_as_single_score}
\alias{summarise_features_as_single_score}
\title{Summarisation of features as single score}
\usage{
summarise_features_as_single_score(
  x,
  features = NULL,
  method = c("mean", "median", "pc1", "gsva", "ssgsea", "plage", "zscore"),
  na.rm = T,
  log.warning = F,
  ...
)
}
\arguments{
\item{x}{the input matrix, where rows are observations and columns are variables}

\item{features}{the features to summarise. If \code{features} is an integer vector,
values are considered as \code{x} column indices, whether if is a character vector
it must match the \code{x} column names. If \code{NULL}, then all the columns of
\code{x} are considered}

\item{method}{the method to use for summarising the set of features. Six methods
are currently supported}

\item{na.rm}{logical, whether to remove NAs}

\item{...}{further arguments to \code{method}}
}
\value{
a named (if \code{x} has row names) vector containing one summary
score per observation
}
\description{
Summarise the set of features as a single score to
allow easy comparison between observations
}
\examples{
\dontrun{
nvar = 20
nobs = 100
x = matrix(
  data = rnorm(n = nobs*nvar, mean = 3, sd = 0.2), nrow = nobs, ncol = nvar,
  dimnames = list(paste0("O",seq(nobs)), paste0("V", seq(nvar)))
)

sscore = summarise_as_single_score(x = x, features = colnames(x)[1:5], method = "mean")
sscore = summarise_as_single_score(x = x, features = colnames(x)[1:5], method = "median")
sscore = summarise_as_single_score(x = x, features = colnames(x)[1:5], method = "pc1")
sscore = summarise_as_single_score(x = x, features = colnames(x)[1:5], method = "gsva")
sscore = summarise_as_single_score(x = x, features = colnames(x)[1:5], method = "ssgsea")
sscore = summarise_as_single_score(x = x, features = colnames(x)[1:5], method = "plage")
}

}
\author{
Alessandro Barberis
}
\keyword{internal}
