% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Evaluated.R
\name{summary_table,Evaluated-method}
\alias{summary_table,Evaluated-method}
\title{Get the Summary of an Evaluated Object}
\usage{
\S4method{summary_table}{Evaluated}(object, best = c("opt", "1se"), key)
}
\description{
Get the Summary of an Evaluated Object
}
\author{
Alessandro Barberis
}
