% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_EvaluatedList.R
\name{summary_table,EvaluatedList-method}
\alias{summary_table,EvaluatedList-method}
\title{Get the Summary of an EvaluatedList Object}
\usage{
\S4method{summary_table}{EvaluatedList}(object, key, best = c("opt", "1se"), ...)
}
\description{
Get the Summary of an EvaluatedList Object
}
\author{
Alessandro Barberis
}
