% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Filtered.R
\name{summary_table,Filtered-method}
\alias{summary_table,Filtered-method}
\title{Summary Table of a Filtered Object}
\usage{
\S4method{summary_table}{Filtered}(object)
}
\arguments{
\item{object}{an object of class \linkS4class{Filtered}}
}
\value{
A data frame with the following elements:
\describe{
\item{filtering}{character string, the filtering strategy}
\item{filtered}{character, a textual summary of the performed pre-processing}
}
}
\description{
Summary Table of a Filtered Object
}
\author{
Alessandro Barberis
}
