% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Renoir.R
\name{summary_table,Renoir-method}
\alias{summary_table,Renoir-method}
\title{Get Object Summary}
\usage{
\S4method{summary_table}{Renoir}(object, ...)
}
\arguments{
\item{object}{an object of class \linkS4class{Renoir}}

\item{...}{further arguments to \code{\link{summary_table.EvaluatedList}}}
}
\value{
A \code{data.frame}
}
\description{
Get Object Summary
}
\author{
Alessandro Barberis
}
