% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Screened.R
\name{summary_table,Screened-method}
\alias{summary_table,Screened-method}
\title{Summary Table of a Screened Object}
\usage{
\S4method{summary_table}{Screened}(object)
}
\arguments{
\item{object}{an object of class \linkS4class{Screened}}
}
\value{
A data frame with the following elements:
\describe{
\item{\code{screening}}{character string, the screening strategy}
\item{\code{screened}}{integer, the number of screened features}
}
}
\description{
Summary Table of a Screened Object
}
\author{
Alessandro Barberis
}
