% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Tested.R
\name{summary_table,Tested-method}
\alias{summary_table,Tested-method}
\title{Get the Summary of a Tested Object}
\usage{
\S4method{summary_table}{Tested}(object, best = c("opt", "1se"))
}
\description{
Get the Summary of a Tested Object
}
\author{
Alessandro Barberis
}
