% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_TestedList.R
\name{summary_table,TestedList-method}
\alias{summary_table,TestedList-method}
\title{Get the Summary of a TestedList Object}
\usage{
\S4method{summary_table}{TestedList}(object, best)
}
\description{
Get the Summary of a TestedList Object
}
