% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Trained.R
\name{summary_table,Trained-method}
\alias{summary_table,Trained-method}
\title{Summary Table of a Trained Object}
\usage{
\S4method{summary_table}{Trained}(object)
}
\arguments{
\item{object}{an object of class \linkS4class{Trained}}
}
\value{
A data frame with the following elements:
\describe{
\item{\code{learning}}{character string, the learning method}
\item{\code{nfeatures}}{integer, the number of features}
\item{\code{configuration}}{string, the configuration, i.e. the fixed hyperparameters of the model}
}
}
\description{
Summary Table of a Trained Object
}
\author{
Alessandro Barberis
}
