% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Evaluated.R
\name{summary_table.Evaluated}
\alias{summary_table.Evaluated}
\title{Summary Table of an Evaluated Object}
\usage{
summary_table.Evaluated(object, best = c("opt", "1se"), key)
}
\arguments{
\item{object}{object of class \linkS4class{Evaluated}}

\item{best}{character string, the model selection rule to consider (\code{"best"} or \code{"1se"})}

\item{key}{character vector indicating the elements to use for the name generation.
Allowed strings are \code{"id"}, \code{"config"}, \code{"response"}, and \code{"sampling"}}
}
\value{
A \code{data.frame} with the following elements
\describe{
\item{\code{name}}{character string indicating a name for the considered settings. It is generated by \code{\link{generate_name}}}
\item{\code{config}}{name of the selected tuned configuration}
\item{\code{response}}{response type}
\item{\code{sampling}}{sampling strategy}
\item{\code{training_set_size}}{training-set size}
\item{\code{k}}{number of iteration of the sampling procedure}
\item{\code{tuning}}{character string, the tuning strategy}
\item{\code{screening}}{character string, the screening strategy}
\item{\code{screened}}{integer, the number of screened features}
\item{\code{learning}}{character string, the learning method}
\item{\code{nfeatures}}{integer, the number of features in the model}
\item{\code{configuration}}{string, the configuration, i.e. the fixed hyperparameters of the model}
\item{\code{imodel}}{index of the model}
\item{\code{set}}{name of the set}
\item{\code{measure}}{name of the considered performance metric}
\item{\code{score}}{performance metrics}
\item{\code{mean_score}}{mean performance metric}
\item{\code{standard_error}}{standard error of the mean performance metric}
\item{\code{upper_ci}}{upper bound of confidence interval}
\item{\code{lower_ci}}{lower bound of confidence interval}
\item{\code{best_model}}{logical indicating which is the best model}
\item{\code{selection_rule}}{character string indicating the rule adopted for selection of tuned/trained model across k runs}
}
}
\description{
This function produces a summary table of an object of class \linkS4class{Evaluated}.
}
\seealso{
\code{\link{summary_table.Renoir}},
\code{\link{summary_table.EvaluatedList}},
\code{\link{summary_table.TestedList}},
\code{\link{summary_table.Tested}},
\code{\link{summary_table.Tuned}},
\code{\link{summary_table.Trained}},
\code{\link{generate_name}}
}
\author{
Alessandro Barberis
}
\keyword{internal}
