% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_EvaluatedList.R
\name{summary_table.EvaluatedList}
\alias{summary_table.EvaluatedList}
\title{Get EvaluatedList Object Summary}
\usage{
summary_table.EvaluatedList(object, key, best = c("opt", "1se"), ...)
}
\arguments{
\item{object}{object of class \linkS4class{EvaluatedList}, a list of
\linkS4class{Evaluated} elements}

\item{key}{(optional) character vector indicating the data to use for uniquely
identifying the evaluated elements from the list.
It can be a combination of \code{"id"}, \code{"config"},
\code{"response"}, and \code{"sampling"}. If missing, it is internally
identified with \code{\link{identify_primary_key_from_EvaluatedList}}}
}
\value{
A \code{data.frame} with the following elements
\describe{
\item{\code{name}}{character string indicating a name for the considered settings. It is generated by \code{\link{generate_name}}}
\item{\code{config}}{name of the configuration}
\item{\code{response}}{response type}
\item{\code{sampling}}{sampling strategy}
\item{\code{training_set_size}}{training-set size}
\item{\code{k}}{number of iteration of the sampling procedure}
\item{\code{tuning}}{character string, the tuning strategy}
\item{\code{screening}}{character string, the screening strategy}
\item{\code{screened}}{integer, the number of screened features}
\item{\code{learning}}{character string, the learning method}
\item{\code{nfeatures}}{integer, the number of features in the model}
\item{\code{configuration}}{string, the configuration, i.e. the fixed hyperparameters of the model}
\item{\code{imodel}}{index of the model}
\item{\code{set}}{name of the set}
\item{\code{measure}}{name of the considered performance metric}
\item{\code{score}}{performance metrics}
\item{\code{mean_score}}{mean performance metric}
\item{\code{standard_error}}{standard error of the mean performance metric}
\item{\code{upper_ci}}{upper bound of confidence interval}
\item{\code{lower_ci}}{lower bound of confidence interval}
\item{\code{best_model}}{logical indicating which is the best model}
\item{\code{selection_rule}}{character string indicating the adopted model selection rule}
}
}
\description{
This function produces a summary of an object of class \linkS4class{EvaluatedList}.
}
\seealso{
\code{\link{summary_table.Renoir}},
\code{\link{summary_table.Evaluated}},
\code{\link{summary_table.TestedList}},
\code{\link{summary_table.Tested}},
\code{\link{summary_table.Tuned}},
\code{\link{summary_table.Trained}}
}
\author{
Alessandro Barberis
}
