% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Filtered.R
\name{summary_table.Filtered}
\alias{summary_table.Filtered}
\title{Summary Table of a Filtered Object}
\usage{
summary_table.Filtered(object)
}
\arguments{
\item{object}{an object of class \linkS4class{Filtered}}
}
\value{
A data frame with the following elements:
\describe{
\item{filtering}{character string, the filtering strategy}
\item{filtered}{character, a textual summary of the performed pre-processing}
}
}
\description{
This function produces a summary table of an object of class \linkS4class{Filtered}.
}
\seealso{
\code{\link{summary_table.Renoir}},
\code{\link{summary_table.Evaluated}},
\code{\link{summary_table.EvaluatedList}},
\code{\link{summary_table.Tested}}
\code{\link{summary_table.TestedList}},
\code{\link{summary_table.Screened}}
\code{\link{summary_table.Trained}}
}
\author{
Alessandro Barberis
}
