% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Renoir.R
\name{summary_table}
\alias{summary_table}
\title{Object Summaries}
\arguments{
\item{object}{an object for which a summary is desired}

\item{...}{additional arguments affecting the summary produced}
}
\value{
The form of the returned object depends on the class of its
argument. See the documentation of the particular method for
further details.
}
\description{
Generic function used to produce summary tables of various
objects generated by `renoir`. The function invokes particular
methods which depend on the class of the first argument.
}
\details{
Methods for different classes are implemented, e.g.
\linkS4class{Renoir}, \linkS4class{EvaluatedList}, \linkS4class{Evaluated},
\linkS4class{TestedList}, \linkS4class{Tested}.
}
\seealso{
\code{\link{summary_table.Renoir}},
\code{\link{summary_table.EvaluatedList}},
\code{\link{summary_table.Evaluated}},
\code{\link{summary_table.TestedList}},
\code{\link{summary_table.Tested}},
\code{\link{summary_table.Tuned}},
\code{\link{summary_table.Trained}},
\code{\link{summary_table.Screened}}
}
