% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Renoir.R
\name{summary_table.Renoir}
\alias{summary_table.Renoir}
\title{Get Renoir Object Summary}
\usage{
summary_table.Renoir(object, ...)
}
\arguments{
\item{object}{an object of class \linkS4class{Renoir}}

\item{...}{further arguments to \code{\link{summary_table.EvaluatedList}}}
}
\value{
A \code{data.frame}
}
\description{
This function produces a summary of an object of class \linkS4class{Renoir}.
}
\seealso{
\code{\link{summary_table.EvaluatedList}},
\code{\link{summary_table.Evaluated}},
\code{\link{summary_table.TestedList}},
\code{\link{summary_table.Tested}}
\code{\link{summary_table.Tuned}}
\code{\link{summary_table.Trained}}
}
\author{
Alessandro Barberis
}
\keyword{internal}
