% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Screened.R
\name{summary_table.Screened}
\alias{summary_table.Screened}
\title{Summary Table of a Screened Object}
\usage{
summary_table.Screened(object)
}
\arguments{
\item{object}{an object of class \linkS4class{Screened}}
}
\value{
A data frame with the following elements:
\describe{
\item{\code{screening}}{character string, the screening strategy}
\item{\code{screened}}{integer, the number of screened features}
}
}
\description{
This function produces a summary table of an object of class \linkS4class{Screened}.
}
\seealso{
\code{\link{summary_table.Renoir}},
\code{\link{summary_table.Evaluated}},
\code{\link{summary_table.EvaluatedList}},
\code{\link{summary_table.Tested}}
\code{\link{summary_table.TestedList}},
\code{\link{summary_table.Tuned}}
}
\author{
Alessandro Barberis
}
