% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Tested.R
\name{summary_table.Tested}
\alias{summary_table.Tested}
\title{Get the Summary of a Tested Object}
\usage{
summary_table.Tested(object, best = c("opt", "1se"))
}
\arguments{
\item{object}{object of class \linkS4class{Tested}}

\item{best}{character string, the model selection rule to consider (\code{"best"} or \code{"1se"})}
}
\value{
A \code{data.frame} with the following elements
\describe{
\item{\code{set}}{name of the set}
\item{\code{measure}}{name of the considered performance metric}
\item{\code{score}}{performance metrics}
\item{\code{mean_score}}{mean performance metric}
\item{\code{standard_error}}{standard error of the mean performance metric}
\item{\code{upper_ci}}{upper bound of confidence interval}
\item{\code{lower_ci}}{lower bound of confidence interval}
\item{\code{best_model}}{logical indicating which is the best model}
\item{\code{imodel}}{index of the model}
\item{\code{selection_rule}}{character string indicating the adopted model selection rule}
}
}
\description{
This function produces a summary of an object of class \linkS4class{Tested}.
}
\seealso{
\code{\link{summary_table.Renoir}},
\code{\link{summary_table.Evaluated}},
\code{\link{summary_table.EvaluatedList}},
\code{\link{summary_table.TestedList}},
\code{\link{summary_table.Tuned}},
\code{\link{summary_table.Trained}}
}
\author{
Alessandro Barberis
}
\keyword{internal}
