% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Trained.R
\name{summary_table.Trained}
\alias{summary_table.Trained}
\title{Summary Table of a Trained Object}
\usage{
summary_table.Trained(object)
}
\arguments{
\item{object}{an object of class \linkS4class{Trained}}
}
\value{
A data frame with the following elements:
\describe{
\item{\code{learning}}{character string, the learning method}
\item{\code{nfeatures}}{integer, the number of features}
\item{\code{configuration}}{string, the configuration, i.e. the fixed hyperparameters of the model}
}
}
\description{
This function produces a summary table of an object of class \linkS4class{Trained}.
}
\seealso{
\code{\link{summary_table.Renoir}},
\code{\link{summary_table.Evaluated}},
\code{\link{summary_table.EvaluatedList}},
\code{\link{summary_table.Tested}}
\code{\link{summary_table.TestedList}},
\code{\link{summary_table.Tuned}}
}
\author{
Alessandro Barberis
}
\keyword{internal}
