% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Tuned.R
\name{summary_table.Tuned}
\alias{summary_table.Tuned}
\title{Summary Table of a Tuned Object}
\usage{
summary_table.Tuned(object)
}
\arguments{
\item{object}{an object of class \linkS4class{Tuned}}
}
\value{
A data frame with the following elements:
\describe{
\item{\code{tuning}}{character string, the tuning strategy}
\item{\code{screening}}{character string, the screening strategy}
\item{\code{screened}}{integer, the number of screened features}
\item{\code{learning}}{character string, the learning method}
\item{\code{nfeatures}}{integer, the number of features}
\item{\code{configuration}}{string, the configuration, i.e. the fixed hyperparameters of the model}
}
}
\description{
This function produces a summary table of an object of class \linkS4class{Tuned}.
}
\seealso{
\code{\link{summary_table.Renoir}},
\code{\link{summary_table.Evaluated}},
\code{\link{summary_table.EvaluatedList}},
\code{\link{summary_table.Tested}}
\code{\link{summary_table.TestedList}},
\code{\link{summary_table.Screened}}
\code{\link{summary_table.Trained}}
}
\author{
Alessandro Barberis
}
\keyword{internal}
