% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Trainer.R
\name{supported_learning_methods}
\alias{supported_learning_methods}
\title{List supported learning methods}
\usage{
supported_learning_methods(x = c("all", "regression", "classification"))
}
\arguments{
\item{x}{character string indicating whether to return all learning methods,
the learning method for regression, or the learning methods for classification.}
}
\value{
A vector containing the supported learning methods.
}
\description{
This function lists the available learning methods.
}
\keyword{internal}
