% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Scorer.R
\name{switch_to_supported_ungrouped_measure}
\alias{switch_to_supported_ungrouped_measure}
\title{Switch accuracy}
\usage{
switch_to_supported_ungrouped_measure(measure)
}
\value{
measure name
}
\description{
Utility function used to switch accuracy measure when
\code{grouped = FALSE}
}
