% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Tester.R
\name{test,TrainedList,Tester,missing,missing,list-method}
\alias{test,TrainedList,Tester,missing,missing,list-method}
\title{Test}
\usage{
\S4method{test}{TrainedList,Tester,missing,missing,list}(models, tester, forecaster, scorer, indices, ...)
}
\arguments{
\item{models}{a list of \linkS4class{TrainedList} objects}

\item{tester}{a \linkS4class{Tester} object}

\item{indices}{list of indices used to subset x, y, weights, offset when training the model}

\item{...}{further arguments to scorer function}

\item{screened}{list of screened features}

\item{newx}{the predictors matrix}

\item{newoffset}{offset}

\item{newy}{the response}

\item{weights}{observations weights (i.e. response prior)}

\item{looper}{a \linkS4class{Looper} object}

\item{logger}{a \linkS4class{Logger} object}

\item{score.type}{the score type to return, one of
\describe{
  \item{\code{score}}{the accuracy measure}
  \item{\code{mean_score}}{estimate of the mean error obtained from the samples}
  \item{\code{summary_score}}{estimate of the mean error and its standard error}
}}
}
\value{
list of 2 elements containing a score estimate and its standard error
}
\description{
Tests the trained model on the provided data
}
