% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_generics.R, R/class_Tester.R
\name{test}
\alias{test}
\alias{test,TrainedList,Tester,ANY,ANY,ANY-method}
\alias{test,Trained,Tester,missing,missing,missing-method}
\alias{test,Trained,missing,Forecaster,Scorer,missing-method}
\alias{test,Trained,missing,Forecaster,ANY,numeric-method}
\alias{test,Trained,missing,Forecaster,ScorerList,missing-method}
\alias{test,TrainedList,missing,Forecaster,ScorerList,missing-method}
\alias{test,TrainedList,missing,Forecaster,Scorer,missing-method}
\alias{test,TrainedList,Tester,missing,missing,list-method}
\alias{test,TrainedList,missing,Forecaster,Scorer,list-method}
\alias{test,TrainedList,missing,Forecaster,ScorerList,list-method}
\alias{test,ANY,Tester,missing,missing,ANY-method}
\alias{test,TunedList,missing,Forecaster,Scorer,list-method}
\alias{test,TunedList,missing,Forecaster,ScorerList,list-method}
\alias{test,list,missing,Forecaster,Scorer,list-method}
\title{Test}
\usage{
test(models, tester, forecaster, scorer, indices, ...)

\S4method{test}{TrainedList,Tester,ANY,ANY,ANY}(models, tester, forecaster, scorer, indices, ...)

\S4method{test}{Trained,Tester,missing,missing,missing}(models, tester, logger, ...)

\S4method{test}{Trained,missing,Forecaster,Scorer,missing}(
  models,
  forecaster,
  scorer,
  newx,
  newoffset,
  newy,
  weights = NULL,
  score.type = c("mean_score", "summary_score", "score", "score_and_summary"),
  min.obs = 3,
  logger,
  ...
)

\S4method{test}{Trained,missing,Forecaster,ANY,numeric}(
  models,
  forecaster,
  scorer,
  newx,
  newoffset,
  newy,
  weights = NULL,
  indices,
  ...
)

\S4method{test}{Trained,missing,Forecaster,ScorerList,missing}(
  models,
  forecaster,
  scorer,
  newx,
  newoffset,
  newy,
  weights = NULL,
  score.type = c("mean_score", "summary_score", "score", "score_and_summary"),
  min.obs = 3,
  logger = Logger(verbose = FALSE),
  ...
)

\S4method{test}{TrainedList,missing,Forecaster,ScorerList,missing}(models, forecaster, scorer, logger, ...)

\S4method{test}{TrainedList,missing,Forecaster,Scorer,missing}(models, forecaster, scorer, logger, ...)

\S4method{test}{TrainedList,Tester,missing,missing,list}(models, tester, forecaster, scorer, indices, ...)

\S4method{test}{TrainedList,missing,Forecaster,Scorer,list}(
  models,
  forecaster,
  scorer,
  indices,
  screened = NULL,
  newx,
  newoffset = NULL,
  newy,
  weights = NULL,
  score.type = c("mean_score", "summary_score", "score", "score_and_summary"),
  min.obs = 3,
  looper = Looper(),
  logger,
  check.input = FALSE,
  ...
)

\S4method{test}{TrainedList,missing,Forecaster,ScorerList,list}(
  models,
  forecaster,
  scorer,
  indices,
  screened = NULL,
  newx,
  newoffset = NULL,
  newy,
  weights = NULL,
  score.type = c("mean_score", "summary_score", "score", "score_and_summary"),
  min.obs = 3,
  looper,
  logger,
  ...
)

\S4method{test}{ANY,Tester,missing,missing,ANY}(models, tester, forecaster, scorer, indices, ...)

\S4method{test}{TunedList,missing,Forecaster,Scorer,list}(
  models,
  forecaster,
  scorer,
  indices,
  screened = NULL,
  newx,
  newoffset = NULL,
  newy,
  weights = NULL,
  score.type = c("mean_score", "summary_score", "score", "score_and_summary"),
  min.obs = 3,
  looper = Looper(),
  logger,
  ...
)

\S4method{test}{TunedList,missing,Forecaster,ScorerList,list}(
  models,
  forecaster,
  scorer,
  indices,
  screened = NULL,
  newx,
  newoffset = NULL,
  newy,
  weights = NULL,
  score.type = c("mean_score", "summary_score", "score", "score_and_summary"),
  min.obs = 3,
  looper = Looper(),
  logger,
  ...
)

\S4method{test}{list,missing,Forecaster,Scorer,list}(
  models,
  forecaster,
  scorer,
  indices,
  screened = NULL,
  newx,
  newoffset = NULL,
  newy,
  weights = NULL,
  score.type = c("mean_score", "summary_score", "score", "score_and_summary"),
  min.obs = 3,
  looper = Looper(),
  logger,
  ...
)
}
\arguments{
\item{models}{a list of \linkS4class{TrainedList} objects}

\item{tester}{a \linkS4class{Tester} object}

\item{indices}{list of indices used to subset x, y, weights, offset when training the model}

\item{...}{further arguments to scorer function}

\item{logger}{a \linkS4class{Logger} object}

\item{newx}{the predictors matrix}

\item{newoffset}{offset}

\item{newy}{the response}

\item{weights}{observations weights (i.e. response prior)}

\item{score.type}{the score type to return, one of
\describe{
\item{\code{score}}{the accuracy measure}
\item{\code{mean_score}}{estimate of the mean error obtained from the samples}
\item{\code{summary_score}}{estimate of the mean error and its standard error}
}}

\item{screened}{list of screened features}

\item{looper}{a \linkS4class{Looper} object}
}
\value{
list of 2 elements containing a score estimate and its standard error
}
\description{
Tests the trained model on the provided data
}
