% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Tested.R
\name{Tested}
\alias{Tested}
\title{Constructor for the S4 Tested object.}
\usage{
Tested(
  set = character(),
  measure = character(),
  score = numeric(),
  mscore = numeric(),
  sem = numeric(),
  ci = numeric(),
  opt = integer(),
  `1se` = integer()
)
}
\arguments{
\item{set}{name of the set}

\item{measure}{name of the considered performance metric}

\item{score}{performance metrics}

\item{mscore}{mean of the performance metrics}

\item{sem}{standard error of the mean}

\item{ci}{confidence interval}

\item{opt}{index of model associated to optimal performance}

\item{1se}{index of least complex model associated to performance within 1SE from the optimal}
}
\description{
Constructor for the S4 \linkS4class{Tested} object.
}
\author{
Alessandro Barberis
}
