% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_TestedList.R
\name{TestedList}
\alias{TestedList}
\title{TestedList Constructor
Constructor for the S4 \linkS4class{TestedList} object.}
\usage{
TestedList(...)
}
\arguments{
\item{...}{a comma-separated list of S4 \linkS4class{TestedList} objects.}
}
\description{
TestedList Constructor
Constructor for the S4 \linkS4class{TestedList} object.
}
