% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Tester.R
\name{Tester}
\alias{Tester}
\title{Constructor for the S4 Tester object.}
\usage{
Tester(forecaster, scorer, logger = Logger())
}
\arguments{
\item{forecaster}{a \linkS4class{ForecasterList}}

\item{scorer}{a \linkS4class{ScorerList}}

\item{logger}{a \linkS4class{Logger}}

\item{looper}{a \linkS4class{Looper}}
}
\description{
Constructor for the S4 \linkS4class{Tester} object.
}
