% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_TesterList.R
\name{TesterList}
\alias{TesterList}
\title{Constructor for the S4 TesterList object.

Constructor for the S4 \linkS4class{TesterList} object.}
\usage{
TesterList(...)
}
\description{
TesterList Constructor
}
