% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{TNR}
\alias{TNR}
\title{True negative rate, specificity (SPC), selectivity, (1 - FPR)}
\usage{
TNR(TP, TN, FP, FN)
}
\arguments{
\item{TP}{number of true positives}

\item{TN}{number of true negatives}

\item{FP}{number of false positives}

\item{FN}{number of false negatives}
}
\value{
A numeric value.
}
\description{
This function computes the true negative rate.
}
\author{
Alessandro Barberis
}
