% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{TPR}
\alias{TPR}
\title{True Positive Rate}
\usage{
TPR(TP, TN, FP, FN)
}
\arguments{
\item{TP}{number of true positives}

\item{TN}{number of true negatives}

\item{FP}{number of false positives}

\item{FN}{number of false negatives}
}
\value{
A numeric value, the true positive rate.
}
\description{
This function computes the true positive rate. It is also known as
recall, sensitivity, probability of detection, power, hit rate (1 - FNR).
}
\details{
The true positive rate measures the ability of a classifier
of correctly predicting the presence of a condition.
It is defined as

\deqn{sensitivity = recall = \textit{true positive rate} (TPR) = \frac{TP}{P} = \frac{TP}{TP + FN} = 1 - FNR}

The optimal value is 1 and the worst value is 0.
}
\author{
Alessandro Barberis
}
