% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_generics.R, R/class_Trainer.R
\name{train}
\alias{train}
\alias{train,Trainer,ANY-method}
\alias{train,Trainer,missing-method}
\alias{train,Trainer,list-method}
\title{Train the model}
\usage{
train(trainer, features, ...)

\S4method{train}{Trainer,ANY}(
  trainer,
  x,
  y,
  weights = NULL,
  offset = NULL,
  resp.type,
  observations = NULL,
  features = NULL,
  ...
)

\S4method{train}{Trainer,missing}(trainer, features, ...)

\S4method{train}{Trainer,list}(
  trainer,
  x,
  y,
  weights = NULL,
  offset = NULL,
  resp.type,
  observations = NULL,
  features,
  logger = Logger(verbose = FALSE),
  ...
)
}
\arguments{
\item{trainer}{a \linkS4class{Trainer} object providing the learning method}

\item{features}{(optional) indices of predictors to keep}

\item{...}{additional arguments to training method}

\item{x}{the input matrix, where rows are observations and columns are variables.}

\item{y}{the response variable. Its number of rows must match the number of rows of \code{x}.}

\item{weights}{priors of the observations}

\item{offset}{used only for GLM methods, it is an a priori known component to be included in the linear predictor during fitting}

\item{resp.type}{the response type}

\item{observations}{(optional) indices of observations to keep}
}
\value{
a \linkS4class{Trained} object or a \linkS4class{TrainedList} object if learning method
runs multiple configurations (i.e. multiple sets of hyper-parameters).

Results are merged into one \linkS4class{TrainedList} object
}
\description{
Common interface to train a model. The learning method is provided in \code{trainer}
}
\author{
Alessandro Barberis
}
\keyword{internal}
