% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Trained.R
\name{Trained}
\alias{Trained}
\title{Constructor for the S4 Trained object.}
\usage{
Trained(
  fit = NULL,
  config = list(),
  learning.method = character(),
  nfeatures = integer()
)
}
\description{
Constructor for the S4 \linkS4class{Trained} object.
}
