% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_TrainedList.R
\name{TrainedList}
\alias{TrainedList}
\title{TrainedList Constructor
Constructor for the S4 TrainedList object.}
\usage{
TrainedList(...)
}
\description{
Constructor for the S4 \linkS4class{TrainedList} object.
}
