% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_tests.R
\name{ttest}
\alias{ttest}
\title{Student's t-Test}
\usage{
ttest(x, g, alternative = "two.sided", ...)
}
\description{
Computes the two sample t-tests for each variable and returns the relative
non-adjusted p-value.
}
\keyword{internal}
