% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Tuner.R
\name{tune_via_grid_search}
\alias{tune_via_grid_search}
\title{Tune via grid search}
\usage{
tune_via_grid_search(
  sampler,
  looper = Looper(),
  screener,
  trainer,
  tester,
  selector,
  hyperparameters,
  logger,
  x,
  y,
  weights,
  offset,
  resp.type,
  screening = NULL,
  keep.fit = FALSE,
  rule = c("opt", "1se"),
  observations = NULL,
  marker,
  recorder,
  ...
)
}
\arguments{
\item{sampler}{a \linkS4class{Sampler} object, the sampling procedure to use for tuning}

\item{looper}{a \linkS4class{Looper} object}

\item{screener}{a \linkS4class{Screener} object}

\item{trainer}{a \linkS4class{Trainer} object}

\item{tester}{a \linkS4class{Tester} object}

\item{selector}{a \linkS4class{Selector} object}

\item{hyperparameters}{list containing the hyperparameters to be tuned}

\item{logger}{a \linkS4class{Logger} object}

\item{x}{the input matrix, where rows are observations and columns are variables.}

\item{y}{the response variable. Its number of rows must match the number of rows of \code{x}.}

\item{weights}{priors of the observations}

\item{offset}{used only for GLM methods, it is an a priori known component to be included in the linear predictor during fitting}

\item{resp.type}{the response type}

\item{screening}{(optional) integer, if provided a features screening is performed before fitting.
If an integer vector is provided then the number of features to select before fitting is
considered an hyperparameter for the screener and will be tuned}

\item{keep.fit}{logical, whether to keep the trained objects returned by each resampling fit.}

\item{rule}{character string, the model selection rule to apply}

\item{observations}{(optional) integer vector, providing the indices of observations to keep.
If missing the entire set of observations is considered.}

\item{marker}{a \linkS4class{Marker} object}

\item{recorder}{a \linkS4class{Recorder} object}

\item{...}{further arguments to the train method}
}
\description{
This function tunes the hyperparameters of a learning method via
a grid-search strategy.
}
\author{
Alessandro Barberis
}
\keyword{internal}
