% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Tuner.R
\docType{class}
\name{Tuner-class}
\alias{Tuner-class}
\title{Tuner Class}
\description{
An S4 class providing the methods to perform the tuning of
hyperparameters and the training of the final model.

The object consists of 6 slots
}
\section{Slots}{

\describe{
\item{\code{id}}{the name of the tuning method to use}

\item{\code{tuner}}{function to use for tuning the hyperparameters of the model}

\item{\code{parameters}}{list containing the parameters to fix for the chosen tuning method}

\item{\code{logger}}{a \linkS4class{Logger}}

\item{\code{sampler}}{a \linkS4class{Sampler} object, the sampling procedure to use for tuning}

\item{\code{looper}}{a \linkS4class{Looper} object}

\item{\code{screener}}{a \linkS4class{ScreenerList} object}
}}

\author{
Alessandro Barberis
}
