% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_best_model_selection}
\alias{update_best_model_selection}
\title{Update best model selection}
\usage{
update_best_model_selection(
  object,
  tolerance = 0,
  set = c("test.set", "train.set", "full.set")
)
}
\arguments{
\item{object}{an object of class \code{renoir}}

\item{tolerance}{the new tolerance to use for the selection of the best model}

\item{set}{which set of assessment data to consider for the selection of the best model
(\code{'test.set'} by default)}
}
\value{
an updated \code{renoir} object
}
\description{
Update best model selection
}
\keyword{internal}
