% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_features_importance}
\alias{update_features_importance}
\title{Update best model selection}
\usage{
update_features_importance(
  object,
  recruitment.term = c("coefficient", "presence")
)
}
\arguments{
\item{object}{an object of class \code{renoir}}

\item{recruitment.term}{whether to use a binary variable (\code{presence = 0} or \code{presence = 1})
multiplied by the sign of the coefficient, or to use directly the coefficient of a feature
in the computation of the significance score.}
}
\value{
an updated \code{renoir} object
}
\description{
Update best model selection
}
\keyword{internal}
