% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_EvaluatedList.R
\name{which_best_evaluated}
\alias{which_best_evaluated}
\title{Which is the best training set size}
\usage{
which_best_evaluated(
  object,
  set,
  measure,
  evaluator,
  scorer,
  confidence,
  distribution
)
}
\arguments{
\item{object}{\linkS4class{EvaluatedList} object}

\item{scorer}{a \linkS4class{Scorer} object or a \linkS4class{ScorerList} object
containing the \linkS4class{Scorer} object that computed the \code{measure} of interest}
}
\value{
integer indicating the index of the best \linkS4class{Evaluated} object
in the input list
}
\description{
Select best \linkS4class{Evaluated} object from a list.
The selection is performed by looking at the optimal confidence interval
(defined as the confidence interval with minimum upper confidence limit,
or maximum lower confidence limit, depending on the considered accuracy measure)
}
