% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_model_selection.R
\name{which_best_glmnet}
\alias{which_best_glmnet}
\title{Which is the best model}
\usage{
which_best_glmnet(object, best.config = c("min", "1se"), measure, type.measure)
}
\arguments{
\item{object}{a list of \code{renoir.trained} objects}

\item{best.config}{which configuration to consider}

\item{measure}{a vector containing accuracy measures}

\item{type.measure}{the type of accuracy measure}

\item{learning.method}{method for fitting the data}
}
\value{
the index of the best model

the index of the best model
}
\description{
Given a list of models, it returns the best fit according to the
criterium that a model with less coefficients is better

Given a list of models, it returns the best fit
}
\keyword{Which}
\keyword{best}
\keyword{internal}
\keyword{is}
\keyword{model}
\keyword{the}
