% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_model_selection.R
\name{which_best_resampling_iteration}
\alias{which_best_resampling_iteration}
\title{Which is the best model}
\usage{
which_best_resampling_iteration(
  trained,
  assessed,
  learning.method,
  best.config = c("min", "1se"),
  tolerance = 0
)
}
\arguments{
\item{trained}{a list of objects of class \code{RenoirTrainedAndTested}, where each element
corresponds to a different run over the same training set size}

\item{assessed}{a list of assessment measures, one for each model in \code{trained}}

\item{learning.method}{method for fitting the data}

\item{tolerance}{numeric (range [0,1]), the tolerance in the selection of the best model}
}
\value{
the index of the best model for each accuracy measure, for both best configurations
}
\description{
This function returns the index of the best model over a list
of models obtained via a resampling procedure.
The function looks over the performance measures and
for each of them select the best model, i.e. the model which minimises
or maximises the measure. If a tolerance is given, the function compares
all models having an accuracy measure within the range
\eqn{[error - tolerance*error, error + tolerance*error]}
and selects the one with less coefficients.
}
\keyword{internal}
