% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_model_selection.R
\name{which_best_training_set_size}
\alias{which_best_training_set_size}
\title{Which is the best training set size}
\usage{
which_best_training_set_size(
  object,
  set = c("test.set", "train.set", "full.set")
)
}
\arguments{
\item{object}{a list of objects of class \code{renoir.tat}, where each element
correspons to a different training set size}

\item{set}{which set of assessment data to consider for the selection of the best training set size.
The available options are \code{'test.set'} (default), \code{'train.set'} and \code{'full.set'}
(which consists of \code{test.set + train.set})}
}
\value{
the index of the best training set size
}
\description{
This function returns the index of the best training set size from a
list of \code{renoir.tat} object.
}
\keyword{internal}
