% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_model_selection.R
\name{which_best_training_set_sizes}
\alias{which_best_training_set_sizes}
\title{Which is the best training set size}
\usage{
which_best_training_set_sizes(object)
}
\arguments{
\item{object}{a list of objects of class \code{renoir.tat}, where each element
correspons to a different training set size}
}
\value{
a list of lists containing the indices of the best training set size for each
set of data ('test.set', 'train.set' and 'full.set')
\describe{
  \item{\code{test.set}}{a list of 2 elements (one for best configuration),
  each being a list where each element is the index of the best training set size
  associated to a specific accuracy measure computed by using the test.set}
  \item{\code{train.set}}{a list of 2 elements (one for best configuration),
  each being a list where each element is the index of the best training set size
  associated to a specific accuracy measure computed by using the train.set}
  \item{\code{full.set}}{a list of 2 elements (one for best configuration),
  each being a list where each element is the index of the best training set size
  associated to a specific accuracy measure computed by using the full.set}
}
}
\description{
This function returns the indices of the best training set size from a
list of \code{renoir.tat} objects.
}
\keyword{internal}
