% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_tests.R
\name{wtest}
\alias{wtest}
\title{Wilcoxon Rank Sum and Signed Rank Tests}
\usage{
wtest(x, g, alternative = "two.sided", ...)
}
\arguments{
\item{x}{numeric vector of data values.  Non-finite (e.g., infinite or
    missing) values will be omitted.}

\item{alternative}{a character string specifying the alternative
    hypothesis, must be one of \code{"two.sided"} (default),
    \code{"greater"} or \code{"less"}.  You can specify just the initial
    letter.}

\item{...}{further arguments to be passed to or from methods.}
}
\description{
Wilcoxon Rank Sum and Signed Rank Tests
}
\keyword{internal}
