% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Evaluator.R
\docType{class}
\name{Evaluator-class}
\alias{Evaluator-class}
\alias{Evaluator}
\title{Evaluator Class}
\usage{
Evaluator(
  sampler = Sampler(method = "random", k = 1L),
  looper = Looper(),
  scorer = ScorerList(Scorer(id = "mse")),
  logger = Logger()
)
}
\arguments{
\item{sampler}{a \linkS4class{Sampler}}

\item{looper}{a \linkS4class{Looper}}

\item{scorer}{a \linkS4class{ScorerList} or a \linkS4class{Scorer} object}

\item{logger}{a \linkS4class{Logger}}
}
\value{
An object of class \linkS4class{Evaluator}.
}
\description{
An S4 class providing the methods to perform an evaluation of a learning method.

The object consists of 3 slots

Constructor for the S4 \linkS4class{Evaluator} object.
}
\section{Slots}{

\describe{
\item{\code{sampler}}{a \linkS4class{Sampler} object}

\item{\code{looper}}{a \linkS4class{Looper} object}

\item{\code{scorer}}{a \linkS4class{ScorerList} object}

\item{\code{logger}}{a \linkS4class{Logger} object}
}}

\author{
Alessandro Barberis
}
