% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Forecaster.R
\docType{class}
\name{Forecaster-class}
\alias{Forecaster-class}
\title{Forecaster Class
An S4 class providing the methods to test the trained models on provided data.}
\description{
The object consists of 3 slots
}
\section{Slots}{

\describe{
\item{\code{id}}{the object identifier, it must be equivalent to the learning method name (i.e. the trainer id)}

\item{\code{forecaster}}{function to use to compute predictions.
It must have the following formal arguments:
\describe{
   \item{object}{a model for which prediction is desired.
   It is a \linkS4class{Trained} object as returned by the training process,
   where the fit is stored in the \code{fit} slot}
   \item{newx}{the new values at which prediction are to be made}
   \item{type}{the type of prediction required}
   \item{newoffset}{if an offset was used in the fit then one must be provided}
   \item{...}{additional arguments affecting the predictions produced}
}}

\item{\code{selector}}{helper function to select the appropriate prediction type given
the response type and an accuracy measure
It must have the following formal arguments:
\describe{
   \item{learning.method}{the learning method, used to internally select the correct prediction type}
   \item{type.measure}{the accuracy type measure, used to internally select the correct prediction type}
}}
}}

\author{
Alessandro Barberis
}
