% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_generics.R, R/class_Evaluated.R,
%   R/class_EvaluatedList.R, R/class_Renoir.R
\name{get_model}
\alias{get_model}
\alias{get_model,Evaluated-method}
\alias{get_model,EvaluatedList-method}
\alias{get_model,Renoir-method}
\title{Get Trained/Tuned Model}
\usage{
get_model(object, ...)

\S4method{get_model}{Evaluated}(object, index)

\S4method{get_model}{EvaluatedList}(object, n, index)

\S4method{get_model}{Renoir}(object, n, index)
}
\arguments{
\item{object}{an object of class \linkS4class{Renoir}}

\item{index}{index of model to select}

\item{n}{sample size (i.e. training set size) to select}
}
\value{
a \linkS4class{Trained} or \linkS4class{Tuned} object

a \linkS4class{Trained} object

a \linkS4class{Trained} or \linkS4class{Tuned} object
}
\description{
This function extracts a model from a
\linkS4class{Evaluated} object.

This function extracts a model from a
\linkS4class{EvaluatedList} object.

This function extracts a model from a
\linkS4class{Renoir} object.
}
\author{
Alessandro Barberis
}
