% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Logger.R
\docType{class}
\name{Logger-class}
\alias{Logger-class}
\title{Logger Class}
\description{
An S4 class providing a logger.
}
\section{Slots}{

\describe{
\item{\code{path}}{path to the log file to use}

\item{\code{con}}{\code{file}}

\item{\code{verbose}}{logical, indicating whether R should report extra information on progress}

\item{\code{level}}{character, the log level}
}}

\author{
Alessandro Barberis
}
