% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_pca}
\alias{plot_pca}
\title{Plot PCA}
\usage{
plot_pca(
  PCA,
  groups,
  pc = c(1, 2),
  title = NULL,
  palette = NULL,
  palette.type = "qual"
)
}
\arguments{
\item{PCA}{object as returned by \code{\link[stats]{prcomp}}}

\item{groups}{data frame with 1 column containing the groups}

\item{pc}{vector with 2 elements, the principal components to plot}

\item{title}{plot title}

\item{palette}{(optional) when a single value,
if a string, it is considered as a name of palette;
if a number, it is the index into the lists of palettes of appropriate type.
The parameter is passed to \code{\link[ggplot2]{scale_colour_brewer}}}

\item{palette.type}{palette type. One of seq (sequential), div (diverging) or qual (qualitative)}
}
\description{
Plot PCA
}
\keyword{internal}
