% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Scorer.R
\docType{class}
\name{Scorer-class}
\alias{Scorer-class}
\title{Scorer Class
An S4 class providing the methods to compute an accuracy measure.}
\description{
The object consists of 6 slots
}
\section{Slots}{

\describe{
\item{\code{id}}{the accuracy measure}

\item{\code{scorer}}{function to be used for scoring}

\item{\code{parameters}}{list containing the parameters to fix for the scoring function}

\item{\code{grouped}}{logical, whether to compute separate statistics when lists of values are provided}

\item{\code{selector}}{function to select optimal measure given a vector of values}

\item{\code{optimum}}{the optimisation problem to solve for selecting the optimal accuracy measure.
Two possible values are \code{min} and \code{max}}

\item{\code{quantifier}}{a \linkS4class{Quantifier}}

\item{\code{logger}}{a \linkS4class{Logger}}
}}

\author{
Alessandro Barberis
}
