% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Screened.R
\docType{class}
\name{Screened-class}
\alias{Screened-class}
\title{Screened Class}
\description{
An S4 class providing a container for the results of a screening procedure.
}
\section{Slots}{

\describe{
\item{\code{method}}{the used screening method}

\item{\code{n}}{the maximum number of variables kept}

\item{\code{index}}{the index of kept features}

\item{\code{score}}{the score values resulting from screening procedure}
}}

\author{
Alessandro Barberis
}
