% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Screener.R
\docType{class}
\name{Screener-class}
\alias{Screener-class}
\title{Screener Class
An S4 class providing the methods to perform a features screening.}
\description{
The object consists of 3 slots
}
\section{Slots}{

\describe{
\item{\code{id}}{a name to identify the screener}

\item{\code{screener}}{screener function}

\item{\code{parameters}}{list containing the parameters for the chosen screening method}

\item{\code{logger}}{a \linkS4class{Logger}}
}}

\author{
Alessandro Barberis
}
