% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Selector.R
\docType{class}
\name{Selector-class}
\alias{Selector-class}
\title{Selector Class
An S4 class providing the methods to select a model from a list.}
\description{
The object consists of 3 slots
}
\section{Slots}{

\describe{
\item{\code{id}}{the object identifier, it must be equivalent to the learning method name (i.e. the trainer id)}

\item{\code{selector}}{function to use for selecting the model}

\item{\code{logger}}{an object of class \linkS4class{Logger}}
}}

\author{
Alessandro Barberis
}
