% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_generics.R, R/class_Tested.R,
%   R/class_TestedList.R, R/class_Evaluated.R, R/class_EvaluatedList.R,
%   R/class_Trained.R, R/class_Screened.R, R/class_Tuned.R, R/class_Filtered.R,
%   R/class_Renoir.R
\name{summary_table}
\alias{summary_table}
\alias{summary_table,Tested-method}
\alias{summary_table,TestedList-method}
\alias{summary_table,Evaluated-method}
\alias{summary_table,EvaluatedList-method}
\alias{summary_table,Trained-method}
\alias{summary_table,Screened-method}
\alias{summary_table,Tuned-method}
\alias{summary_table,Filtered-method}
\alias{summary_table,Renoir-method}
\title{Get the Summary of a Tested Object}
\usage{
summary_table(object, ...)

\S4method{summary_table}{Tested}(object, best = c("opt", "1se"))

\S4method{summary_table}{TestedList}(object, best)

\S4method{summary_table}{Evaluated}(object, best = c("opt", "1se"), key)

\S4method{summary_table}{EvaluatedList}(object, key, best = c("opt", "1se"), ...)

\S4method{summary_table}{Trained}(object)

\S4method{summary_table}{Screened}(object)

\S4method{summary_table}{Tuned}(object)

\S4method{summary_table}{Filtered}(object)

\S4method{summary_table}{Renoir}(object, ...)
}
\arguments{
\item{object}{an object for which a summary is desired}

\item{...}{additional arguments affecting the summary produced}
}
\value{
The form of the returned object depends on the class of its
argument. See the documentation of the particular method for
further details.
}
\description{
Generic function used to produce summary tables of various
objects generated by \code{renoir}. The function invokes particular
methods which depend on the class of the first argument.
}
\details{
Methods for different classes are implemented, e.g.
\linkS4class{Renoir}, \linkS4class{EvaluatedList}, \linkS4class{Evaluated},
\linkS4class{TestedList}, \linkS4class{Tested}.
}
\seealso{
\code{\link{summary_table.Renoir}},
\code{\link{summary_table.EvaluatedList}},
\code{\link{summary_table.Evaluated}},
\code{\link{summary_table.TestedList}},
\code{\link{summary_table.Tested}},
\code{\link{summary_table.Tuned}},
\code{\link{summary_table.Trained}},
\code{\link{summary_table.Screened}}
}
\author{
Alessandro Barberis
}
