% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Tested.R
\docType{class}
\name{Tested-class}
\alias{Tested-class}
\title{Tested
An S4 class to represent a tested model}
\description{
The Tested object consists of 3 slots
}
\section{Slots}{

\describe{
\item{\code{set}}{object of class "character"}

\item{\code{measure}}{object of class "character"}

\item{\code{score}}{a numeric vector}

\item{\code{mscore}}{a length-one numeric vector}

\item{\code{sem}}{a length-one numeric vector}

\item{\code{ci}}{a numeric vector}

\item{\code{opt}}{a length-one integer vector}

\item{\code{1se}}{a length-one integer vector}
}}

