% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Tester.R
\docType{class}
\name{Tester-class}
\alias{Tester-class}
\title{Tester Class
An S4 class providing the methods to test the trained models on provided data.}
\description{
The object consists of 6 slots
}
\section{Slots}{

\describe{
\item{\code{forecaster}}{a \linkS4class{Forecaster}}

\item{\code{scorer}}{a \linkS4class{Scorer}}

\item{\code{looper}}{a \linkS4class{Looper}}

\item{\code{logger}}{a \linkS4class{Logger}}
}}

\author{
Alessandro Barberis
}
